+++ frontend/frontend/public/index.html
@@ -3,12 +3,11 @@
 <head>
   <meta charset="UTF-8">
   <title>Máquina de Galton</title>
-  <link rel="stylesheet" href="style.css">
+  <link rel="stylesheet" href="App.css">
   <script src="https://d3js.org/d3.v6.min.js"></script>
 </head>
 <body>
 <div id="root"></div>
-<script src="script.js"></script>
+<script src="App.js"></script>
 </body>
 </html>
-
