+++ frontend/frontend/src/index.js
@@ -1,8 +1,9 @@
 import React from 'react';
-import ReactDOM from 'react-dom/client';
-import './index.css';
 import App from './App';
 import reportWebVitals from './reportWebVitals';
+import ReactDOM from 'react-dom';
+import './App.css';
+
 
 const root = ReactDOM.createRoot(document.getElementById('root'));
 root.render(
