--- frontend/frontend/src/style.css
@@ -1,7 +1,0 @@
-body {
-    font-family: Arial, sans-serif;
-}
-
-svg {
-    border: 1px solid #ccc;
-}
