+++ frontend/src/main/resources/static/css/main.3282904e.css
@@ -1,0 +1,2 @@
+body{font-family:Arial,sans-serif}svg{border:1px solid #ccc}
+/*# sourceMappingURL=main.3282904e.css.map*/
