+++ frontend/src/main/resources/static/index.html
@@ -1,0 +1,1 @@
+<!doctype html><html lang="en"><head><meta charset="UTF-8"><title>Máquina de Galton</title><link rel="stylesheet" href="App.css"><script src="https://d3js.org/d3.v6.min.js"></script><script defer="defer" src="/static/js/main.fd488f23.js"></script><link href="/static/css/main.3282904e.css" rel="stylesheet"></head><body><div id="root"></div><script src="App.js"></script></body></html>
