+++ frontend/src/main/resources/static/js/main.fd488f23.js
@@ -1,0 +1,3 @@
+/*! For license information please see main.fd488f23.js.LICENSE.txt */
+(()=>{"use strict";var e={730:(e,t,n)=>{var r=n(43),a=n(853);function l(e){for(var t="https://reactjs.org/docs/error-decoder.html?invariant="+e,n=1;n<arguments.length;n++)t+="&args[]="+encodeURIComponent(arguments[n]);return"Minified React error #"+e+"; visit "+t+" for the full message or use the non-minified dev environment for full errors and additional helpful warnings."}var i=new Set,o={};function u(e,t){s(e,t),s(e+"Capture",t)}function s(e,t){for(o[e]=t,e=0;e<t.length;e++)i.add(t[e])}var c=!("undefined"===typeof window||"undefined"===typeof window.document||"undefined"===typeof window.document.createElement),f=Object.prototype.hasOwnProperty,d=/^[:A-Z_a-z\u00C0-\u00D6\u00D8-\u00F6\u00F8-\u02FF\u0370-\u037D\u037F-\u1FFF\u200C-\u200D\u2070-\u218F\u2C00-\u2FEF\u3001-\uD7FF\uF900-\uFDCF\uFDF0-\uFFFD][:A-Z_a-z\u00C0-\u00D6\u00D8-\u00F6\u00F8-\u02FF\u0370-\u037D\u037F-\u1FFF\u200C-\u200D\u2070-\u218F\u2C00-\u2FEF\u3001-\uD7FF\uF900-\uFDCF\uFDF0-\uFFFD\-.0-9\u00B7\u0300-\u036F\u203F-\u2040]*$/,p={},h={};function m(e,t,n,r,a,l,i){this.acceptsBooleans=2===t||3===t||4===t,this.attributeName=r,this.attributeNamespace=a,this.mustUseProperty=n,this.propertyName=e,this.type=t,this.sanitizeURL=l,this.removeEmptyString=i}var v={};"children dangerouslySetInnerHTML defaultValue defaultChecked innerHTML suppressContentEditableWarning suppressHydrationWarning style".split(" ").forEach((function(e){v[e]=new m(e,0,!1,e,null,!1,!1)})),[["acceptCharset","accept-charset"],["className","class"],["htmlFor","for"],["httpEquiv","http-equiv"]].forEach((function(e){var t=e[0];v[t]=new m(t,1,!1,e[1],null,!1,!1)})),["contentEditable","draggable","spellCheck","value"].forEach((function(e){v[e]=new m(e,2,!1,e.toLowerCase(),null,!1,!1)})),["autoReverse","externalResourcesRequired","focusable","preserveAlpha"].forEach((function(e){v[e]=new m(e,2,!1,e,null,!1,!1)})),"allowFullScreen async autoFocus autoPlay controls default defer disabled disablePictureInPicture disableRemotePlayback formNoValidate hidden loop noModule noValidate open playsInline readOnly required reversed scoped seamless itemScope".split(" ").forEach((function(e){v[e]=new m(e,3,!1,e.toLowerCase(),null,!1,!1)})),["checked","multiple","muted","selected"].forEach((function(e){v[e]=new m(e,3,!0,e,null,!1,!1)})),["capture","download"].forEach((function(e){v[e]=new m(e,4,!1,e,null,!1,!1)})),["cols","rows","size","span"].forEach((function(e){v[e]=new m(e,6,!1,e,null,!1,!1)})),["rowSpan","start"].forEach((function(e){v[e]=new m(e,5,!1,e.toLowerCase(),null,!1,!1)}));var g=/[\-:]([a-z])/g;function y(e){return e[1].toUpperCase()}function b(e,t,n,r){var a=v.hasOwnProperty(t)?v[t]:null;(null!==a?0!==a.type:r||!(2<t.length)||"o"!==t[0]&&"O"!==t[0]||"n"!==t[1]&&"N"!==t[1])&&(function(e,t,n,r){if(null===t||"undefined"===typeof t||function(e,t,n,r){if(null!==n&&0===n.type)return!1;switch(typeof t){case"function":case"symbol":return!0;case"boolean":return!r&&(null!==n?!n.acceptsBooleans:"data-"!==(e=e.toLowerCase().slice(0,5))&&"aria-"!==e);default:return!1}}(e,t,n,r))return!0;if(r)return!1;if(null!==n)switch(n.type){case 3:return!t;case 4:return!1===t;case 5:return isNaN(t);case 6:return isNaN(t)||1>t}return!1}(t,n,a,r)&&(n=null),r||null===a?function(e){return!!f.call(h,e)||!f.call(p,e)&&(d.test(e)?h[e]=!0:(p[e]=!0,!1))}(t)&&(null===n?e.removeAttribute(t):e.setAttribute(t,""+n)):a.mustUseProperty?e[a.propertyName]=null===n?3!==a.type&&"":n:(t=a.attributeName,r=a.attributeNamespace,null===n?e.removeAttribute(t):(n=3===(a=a.type)||4===a&&!0===n?"":""+n,r?e.setAttributeNS(r,t,n):e.setAttribute(t,n))))}"accent-height alignment-baseline arabic-form baseline-shift cap-height clip-path clip-rule color-interpolation color-interpolation-filters color-profile color-rendering dominant-baseline enable-background fill-opacity fill-rule flood-color flood-opacity font-family font-size font-size-adjust font-stretch font-style font-variant font-weight glyph-name glyph-orientation-horizontal glyph-orientation-vertical horiz-adv-x horiz-origin-x image-rendering letter-spacing lighting-color marker-end marker-mid marker-start overline-position overline-thickness paint-order panose-1 pointer-events rendering-intent shape-rendering stop-color stop-opacity strikethrough-position strikethrough-thickness stroke-dasharray stroke-dashoffset stroke-linecap stroke-linejoin stroke-miterlimit stroke-opacity stroke-width text-anchor text-decoration text-rendering underline-position underline-thickness unicode-bidi unicode-range units-per-em v-alphabetic v-hanging v-ideographic v-mathematical vector-effect vert-adv-y vert-origin-x vert-origin-y word-spacing writing-mode xmlns:xlink x-height".split(" ").forEach((function(e){var t=e.replace(g,y);v[t]=new m(t,1,!1,e,null,!1,!1)})),"xlink:actuate xlink:arcrole xlink:role xlink:show xlink:title xlink:type".split(" ").forEach((function(e){var t=e.replace(g,y);v[t]=new m(t,1,!1,e,"http://www.w3.org/1999/xlink",!1,!1)})),["xml:base","xml:lang","xml:space"].forEach((function(e){var t=e.replace(g,y);v[t]=new m(t,1,!1,e,"http://www.w3.org/XML/1998/namespace",!1,!1)})),["tabIndex","crossOrigin"].forEach((function(e){v[e]=new m(e,1,!1,e.toLowerCase(),null,!1,!1)})),v.xlinkHref=new m("xlinkHref",1,!1,"xlink:href","http://www.w3.org/1999/xlink",!0,!1),["src","href","action","formAction"].forEach((function(e){v[e]=new m(e,1,!1,e.toLowerCase(),null,!0,!0)}));var w=r.__SECRET_INTERNALS_DO_NOT_USE_OR_YOU_WILL_BE_FIRED,k=Symbol.for("react.element"),_=Symbol.for("react.portal"),S=Symbol.for("react.fragment"),x=Symbol.for("react.strict_mode"),E=Symbol.for("react.profiler"),C=Symbol.for("react.provider"),N=Symbol.for("react.context"),P=Symbol.for("react.forward_ref"),z=Symbol.for("react.suspense"),T=Symbol.for("react.suspense_list"),L=Symbol.for("react.memo"),M=Symbol.for("react.lazy");Symbol.for("react.scope"),Symbol.for("react.debug_trace_mode");var R=Symbol.for("react.offscreen");Symbol.for("react.legacy_hidden"),Symbol.for("react.cache"),Symbol.for("react.tracing_marker");var O=Symbol.iterator;function F(e){return null===e||"object"!==typeof e?null:"function"===typeof(e=O&&e[O]||e["@@iterator"])?e:null}var D,I=Object.assign;function A(e){if(void 0===D)try{throw Error()}catch(n){var t=n.stack.trim().match(/\n( *(at )?)/);D=t&&t[1]||""}return"\n"+D+e}var $=!1;function U(e,t){if(!e||$)return"";$=!0;var n=Error.prepareStackTrace;Error.prepareStackTrace=void 0;try{if(t)if(t=function(){throw Error()},Object.defineProperty(t.prototype,"props",{set:function(){throw Error()}}),"object"===typeof Reflect&&Reflect.construct){try{Reflect.construct(t,[])}catch(s){var r=s}Reflect.construct(e,[],t)}else{try{t.call()}catch(s){r=s}e.call(t.prototype)}else{try{throw Error()}catch(s){r=s}e()}}catch(s){if(s&&r&&"string"===typeof s.stack){for(var a=s.stack.split("\n"),l=r.stack.split("\n"),i=a.length-1,o=l.length-1;1<=i&&0<=o&&a[i]!==l[o];)o--;for(;1<=i&&0<=o;i--,o--)if(a[i]!==l[o]){if(1!==i||1!==o)do{if(i--,0>--o||a[i]!==l[o]){var u="\n"+a[i].replace(" at new "," at ");return e.displayName&&u.includes("<anonymous>")&&(u=u.replace("<anonymous>",e.displayName)),u}}while(1<=i&&0<=o);break}}}finally{$=!1,Error.prepareStackTrace=n}return(e=e?e.displayName||e.name:"")?A(e):""}function j(e){switch(e.tag){case 5:return A(e.type);case 16:return A("Lazy");case 13:return A("Suspense");case 19:return A("SuspenseList");case 0:case 2:case 15:return e=U(e.type,!1);case 11:return e=U(e.type.render,!1);case 1:return e=U(e.type,!0);default:return""}}function V(e){if(null==e)return null;if("function"===typeof e)return e.displayName||e.name||null;if("string"===typeof e)return e;switch(e){case S:return"Fragment";case _:return"Portal";case E:return"Profiler";case x:return"StrictMode";case z:return"Suspense";case T:return"SuspenseList"}if("object"===typeof e)switch(e.$$typeof){case N:return(e.displayName||"Context")+".Consumer";case C:return(e._context.displayName||"Context")+".Provider";case P:var t=e.render;return(e=e.displayName)||(e=""!==(e=t.displayName||t.name||"")?"ForwardRef("+e+")":"ForwardRef"),e;case L:return null!==(t=e.displayName||null)?t:V(e.type)||"Memo";case M:t=e._payload,e=e._init;try{return V(e(t))}catch(n){}}return null}function B(e){var t=e.type;switch(e.tag){case 24:return"Cache";case 9:return(t.displayName||"Context")+".Consumer";case 10:return(t._context.displayName||"Context")+".Provider";case 18:return"DehydratedFragment";case 11:return e=(e=t.render).displayName||e.name||"",t.displayName||(""!==e?"ForwardRef("+e+")":"ForwardRef");case 7:return"Fragment";case 5:return t;case 4:return"Portal";case 3:return"Root";case 6:return"Text";case 16:return V(t);case 8:return t===x?"StrictMode":"Mode";case 22:return"Offscreen";case 12:return"Profiler";case 21:return"Scope";case 13:return"Suspense";case 19:return"SuspenseList";case 25:return"TracingMarker";case 1:case 0:case 17:case 2:case 14:case 15:if("function"===typeof t)return t.displayName||t.name||null;if("string"===typeof t)return t}return null}function H(e){switch(typeof e){case"boolean":case"number":case"string":case"undefined":case"object":return e;default:return""}}function W(e){var t=e.type;return(e=e.nodeName)&&"input"===e.toLowerCase()&&("checkbox"===t||"radio"===t)}function Q(e){e._valueTracker||(e._valueTracker=function(e){var t=W(e)?"checked":"value",n=Object.getOwnPropertyDescriptor(e.constructor.prototype,t),r=""+e[t];if(!e.hasOwnProperty(t)&&"undefined"!==typeof n&&"function"===typeof n.get&&"function"===typeof n.set){var a=n.get,l=n.set;return Object.defineProperty(e,t,{configurable:!0,get:function(){return a.call(this)},set:function(e){r=""+e,l.call(this,e)}}),Object.defineProperty(e,t,{enumerable:n.enumerable}),{getValue:function(){return r},setValue:function(e){r=""+e},stopTracking:function(){e._valueTracker=null,delete e[t]}}}}(e))}function q(e){if(!e)return!1;var t=e._valueTracker;if(!t)return!0;var n=t.getValue(),r="";return e&&(r=W(e)?e.checked?"true":"false":e.value),(e=r)!==n&&(t.setValue(e),!0)}function X(e){if("undefined"===typeof(e=e||("undefined"!==typeof document?document:void 0)))return null;try{return e.activeElement||e.body}catch(t){return e.body}}function Y(e,t){var n=t.checked;return I({},t,{defaultChecked:void 0,defaultValue:void 0,value:void 0,checked:null!=n?n:e._wrapperState.initialChecked})}function K(e,t){var n=null==t.defaultValue?"":t.defaultValue,r=null!=t.checked?t.checked:t.defaultChecked;n=H(null!=t.value?t.value:n),e._wrapperState={initialChecked:r,initialValue:n,controlled:"checkbox"===t.type||"radio"===t.type?null!=t.checked:null!=t.value}}function G(e,t){null!=(t=t.checked)&&b(e,"checked",t,!1)}function Z(e,t){G(e,t);var n=H(t.value),r=t.type;if(null!=n)"number"===r?(0===n&&""===e.value||e.value!=n)&&(e.value=""+n):e.value!==""+n&&(e.value=""+n);else if("submit"===r||"reset"===r)return void e.removeAttribute("value");t.hasOwnProperty("value")?ee(e,t.type,n):t.hasOwnProperty("defaultValue")&&ee(e,t.type,H(t.defaultValue)),null==t.checked&&null!=t.defaultChecked&&(e.defaultChecked=!!t.defaultChecked)}function J(e,t,n){if(t.hasOwnProperty("value")||t.hasOwnProperty("defaultValue")){var r=t.type;if(!("submit"!==r&&"reset"!==r||void 0!==t.value&&null!==t.value))return;t=""+e._wrapperState.initialValue,n||t===e.value||(e.value=t),e.defaultValue=t}""!==(n=e.name)&&(e.name=""),e.defaultChecked=!!e._wrapperState.initialChecked,""!==n&&(e.name=n)}function ee(e,t,n){"number"===t&&X(e.ownerDocument)===e||(null==n?e.defaultValue=""+e._wrapperState.initialValue:e.defaultValue!==""+n&&(e.defaultValue=""+n))}var te=Array.isArray;function ne(e,t,n,r){if(e=e.options,t){t={};for(var a=0;a<n.length;a++)t["$"+n[a]]=!0;for(n=0;n<e.length;n++)a=t.hasOwnProperty("$"+e[n].value),e[n].selected!==a&&(e[n].selected=a),a&&r&&(e[n].defaultSelected=!0)}else{for(n=""+H(n),t=null,a=0;a<e.length;a++){if(e[a].value===n)return e[a].selected=!0,void(r&&(e[a].defaultSelected=!0));null!==t||e[a].disabled||(t=e[a])}null!==t&&(t.selected=!0)}}function re(e,t){if(null!=t.dangerouslySetInnerHTML)throw Error(l(91));return I({},t,{value:void 0,defaultValue:void 0,children:""+e._wrapperState.initialValue})}function ae(e,t){var n=t.value;if(null==n){if(n=t.children,t=t.defaultValue,null!=n){if(null!=t)throw Error(l(92));if(te(n)){if(1<n.length)throw Error(l(93));n=n[0]}t=n}null==t&&(t=""),n=t}e._wrapperState={initialValue:H(n)}}function le(e,t){var n=H(t.value),r=H(t.defaultValue);null!=n&&((n=""+n)!==e.value&&(e.value=n),null==t.defaultValue&&e.defaultValue!==n&&(e.defaultValue=n)),null!=r&&(e.defaultValue=""+r)}function ie(e){var t=e.textContent;t===e._wrapperState.initialValue&&""!==t&&null!==t&&(e.value=t)}function oe(e){switch(e){case"svg":return"http://www.w3.org/2000/svg";case"math":return"http://www.w3.org/1998/Math/MathML";default:return"http://www.w3.org/1999/xhtml"}}function ue(e,t){return null==e||"http://www.w3.org/1999/xhtml"===e?oe(t):"http://www.w3.org/2000/svg"===e&&"foreignObject"===t?"http://www.w3.org/1999/xhtml":e}var se,ce,fe=(ce=function(e,t){if("http://www.w3.org/2000/svg"!==e.namespaceURI||"innerHTML"in e)e.innerHTML=t;else{for((se=se||document.createElement("div")).innerHTML="<svg>"+t.valueOf().toString()+"</svg>",t=se.firstChild;e.firstChild;)e.removeChild(e.firstChild);for(;t.firstChild;)e.appendChild(t.firstChild)}},"undefined"!==typeof MSApp&&MSApp.execUnsafeLocalFunction?function(e,t,n,r){MSApp.execUnsafeLocalFunction((function(){return ce(e,t)}))}:ce);function de(e,t){if(t){var n=e.firstChild;if(n&&n===e.lastChild&&3===n.nodeType)return void(n.nodeValue=t)}e.textContent=t}var pe={animationIterationCount:!0,aspectRatio:!0,borderImageOutset:!0,borderImageSlice:!0,borderImageWidth:!0,boxFlex:!0,boxFlexGroup:!0,boxOrdinalGroup:!0,columnCount:!0,columns:!0,flex:!0,flexGrow:!0,flexPositive:!0,flexShrink:!0,flexNegative:!0,flexOrder:!0,gridArea:!0,gridRow:!0,gridRowEnd:!0,gridRowSpan:!0,gridRowStart:!0,gridColumn:!0,gridColumnEnd:!0,gridColumnSpan:!0,gridColumnStart:!0,fontWeight:!0,lineClamp:!0,lineHeight:!0,opacity:!0,order:!0,orphans:!0,tabSize:!0,widows:!0,zIndex:!0,zoom:!0,fillOpacity:!0,floodOpacity:!0,stopOpacity:!0,strokeDasharray:!0,strokeDashoffset:!0,strokeMiterlimit:!0,strokeOpacity:!0,strokeWidth:!0},he=["Webkit","ms","Moz","O"];function me(e,t,n){return null==t||"boolean"===typeof t||""===t?"":n||"number"!==typeof t||0===t||pe.hasOwnProperty(e)&&pe[e]?(""+t).trim():t+"px"}function ve(e,t){for(var n in e=e.style,t)if(t.hasOwnProperty(n)){var r=0===n.indexOf("--"),a=me(n,t[n],r);"float"===n&&(n="cssFloat"),r?e.setProperty(n,a):e[n]=a}}Object.keys(pe).forEach((function(e){he.forEach((function(t){t=t+e.charAt(0).toUpperCase()+e.substring(1),pe[t]=pe[e]}))}));var ge=I({menuitem:!0},{area:!0,base:!0,br:!0,col:!0,embed:!0,hr:!0,img:!0,input:!0,keygen:!0,link:!0,meta:!0,param:!0,source:!0,track:!0,wbr:!0});function ye(e,t){if(t){if(ge[e]&&(null!=t.children||null!=t.dangerouslySetInnerHTML))throw Error(l(137,e));if(null!=t.dangerouslySetInnerHTML){if(null!=t.children)throw Error(l(60));if("object"!==typeof t.dangerouslySetInnerHTML||!("__html"in t.dangerouslySetInnerHTML))throw Error(l(61))}if(null!=t.style&&"object"!==typeof t.style)throw Error(l(62))}}function be(e,t){if(-1===e.indexOf("-"))return"string"===typeof t.is;switch(e){case"annotation-xml":case"color-profile":case"font-face":case"font-face-src":case"font-face-uri":case"font-face-format":case"font-face-name":case"missing-glyph":return!1;default:return!0}}var we=null;function ke(e){return(e=e.target||e.srcElement||window).correspondingUseElement&&(e=e.correspondingUseElement),3===e.nodeType?e.parentNode:e}var _e=null,Se=null,xe=null;function Ee(e){if(e=ba(e)){if("function"!==typeof _e)throw Error(l(280));var t=e.stateNode;t&&(t=ka(t),_e(e.stateNode,e.type,t))}}function Ce(e){Se?xe?xe.push(e):xe=[e]:Se=e}function Ne(){if(Se){var e=Se,t=xe;if(xe=Se=null,Ee(e),t)for(e=0;e<t.length;e++)Ee(t[e])}}function Pe(e,t){return e(t)}function ze(){}var Te=!1;function Le(e,t,n){if(Te)return e(t,n);Te=!0;try{return Pe(e,t,n)}finally{Te=!1,(null!==Se||null!==xe)&&(ze(),Ne())}}function Me(e,t){var n=e.stateNode;if(null===n)return null;var r=ka(n);if(null===r)return null;n=r[t];e:switch(t){case"onClick":case"onClickCapture":case"onDoubleClick":case"onDoubleClickCapture":case"onMouseDown":case"onMouseDownCapture":case"onMouseMove":case"onMouseMoveCapture":case"onMouseUp":case"onMouseUpCapture":case"onMouseEnter":(r=!r.disabled)||(r=!("button"===(e=e.type)||"input"===e||"select"===e||"textarea"===e)),e=!r;break e;default:e=!1}if(e)return null;if(n&&"function"!==typeof n)throw Error(l(231,t,typeof n));return n}var Re=!1;if(c)try{var Oe={};Object.defineProperty(Oe,"passive",{get:function(){Re=!0}}),window.addEventListener("test",Oe,Oe),window.removeEventListener("test",Oe,Oe)}catch(ce){Re=!1}function Fe(e,t,n,r,a,l,i,o,u){var s=Array.prototype.slice.call(arguments,3);try{t.apply(n,s)}catch(c){this.onError(c)}}var De=!1,Ie=null,Ae=!1,$e=null,Ue={onError:function(e){De=!0,Ie=e}};function je(e,t,n,r,a,l,i,o,u){De=!1,Ie=null,Fe.apply(Ue,arguments)}function Ve(e){var t=e,n=e;if(e.alternate)for(;t.return;)t=t.return;else{e=t;do{0!==(4098&(t=e).flags)&&(n=t.return),e=t.return}while(e)}return 3===t.tag?n:null}function Be(e){if(13===e.tag){var t=e.memoizedState;if(null===t&&(null!==(e=e.alternate)&&(t=e.memoizedState)),null!==t)return t.dehydrated}return null}function He(e){if(Ve(e)!==e)throw Error(l(188))}function We(e){return null!==(e=function(e){var t=e.alternate;if(!t){if(null===(t=Ve(e)))throw Error(l(188));return t!==e?null:e}for(var n=e,r=t;;){var a=n.return;if(null===a)break;var i=a.alternate;if(null===i){if(null!==(r=a.return)){n=r;continue}break}if(a.child===i.child){for(i=a.child;i;){if(i===n)return He(a),e;if(i===r)return He(a),t;i=i.sibling}throw Error(l(188))}if(n.return!==r.return)n=a,r=i;else{for(var o=!1,u=a.child;u;){if(u===n){o=!0,n=a,r=i;break}if(u===r){o=!0,r=a,n=i;break}u=u.sibling}if(!o){for(u=i.child;u;){if(u===n){o=!0,n=i,r=a;break}if(u===r){o=!0,r=i,n=a;break}u=u.sibling}if(!o)throw Error(l(189))}}if(n.alternate!==r)throw Error(l(190))}if(3!==n.tag)throw Error(l(188));return n.stateNode.current===n?e:t}(e))?Qe(e):null}function Qe(e){if(5===e.tag||6===e.tag)return e;for(e=e.child;null!==e;){var t=Qe(e);if(null!==t)return t;e=e.sibling}return null}var qe=a.unstable_scheduleCallback,Xe=a.unstable_cancelCallback,Ye=a.unstable_shouldYield,Ke=a.unstable_requestPaint,Ge=a.unstable_now,Ze=a.unstable_getCurrentPriorityLevel,Je=a.unstable_ImmediatePriority,et=a.unstable_UserBlockingPriority,tt=a.unstable_NormalPriority,nt=a.unstable_LowPriority,rt=a.unstable_IdlePriority,at=null,lt=null;var it=Math.clz32?Math.clz32:function(e){return e>>>=0,0===e?32:31-(ot(e)/ut|0)|0},ot=Math.log,ut=Math.LN2;var st=64,ct=4194304;function ft(e){switch(e&-e){case 1:return 1;case 2:return 2;case 4:return 4;case 8:return 8;case 16:return 16;case 32:return 32;case 64:case 128:case 256:case 512:case 1024:case 2048:case 4096:case 8192:case 16384:case 32768:case 65536:case 131072:case 262144:case 524288:case 1048576:case 2097152:return 4194240&e;case 4194304:case 8388608:case 16777216:case 33554432:case 67108864:return 130023424&e;case 134217728:return 134217728;case 268435456:return 268435456;case 536870912:return 536870912;case 1073741824:return 1073741824;default:return e}}function dt(e,t){var n=e.pendingLanes;if(0===n)return 0;var r=0,a=e.suspendedLanes,l=e.pingedLanes,i=268435455&n;if(0!==i){var o=i&~a;0!==o?r=ft(o):0!==(l&=i)&&(r=ft(l))}else 0!==(i=n&~a)?r=ft(i):0!==l&&(r=ft(l));if(0===r)return 0;if(0!==t&&t!==r&&0===(t&a)&&((a=r&-r)>=(l=t&-t)||16===a&&0!==(4194240&l)))return t;if(0!==(4&r)&&(r|=16&n),0!==(t=e.entangledLanes))for(e=e.entanglements,t&=r;0<t;)a=1<<(n=31-it(t)),r|=e[n],t&=~a;return r}function pt(e,t){switch(e){case 1:case 2:case 4:return t+250;case 8:case 16:case 32:case 64:case 128:case 256:case 512:case 1024:case 2048:case 4096:case 8192:case 16384:case 32768:case 65536:case 131072:case 262144:case 524288:case 1048576:case 2097152:return t+5e3;default:return-1}}function ht(e){return 0!==(e=-1073741825&e.pendingLanes)?e:1073741824&e?1073741824:0}function mt(){var e=st;return 0===(4194240&(st<<=1))&&(st=64),e}function vt(e){for(var t=[],n=0;31>n;n++)t.push(e);return t}function gt(e,t,n){e.pendingLanes|=t,536870912!==t&&(e.suspendedLanes=0,e.pingedLanes=0),(e=e.eventTimes)[t=31-it(t)]=n}function yt(e,t){var n=e.entangledLanes|=t;for(e=e.entanglements;n;){var r=31-it(n),a=1<<r;a&t|e[r]&t&&(e[r]|=t),n&=~a}}var bt=0;function wt(e){return 1<(e&=-e)?4<e?0!==(268435455&e)?16:536870912:4:1}var kt,_t,St,xt,Et,Ct=!1,Nt=[],Pt=null,zt=null,Tt=null,Lt=new Map,Mt=new Map,Rt=[],Ot="mousedown mouseup touchcancel touchend touchstart auxclick dblclick pointercancel pointerdown pointerup dragend dragstart drop compositionend compositionstart keydown keypress keyup input textInput copy cut paste click change contextmenu reset submit".split(" ");function Ft(e,t){switch(e){case"focusin":case"focusout":Pt=null;break;case"dragenter":case"dragleave":zt=null;break;case"mouseover":case"mouseout":Tt=null;break;case"pointerover":case"pointerout":Lt.delete(t.pointerId);break;case"gotpointercapture":case"lostpointercapture":Mt.delete(t.pointerId)}}function Dt(e,t,n,r,a,l){return null===e||e.nativeEvent!==l?(e={blockedOn:t,domEventName:n,eventSystemFlags:r,nativeEvent:l,targetContainers:[a]},null!==t&&(null!==(t=ba(t))&&_t(t)),e):(e.eventSystemFlags|=r,t=e.targetContainers,null!==a&&-1===t.indexOf(a)&&t.push(a),e)}function It(e){var t=ya(e.target);if(null!==t){var n=Ve(t);if(null!==n)if(13===(t=n.tag)){if(null!==(t=Be(n)))return e.blockedOn=t,void Et(e.priority,(function(){St(n)}))}else if(3===t&&n.stateNode.current.memoizedState.isDehydrated)return void(e.blockedOn=3===n.tag?n.stateNode.containerInfo:null)}e.blockedOn=null}function At(e){if(null!==e.blockedOn)return!1;for(var t=e.targetContainers;0<t.length;){var n=Yt(e.domEventName,e.eventSystemFlags,t[0],e.nativeEvent);if(null!==n)return null!==(t=ba(n))&&_t(t),e.blockedOn=n,!1;var r=new(n=e.nativeEvent).constructor(n.type,n);we=r,n.target.dispatchEvent(r),we=null,t.shift()}return!0}function $t(e,t,n){At(e)&&n.delete(t)}function Ut(){Ct=!1,null!==Pt&&At(Pt)&&(Pt=null),null!==zt&&At(zt)&&(zt=null),null!==Tt&&At(Tt)&&(Tt=null),Lt.forEach($t),Mt.forEach($t)}function jt(e,t){e.blockedOn===t&&(e.blockedOn=null,Ct||(Ct=!0,a.unstable_scheduleCallback(a.unstable_NormalPriority,Ut)))}function Vt(e){function t(t){return jt(t,e)}if(0<Nt.length){jt(Nt[0],e);for(var n=1;n<Nt.length;n++){var r=Nt[n];r.blockedOn===e&&(r.blockedOn=null)}}for(null!==Pt&&jt(Pt,e),null!==zt&&jt(zt,e),null!==Tt&&jt(Tt,e),Lt.forEach(t),Mt.forEach(t),n=0;n<Rt.length;n++)(r=Rt[n]).blockedOn===e&&(r.blockedOn=null);for(;0<Rt.length&&null===(n=Rt[0]).blockedOn;)It(n),null===n.blockedOn&&Rt.shift()}var Bt=w.ReactCurrentBatchConfig,Ht=!0;function Wt(e,t,n,r){var a=bt,l=Bt.transition;Bt.transition=null;try{bt=1,qt(e,t,n,r)}finally{bt=a,Bt.transition=l}}function Qt(e,t,n,r){var a=bt,l=Bt.transition;Bt.transition=null;try{bt=4,qt(e,t,n,r)}finally{bt=a,Bt.transition=l}}function qt(e,t,n,r){if(Ht){var a=Yt(e,t,n,r);if(null===a)Hr(e,t,r,Xt,n),Ft(e,r);else if(function(e,t,n,r,a){switch(t){case"focusin":return Pt=Dt(Pt,e,t,n,r,a),!0;case"dragenter":return zt=Dt(zt,e,t,n,r,a),!0;case"mouseover":return Tt=Dt(Tt,e,t,n,r,a),!0;case"pointerover":var l=a.pointerId;return Lt.set(l,Dt(Lt.get(l)||null,e,t,n,r,a)),!0;case"gotpointercapture":return l=a.pointerId,Mt.set(l,Dt(Mt.get(l)||null,e,t,n,r,a)),!0}return!1}(a,e,t,n,r))r.stopPropagation();else if(Ft(e,r),4&t&&-1<Ot.indexOf(e)){for(;null!==a;){var l=ba(a);if(null!==l&&kt(l),null===(l=Yt(e,t,n,r))&&Hr(e,t,r,Xt,n),l===a)break;a=l}null!==a&&r.stopPropagation()}else Hr(e,t,r,null,n)}}var Xt=null;function Yt(e,t,n,r){if(Xt=null,null!==(e=ya(e=ke(r))))if(null===(t=Ve(e)))e=null;else if(13===(n=t.tag)){if(null!==(e=Be(t)))return e;e=null}else if(3===n){if(t.stateNode.current.memoizedState.isDehydrated)return 3===t.tag?t.stateNode.containerInfo:null;e=null}else t!==e&&(e=null);return Xt=e,null}function Kt(e){switch(e){case"cancel":case"click":case"close":case"contextmenu":case"copy":case"cut":case"auxclick":case"dblclick":case"dragend":case"dragstart":case"drop":case"focusin":case"focusout":case"input":case"invalid":case"keydown":case"keypress":case"keyup":case"mousedown":case"mouseup":case"paste":case"pause":case"play":case"pointercancel":case"pointerdown":case"pointerup":case"ratechange":case"reset":case"resize":case"seeked":case"submit":case"touchcancel":case"touchend":case"touchstart":case"volumechange":case"change":case"selectionchange":case"textInput":case"compositionstart":case"compositionend":case"compositionupdate":case"beforeblur":case"afterblur":case"beforeinput":case"blur":case"fullscreenchange":case"focus":case"hashchange":case"popstate":case"select":case"selectstart":return 1;case"drag":case"dragenter":case"dragexit":case"dragleave":case"dragover":case"mousemove":case"mouseout":case"mouseover":case"pointermove":case"pointerout":case"pointerover":case"scroll":case"toggle":case"touchmove":case"wheel":case"mouseenter":case"mouseleave":case"pointerenter":case"pointerleave":return 4;case"message":switch(Ze()){case Je:return 1;case et:return 4;case tt:case nt:return 16;case rt:return 536870912;default:return 16}default:return 16}}var Gt=null,Zt=null,Jt=null;function en(){if(Jt)return Jt;var e,t,n=Zt,r=n.length,a="value"in Gt?Gt.value:Gt.textContent,l=a.length;for(e=0;e<r&&n[e]===a[e];e++);var i=r-e;for(t=1;t<=i&&n[r-t]===a[l-t];t++);return Jt=a.slice(e,1<t?1-t:void 0)}function tn(e){var t=e.keyCode;return"charCode"in e?0===(e=e.charCode)&&13===t&&(e=13):e=t,10===e&&(e=13),32<=e||13===e?e:0}function nn(){return!0}function rn(){return!1}function an(e){function t(t,n,r,a,l){for(var i in this._reactName=t,this._targetInst=r,this.type=n,this.nativeEvent=a,this.target=l,this.currentTarget=null,e)e.hasOwnProperty(i)&&(t=e[i],this[i]=t?t(a):a[i]);return this.isDefaultPrevented=(null!=a.defaultPrevented?a.defaultPrevented:!1===a.returnValue)?nn:rn,this.isPropagationStopped=rn,this}return I(t.prototype,{preventDefault:function(){this.defaultPrevented=!0;var e=this.nativeEvent;e&&(e.preventDefault?e.preventDefault():"unknown"!==typeof e.returnValue&&(e.returnValue=!1),this.isDefaultPrevented=nn)},stopPropagation:function(){var e=this.nativeEvent;e&&(e.stopPropagation?e.stopPropagation():"unknown"!==typeof e.cancelBubble&&(e.cancelBubble=!0),this.isPropagationStopped=nn)},persist:function(){},isPersistent:nn}),t}var ln,on,un,sn={eventPhase:0,bubbles:0,cancelable:0,timeStamp:function(e){return e.timeStamp||Date.now()},defaultPrevented:0,isTrusted:0},cn=an(sn),fn=I({},sn,{view:0,detail:0}),dn=an(fn),pn=I({},fn,{screenX:0,screenY:0,clientX:0,clientY:0,pageX:0,pageY:0,ctrlKey:0,shiftKey:0,altKey:0,metaKey:0,getModifierState:En,button:0,buttons:0,relatedTarget:function(e){return void 0===e.relatedTarget?e.fromElement===e.srcElement?e.toElement:e.fromElement:e.relatedTarget},movementX:function(e){return"movementX"in e?e.movementX:(e!==un&&(un&&"mousemove"===e.type?(ln=e.screenX-un.screenX,on=e.screenY-un.screenY):on=ln=0,un=e),ln)},movementY:function(e){return"movementY"in e?e.movementY:on}}),hn=an(pn),mn=an(I({},pn,{dataTransfer:0})),vn=an(I({},fn,{relatedTarget:0})),gn=an(I({},sn,{animationName:0,elapsedTime:0,pseudoElement:0})),yn=I({},sn,{clipboardData:function(e){return"clipboardData"in e?e.clipboardData:window.clipboardData}}),bn=an(yn),wn=an(I({},sn,{data:0})),kn={Esc:"Escape",Spacebar:" ",Left:"ArrowLeft",Up:"ArrowUp",Right:"ArrowRight",Down:"ArrowDown",Del:"Delete",Win:"OS",Menu:"ContextMenu",Apps:"ContextMenu",Scroll:"ScrollLock",MozPrintableKey:"Unidentified"},_n={8:"Backspace",9:"Tab",12:"Clear",13:"Enter",16:"Shift",17:"Control",18:"Alt",19:"Pause",20:"CapsLock",27:"Escape",32:" ",33:"PageUp",34:"PageDown",35:"End",36:"Home",37:"ArrowLeft",38:"ArrowUp",39:"ArrowRight",40:"ArrowDown",45:"Insert",46:"Delete",112:"F1",113:"F2",114:"F3",115:"F4",116:"F5",117:"F6",118:"F7",119:"F8",120:"F9",121:"F10",122:"F11",123:"F12",144:"NumLock",145:"ScrollLock",224:"Meta"},Sn={Alt:"altKey",Control:"ctrlKey",Meta:"metaKey",Shift:"shiftKey"};function xn(e){var t=this.nativeEvent;return t.getModifierState?t.getModifierState(e):!!(e=Sn[e])&&!!t[e]}function En(){return xn}var Cn=I({},fn,{key:function(e){if(e.key){var t=kn[e.key]||e.key;if("Unidentified"!==t)return t}return"keypress"===e.type?13===(e=tn(e))?"Enter":String.fromCharCode(e):"keydown"===e.type||"keyup"===e.type?_n[e.keyCode]||"Unidentified":""},code:0,location:0,ctrlKey:0,shiftKey:0,altKey:0,metaKey:0,repeat:0,locale:0,getModifierState:En,charCode:function(e){return"keypress"===e.type?tn(e):0},keyCode:function(e){return"keydown"===e.type||"keyup"===e.type?e.keyCode:0},which:function(e){return"keypress"===e.type?tn(e):"keydown"===e.type||"keyup"===e.type?e.keyCode:0}}),Nn=an(Cn),Pn=an(I({},pn,{pointerId:0,width:0,height:0,pressure:0,tangentialPressure:0,tiltX:0,tiltY:0,twist:0,pointerType:0,isPrimary:0})),zn=an(I({},fn,{touches:0,targetTouches:0,changedTouches:0,altKey:0,metaKey:0,ctrlKey:0,shiftKey:0,getModifierState:En})),Tn=an(I({},sn,{propertyName:0,elapsedTime:0,pseudoElement:0})),Ln=I({},pn,{deltaX:function(e){return"deltaX"in e?e.deltaX:"wheelDeltaX"in e?-e.wheelDeltaX:0},deltaY:function(e){return"deltaY"in e?e.deltaY:"wheelDeltaY"in e?-e.wheelDeltaY:"wheelDelta"in e?-e.wheelDelta:0},deltaZ:0,deltaMode:0}),Mn=an(Ln),Rn=[9,13,27,32],On=c&&"CompositionEvent"in window,Fn=null;c&&"documentMode"in document&&(Fn=document.documentMode);var Dn=c&&"TextEvent"in window&&!Fn,In=c&&(!On||Fn&&8<Fn&&11>=Fn),An=String.fromCharCode(32),$n=!1;function Un(e,t){switch(e){case"keyup":return-1!==Rn.indexOf(t.keyCode);case"keydown":return 229!==t.keyCode;case"keypress":case"mousedown":case"focusout":return!0;default:return!1}}function jn(e){return"object"===typeof(e=e.detail)&&"data"in e?e.data:null}var Vn=!1;var Bn={color:!0,date:!0,datetime:!0,"datetime-local":!0,email:!0,month:!0,number:!0,password:!0,range:!0,search:!0,tel:!0,text:!0,time:!0,url:!0,week:!0};function Hn(e){var t=e&&e.nodeName&&e.nodeName.toLowerCase();return"input"===t?!!Bn[e.type]:"textarea"===t}function Wn(e,t,n,r){Ce(r),0<(t=Qr(t,"onChange")).length&&(n=new cn("onChange","change",null,n,r),e.push({event:n,listeners:t}))}var Qn=null,qn=null;function Xn(e){Ar(e,0)}function Yn(e){if(q(wa(e)))return e}function Kn(e,t){if("change"===e)return t}var Gn=!1;if(c){var Zn;if(c){var Jn="oninput"in document;if(!Jn){var er=document.createElement("div");er.setAttribute("oninput","return;"),Jn="function"===typeof er.oninput}Zn=Jn}else Zn=!1;Gn=Zn&&(!document.documentMode||9<document.documentMode)}function tr(){Qn&&(Qn.detachEvent("onpropertychange",nr),qn=Qn=null)}function nr(e){if("value"===e.propertyName&&Yn(qn)){var t=[];Wn(t,qn,e,ke(e)),Le(Xn,t)}}function rr(e,t,n){"focusin"===e?(tr(),qn=n,(Qn=t).attachEvent("onpropertychange",nr)):"focusout"===e&&tr()}function ar(e){if("selectionchange"===e||"keyup"===e||"keydown"===e)return Yn(qn)}function lr(e,t){if("click"===e)return Yn(t)}function ir(e,t){if("input"===e||"change"===e)return Yn(t)}var or="function"===typeof Object.is?Object.is:function(e,t){return e===t&&(0!==e||1/e===1/t)||e!==e&&t!==t};function ur(e,t){if(or(e,t))return!0;if("object"!==typeof e||null===e||"object"!==typeof t||null===t)return!1;var n=Object.keys(e),r=Object.keys(t);if(n.length!==r.length)return!1;for(r=0;r<n.length;r++){var a=n[r];if(!f.call(t,a)||!or(e[a],t[a]))return!1}return!0}function sr(e){for(;e&&e.firstChild;)e=e.firstChild;return e}function cr(e,t){var n,r=sr(e);for(e=0;r;){if(3===r.nodeType){if(n=e+r.textContent.length,e<=t&&n>=t)return{node:r,offset:t-e};e=n}e:{for(;r;){if(r.nextSibling){r=r.nextSibling;break e}r=r.parentNode}r=void 0}r=sr(r)}}function fr(e,t){return!(!e||!t)&&(e===t||(!e||3!==e.nodeType)&&(t&&3===t.nodeType?fr(e,t.parentNode):"contains"in e?e.contains(t):!!e.compareDocumentPosition&&!!(16&e.compareDocumentPosition(t))))}function dr(){for(var e=window,t=X();t instanceof e.HTMLIFrameElement;){try{var n="string"===typeof t.contentWindow.location.href}catch(r){n=!1}if(!n)break;t=X((e=t.contentWindow).document)}return t}function pr(e){var t=e&&e.nodeName&&e.nodeName.toLowerCase();return t&&("input"===t&&("text"===e.type||"search"===e.type||"tel"===e.type||"url"===e.type||"password"===e.type)||"textarea"===t||"true"===e.contentEditable)}function hr(e){var t=dr(),n=e.focusedElem,r=e.selectionRange;if(t!==n&&n&&n.ownerDocument&&fr(n.ownerDocument.documentElement,n)){if(null!==r&&pr(n))if(t=r.start,void 0===(e=r.end)&&(e=t),"selectionStart"in n)n.selectionStart=t,n.selectionEnd=Math.min(e,n.value.length);else if((e=(t=n.ownerDocument||document)&&t.defaultView||window).getSelection){e=e.getSelection();var a=n.textContent.length,l=Math.min(r.start,a);r=void 0===r.end?l:Math.min(r.end,a),!e.extend&&l>r&&(a=r,r=l,l=a),a=cr(n,l);var i=cr(n,r);a&&i&&(1!==e.rangeCount||e.anchorNode!==a.node||e.anchorOffset!==a.offset||e.focusNode!==i.node||e.focusOffset!==i.offset)&&((t=t.createRange()).setStart(a.node,a.offset),e.removeAllRanges(),l>r?(e.addRange(t),e.extend(i.node,i.offset)):(t.setEnd(i.node,i.offset),e.addRange(t)))}for(t=[],e=n;e=e.parentNode;)1===e.nodeType&&t.push({element:e,left:e.scrollLeft,top:e.scrollTop});for("function"===typeof n.focus&&n.focus(),n=0;n<t.length;n++)(e=t[n]).element.scrollLeft=e.left,e.element.scrollTop=e.top}}var mr=c&&"documentMode"in document&&11>=document.documentMode,vr=null,gr=null,yr=null,br=!1;function wr(e,t,n){var r=n.window===n?n.document:9===n.nodeType?n:n.ownerDocument;br||null==vr||vr!==X(r)||("selectionStart"in(r=vr)&&pr(r)?r={start:r.selectionStart,end:r.selectionEnd}:r={anchorNode:(r=(r.ownerDocument&&r.ownerDocument.defaultView||window).getSelection()).anchorNode,anchorOffset:r.anchorOffset,focusNode:r.focusNode,focusOffset:r.focusOffset},yr&&ur(yr,r)||(yr=r,0<(r=Qr(gr,"onSelect")).length&&(t=new cn("onSelect","select",null,t,n),e.push({event:t,listeners:r}),t.target=vr)))}function kr(e,t){var n={};return n[e.toLowerCase()]=t.toLowerCase(),n["Webkit"+e]="webkit"+t,n["Moz"+e]="moz"+t,n}var _r={animationend:kr("Animation","AnimationEnd"),animationiteration:kr("Animation","AnimationIteration"),animationstart:kr("Animation","AnimationStart"),transitionend:kr("Transition","TransitionEnd")},Sr={},xr={};function Er(e){if(Sr[e])return Sr[e];if(!_r[e])return e;var t,n=_r[e];for(t in n)if(n.hasOwnProperty(t)&&t in xr)return Sr[e]=n[t];return e}c&&(xr=document.createElement("div").style,"AnimationEvent"in window||(delete _r.animationend.animation,delete _r.animationiteration.animation,delete _r.animationstart.animation),"TransitionEvent"in window||delete _r.transitionend.transition);var Cr=Er("animationend"),Nr=Er("animationiteration"),Pr=Er("animationstart"),zr=Er("transitionend"),Tr=new Map,Lr="abort auxClick cancel canPlay canPlayThrough click close contextMenu copy cut drag dragEnd dragEnter dragExit dragLeave dragOver dragStart drop durationChange emptied encrypted ended error gotPointerCapture input invalid keyDown keyPress keyUp load loadedData loadedMetadata loadStart lostPointerCapture mouseDown mouseMove mouseOut mouseOver mouseUp paste pause play playing pointerCancel pointerDown pointerMove pointerOut pointerOver pointerUp progress rateChange reset resize seeked seeking stalled submit suspend timeUpdate touchCancel touchEnd touchStart volumeChange scroll toggle touchMove waiting wheel".split(" ");function Mr(e,t){Tr.set(e,t),u(t,[e])}for(var Rr=0;Rr<Lr.length;Rr++){var Or=Lr[Rr];Mr(Or.toLowerCase(),"on"+(Or[0].toUpperCase()+Or.slice(1)))}Mr(Cr,"onAnimationEnd"),Mr(Nr,"onAnimationIteration"),Mr(Pr,"onAnimationStart"),Mr("dblclick","onDoubleClick"),Mr("focusin","onFocus"),Mr("focusout","onBlur"),Mr(zr,"onTransitionEnd"),s("onMouseEnter",["mouseout","mouseover"]),s("onMouseLeave",["mouseout","mouseover"]),s("onPointerEnter",["pointerout","pointerover"]),s("onPointerLeave",["pointerout","pointerover"]),u("onChange","change click focusin focusout input keydown keyup selectionchange".split(" ")),u("onSelect","focusout contextmenu dragend focusin keydown keyup mousedown mouseup selectionchange".split(" ")),u("onBeforeInput",["compositionend","keypress","textInput","paste"]),u("onCompositionEnd","compositionend focusout keydown keypress keyup mousedown".split(" ")),u("onCompositionStart","compositionstart focusout keydown keypress keyup mousedown".split(" ")),u("onCompositionUpdate","compositionupdate focusout keydown keypress keyup mousedown".split(" "));var Fr="abort canplay canplaythrough durationchange emptied encrypted ended error loadeddata loadedmetadata loadstart pause play playing progress ratechange resize seeked seeking stalled suspend timeupdate volumechange waiting".split(" "),Dr=new Set("cancel close invalid load scroll toggle".split(" ").concat(Fr));function Ir(e,t,n){var r=e.type||"unknown-event";e.currentTarget=n,function(e,t,n,r,a,i,o,u,s){if(je.apply(this,arguments),De){if(!De)throw Error(l(198));var c=Ie;De=!1,Ie=null,Ae||(Ae=!0,$e=c)}}(r,t,void 0,e),e.currentTarget=null}function Ar(e,t){t=0!==(4&t);for(var n=0;n<e.length;n++){var r=e[n],a=r.event;r=r.listeners;e:{var l=void 0;if(t)for(var i=r.length-1;0<=i;i--){var o=r[i],u=o.instance,s=o.currentTarget;if(o=o.listener,u!==l&&a.isPropagationStopped())break e;Ir(a,o,s),l=u}else for(i=0;i<r.length;i++){if(u=(o=r[i]).instance,s=o.currentTarget,o=o.listener,u!==l&&a.isPropagationStopped())break e;Ir(a,o,s),l=u}}}if(Ae)throw e=$e,Ae=!1,$e=null,e}function $r(e,t){var n=t[ma];void 0===n&&(n=t[ma]=new Set);var r=e+"__bubble";n.has(r)||(Br(t,e,2,!1),n.add(r))}function Ur(e,t,n){var r=0;t&&(r|=4),Br(n,e,r,t)}var jr="_reactListening"+Math.random().toString(36).slice(2);function Vr(e){if(!e[jr]){e[jr]=!0,i.forEach((function(t){"selectionchange"!==t&&(Dr.has(t)||Ur(t,!1,e),Ur(t,!0,e))}));var t=9===e.nodeType?e:e.ownerDocument;null===t||t[jr]||(t[jr]=!0,Ur("selectionchange",!1,t))}}function Br(e,t,n,r){switch(Kt(t)){case 1:var a=Wt;break;case 4:a=Qt;break;default:a=qt}n=a.bind(null,t,n,e),a=void 0,!Re||"touchstart"!==t&&"touchmove"!==t&&"wheel"!==t||(a=!0),r?void 0!==a?e.addEventListener(t,n,{capture:!0,passive:a}):e.addEventListener(t,n,!0):void 0!==a?e.addEventListener(t,n,{passive:a}):e.addEventListener(t,n,!1)}function Hr(e,t,n,r,a){var l=r;if(0===(1&t)&&0===(2&t)&&null!==r)e:for(;;){if(null===r)return;var i=r.tag;if(3===i||4===i){var o=r.stateNode.containerInfo;if(o===a||8===o.nodeType&&o.parentNode===a)break;if(4===i)for(i=r.return;null!==i;){var u=i.tag;if((3===u||4===u)&&((u=i.stateNode.containerInfo)===a||8===u.nodeType&&u.parentNode===a))return;i=i.return}for(;null!==o;){if(null===(i=ya(o)))return;if(5===(u=i.tag)||6===u){r=l=i;continue e}o=o.parentNode}}r=r.return}Le((function(){var r=l,a=ke(n),i=[];e:{var o=Tr.get(e);if(void 0!==o){var u=cn,s=e;switch(e){case"keypress":if(0===tn(n))break e;case"keydown":case"keyup":u=Nn;break;case"focusin":s="focus",u=vn;break;case"focusout":s="blur",u=vn;break;case"beforeblur":case"afterblur":u=vn;break;case"click":if(2===n.button)break e;case"auxclick":case"dblclick":case"mousedown":case"mousemove":case"mouseup":case"mouseout":case"mouseover":case"contextmenu":u=hn;break;case"drag":case"dragend":case"dragenter":case"dragexit":case"dragleave":case"dragover":case"dragstart":case"drop":u=mn;break;case"touchcancel":case"touchend":case"touchmove":case"touchstart":u=zn;break;case Cr:case Nr:case Pr:u=gn;break;case zr:u=Tn;break;case"scroll":u=dn;break;case"wheel":u=Mn;break;case"copy":case"cut":case"paste":u=bn;break;case"gotpointercapture":case"lostpointercapture":case"pointercancel":case"pointerdown":case"pointermove":case"pointerout":case"pointerover":case"pointerup":u=Pn}var c=0!==(4&t),f=!c&&"scroll"===e,d=c?null!==o?o+"Capture":null:o;c=[];for(var p,h=r;null!==h;){var m=(p=h).stateNode;if(5===p.tag&&null!==m&&(p=m,null!==d&&(null!=(m=Me(h,d))&&c.push(Wr(h,m,p)))),f)break;h=h.return}0<c.length&&(o=new u(o,s,null,n,a),i.push({event:o,listeners:c}))}}if(0===(7&t)){if(u="mouseout"===e||"pointerout"===e,(!(o="mouseover"===e||"pointerover"===e)||n===we||!(s=n.relatedTarget||n.fromElement)||!ya(s)&&!s[ha])&&(u||o)&&(o=a.window===a?a:(o=a.ownerDocument)?o.defaultView||o.parentWindow:window,u?(u=r,null!==(s=(s=n.relatedTarget||n.toElement)?ya(s):null)&&(s!==(f=Ve(s))||5!==s.tag&&6!==s.tag)&&(s=null)):(u=null,s=r),u!==s)){if(c=hn,m="onMouseLeave",d="onMouseEnter",h="mouse","pointerout"!==e&&"pointerover"!==e||(c=Pn,m="onPointerLeave",d="onPointerEnter",h="pointer"),f=null==u?o:wa(u),p=null==s?o:wa(s),(o=new c(m,h+"leave",u,n,a)).target=f,o.relatedTarget=p,m=null,ya(a)===r&&((c=new c(d,h+"enter",s,n,a)).target=p,c.relatedTarget=f,m=c),f=m,u&&s)e:{for(d=s,h=0,p=c=u;p;p=qr(p))h++;for(p=0,m=d;m;m=qr(m))p++;for(;0<h-p;)c=qr(c),h--;for(;0<p-h;)d=qr(d),p--;for(;h--;){if(c===d||null!==d&&c===d.alternate)break e;c=qr(c),d=qr(d)}c=null}else c=null;null!==u&&Xr(i,o,u,c,!1),null!==s&&null!==f&&Xr(i,f,s,c,!0)}if("select"===(u=(o=r?wa(r):window).nodeName&&o.nodeName.toLowerCase())||"input"===u&&"file"===o.type)var v=Kn;else if(Hn(o))if(Gn)v=ir;else{v=ar;var g=rr}else(u=o.nodeName)&&"input"===u.toLowerCase()&&("checkbox"===o.type||"radio"===o.type)&&(v=lr);switch(v&&(v=v(e,r))?Wn(i,v,n,a):(g&&g(e,o,r),"focusout"===e&&(g=o._wrapperState)&&g.controlled&&"number"===o.type&&ee(o,"number",o.value)),g=r?wa(r):window,e){case"focusin":(Hn(g)||"true"===g.contentEditable)&&(vr=g,gr=r,yr=null);break;case"focusout":yr=gr=vr=null;break;case"mousedown":br=!0;break;case"contextmenu":case"mouseup":case"dragend":br=!1,wr(i,n,a);break;case"selectionchange":if(mr)break;case"keydown":case"keyup":wr(i,n,a)}var y;if(On)e:{switch(e){case"compositionstart":var b="onCompositionStart";break e;case"compositionend":b="onCompositionEnd";break e;case"compositionupdate":b="onCompositionUpdate";break e}b=void 0}else Vn?Un(e,n)&&(b="onCompositionEnd"):"keydown"===e&&229===n.keyCode&&(b="onCompositionStart");b&&(In&&"ko"!==n.locale&&(Vn||"onCompositionStart"!==b?"onCompositionEnd"===b&&Vn&&(y=en()):(Zt="value"in(Gt=a)?Gt.value:Gt.textContent,Vn=!0)),0<(g=Qr(r,b)).length&&(b=new wn(b,e,null,n,a),i.push({event:b,listeners:g}),y?b.data=y:null!==(y=jn(n))&&(b.data=y))),(y=Dn?function(e,t){switch(e){case"compositionend":return jn(t);case"keypress":return 32!==t.which?null:($n=!0,An);case"textInput":return(e=t.data)===An&&$n?null:e;default:return null}}(e,n):function(e,t){if(Vn)return"compositionend"===e||!On&&Un(e,t)?(e=en(),Jt=Zt=Gt=null,Vn=!1,e):null;switch(e){case"paste":default:return null;case"keypress":if(!(t.ctrlKey||t.altKey||t.metaKey)||t.ctrlKey&&t.altKey){if(t.char&&1<t.char.length)return t.char;if(t.which)return String.fromCharCode(t.which)}return null;case"compositionend":return In&&"ko"!==t.locale?null:t.data}}(e,n))&&(0<(r=Qr(r,"onBeforeInput")).length&&(a=new wn("onBeforeInput","beforeinput",null,n,a),i.push({event:a,listeners:r}),a.data=y))}Ar(i,t)}))}function Wr(e,t,n){return{instance:e,listener:t,currentTarget:n}}function Qr(e,t){for(var n=t+"Capture",r=[];null!==e;){var a=e,l=a.stateNode;5===a.tag&&null!==l&&(a=l,null!=(l=Me(e,n))&&r.unshift(Wr(e,l,a)),null!=(l=Me(e,t))&&r.push(Wr(e,l,a))),e=e.return}return r}function qr(e){if(null===e)return null;do{e=e.return}while(e&&5!==e.tag);return e||null}function Xr(e,t,n,r,a){for(var l=t._reactName,i=[];null!==n&&n!==r;){var o=n,u=o.alternate,s=o.stateNode;if(null!==u&&u===r)break;5===o.tag&&null!==s&&(o=s,a?null!=(u=Me(n,l))&&i.unshift(Wr(n,u,o)):a||null!=(u=Me(n,l))&&i.push(Wr(n,u,o))),n=n.return}0!==i.length&&e.push({event:t,listeners:i})}var Yr=/\r\n?/g,Kr=/\u0000|\uFFFD/g;function Gr(e){return("string"===typeof e?e:""+e).replace(Yr,"\n").replace(Kr,"")}function Zr(e,t,n){if(t=Gr(t),Gr(e)!==t&&n)throw Error(l(425))}function Jr(){}var ea=null,ta=null;function na(e,t){return"textarea"===e||"noscript"===e||"string"===typeof t.children||"number"===typeof t.children||"object"===typeof t.dangerouslySetInnerHTML&&null!==t.dangerouslySetInnerHTML&&null!=t.dangerouslySetInnerHTML.__html}var ra="function"===typeof setTimeout?setTimeout:void 0,aa="function"===typeof clearTimeout?clearTimeout:void 0,la="function"===typeof Promise?Promise:void 0,ia="function"===typeof queueMicrotask?queueMicrotask:"undefined"!==typeof la?function(e){return la.resolve(null).then(e).catch(oa)}:ra;function oa(e){setTimeout((function(){throw e}))}function ua(e,t){var n=t,r=0;do{var a=n.nextSibling;if(e.removeChild(n),a&&8===a.nodeType)if("/$"===(n=a.data)){if(0===r)return e.removeChild(a),void Vt(t);r--}else"$"!==n&&"$?"!==n&&"$!"!==n||r++;n=a}while(n);Vt(t)}function sa(e){for(;null!=e;e=e.nextSibling){var t=e.nodeType;if(1===t||3===t)break;if(8===t){if("$"===(t=e.data)||"$!"===t||"$?"===t)break;if("/$"===t)return null}}return e}function ca(e){e=e.previousSibling;for(var t=0;e;){if(8===e.nodeType){var n=e.data;if("$"===n||"$!"===n||"$?"===n){if(0===t)return e;t--}else"/$"===n&&t++}e=e.previousSibling}return null}var fa=Math.random().toString(36).slice(2),da="__reactFiber$"+fa,pa="__reactProps$"+fa,ha="__reactContainer$"+fa,ma="__reactEvents$"+fa,va="__reactListeners$"+fa,ga="__reactHandles$"+fa;function ya(e){var t=e[da];if(t)return t;for(var n=e.parentNode;n;){if(t=n[ha]||n[da]){if(n=t.alternate,null!==t.child||null!==n&&null!==n.child)for(e=ca(e);null!==e;){if(n=e[da])return n;e=ca(e)}return t}n=(e=n).parentNode}return null}function ba(e){return!(e=e[da]||e[ha])||5!==e.tag&&6!==e.tag&&13!==e.tag&&3!==e.tag?null:e}function wa(e){if(5===e.tag||6===e.tag)return e.stateNode;throw Error(l(33))}function ka(e){return e[pa]||null}var _a=[],Sa=-1;function xa(e){return{current:e}}function Ea(e){0>Sa||(e.current=_a[Sa],_a[Sa]=null,Sa--)}function Ca(e,t){Sa++,_a[Sa]=e.current,e.current=t}var Na={},Pa=xa(Na),za=xa(!1),Ta=Na;function La(e,t){var n=e.type.contextTypes;if(!n)return Na;var r=e.stateNode;if(r&&r.__reactInternalMemoizedUnmaskedChildContext===t)return r.__reactInternalMemoizedMaskedChildContext;var a,l={};for(a in n)l[a]=t[a];return r&&((e=e.stateNode).__reactInternalMemoizedUnmaskedChildContext=t,e.__reactInternalMemoizedMaskedChildContext=l),l}function Ma(e){return null!==(e=e.childContextTypes)&&void 0!==e}function Ra(){Ea(za),Ea(Pa)}function Oa(e,t,n){if(Pa.current!==Na)throw Error(l(168));Ca(Pa,t),Ca(za,n)}function Fa(e,t,n){var r=e.stateNode;if(t=t.childContextTypes,"function"!==typeof r.getChildContext)return n;for(var a in r=r.getChildContext())if(!(a in t))throw Error(l(108,B(e)||"Unknown",a));return I({},n,r)}function Da(e){return e=(e=e.stateNode)&&e.__reactInternalMemoizedMergedChildContext||Na,Ta=Pa.current,Ca(Pa,e),Ca(za,za.current),!0}function Ia(e,t,n){var r=e.stateNode;if(!r)throw Error(l(169));n?(e=Fa(e,t,Ta),r.__reactInternalMemoizedMergedChildContext=e,Ea(za),Ea(Pa),Ca(Pa,e)):Ea(za),Ca(za,n)}var Aa=null,$a=!1,Ua=!1;function ja(e){null===Aa?Aa=[e]:Aa.push(e)}function Va(){if(!Ua&&null!==Aa){Ua=!0;var e=0,t=bt;try{var n=Aa;for(bt=1;e<n.length;e++){var r=n[e];do{r=r(!0)}while(null!==r)}Aa=null,$a=!1}catch(a){throw null!==Aa&&(Aa=Aa.slice(e+1)),qe(Je,Va),a}finally{bt=t,Ua=!1}}return null}var Ba=[],Ha=0,Wa=null,Qa=0,qa=[],Xa=0,Ya=null,Ka=1,Ga="";function Za(e,t){Ba[Ha++]=Qa,Ba[Ha++]=Wa,Wa=e,Qa=t}function Ja(e,t,n){qa[Xa++]=Ka,qa[Xa++]=Ga,qa[Xa++]=Ya,Ya=e;var r=Ka;e=Ga;var a=32-it(r)-1;r&=~(1<<a),n+=1;var l=32-it(t)+a;if(30<l){var i=a-a%5;l=(r&(1<<i)-1).toString(32),r>>=i,a-=i,Ka=1<<32-it(t)+a|n<<a|r,Ga=l+e}else Ka=1<<l|n<<a|r,Ga=e}function el(e){null!==e.return&&(Za(e,1),Ja(e,1,0))}function tl(e){for(;e===Wa;)Wa=Ba[--Ha],Ba[Ha]=null,Qa=Ba[--Ha],Ba[Ha]=null;for(;e===Ya;)Ya=qa[--Xa],qa[Xa]=null,Ga=qa[--Xa],qa[Xa]=null,Ka=qa[--Xa],qa[Xa]=null}var nl=null,rl=null,al=!1,ll=null;function il(e,t){var n=Ls(5,null,null,0);n.elementType="DELETED",n.stateNode=t,n.return=e,null===(t=e.deletions)?(e.deletions=[n],e.flags|=16):t.push(n)}function ol(e,t){switch(e.tag){case 5:var n=e.type;return null!==(t=1!==t.nodeType||n.toLowerCase()!==t.nodeName.toLowerCase()?null:t)&&(e.stateNode=t,nl=e,rl=sa(t.firstChild),!0);case 6:return null!==(t=""===e.pendingProps||3!==t.nodeType?null:t)&&(e.stateNode=t,nl=e,rl=null,!0);case 13:return null!==(t=8!==t.nodeType?null:t)&&(n=null!==Ya?{id:Ka,overflow:Ga}:null,e.memoizedState={dehydrated:t,treeContext:n,retryLane:1073741824},(n=Ls(18,null,null,0)).stateNode=t,n.return=e,e.child=n,nl=e,rl=null,!0);default:return!1}}function ul(e){return 0!==(1&e.mode)&&0===(128&e.flags)}function sl(e){if(al){var t=rl;if(t){var n=t;if(!ol(e,t)){if(ul(e))throw Error(l(418));t=sa(n.nextSibling);var r=nl;t&&ol(e,t)?il(r,n):(e.flags=-4097&e.flags|2,al=!1,nl=e)}}else{if(ul(e))throw Error(l(418));e.flags=-4097&e.flags|2,al=!1,nl=e}}}function cl(e){for(e=e.return;null!==e&&5!==e.tag&&3!==e.tag&&13!==e.tag;)e=e.return;nl=e}function fl(e){if(e!==nl)return!1;if(!al)return cl(e),al=!0,!1;var t;if((t=3!==e.tag)&&!(t=5!==e.tag)&&(t="head"!==(t=e.type)&&"body"!==t&&!na(e.type,e.memoizedProps)),t&&(t=rl)){if(ul(e))throw dl(),Error(l(418));for(;t;)il(e,t),t=sa(t.nextSibling)}if(cl(e),13===e.tag){if(!(e=null!==(e=e.memoizedState)?e.dehydrated:null))throw Error(l(317));e:{for(e=e.nextSibling,t=0;e;){if(8===e.nodeType){var n=e.data;if("/$"===n){if(0===t){rl=sa(e.nextSibling);break e}t--}else"$"!==n&&"$!"!==n&&"$?"!==n||t++}e=e.nextSibling}rl=null}}else rl=nl?sa(e.stateNode.nextSibling):null;return!0}function dl(){for(var e=rl;e;)e=sa(e.nextSibling)}function pl(){rl=nl=null,al=!1}function hl(e){null===ll?ll=[e]:ll.push(e)}var ml=w.ReactCurrentBatchConfig;function vl(e,t,n){if(null!==(e=n.ref)&&"function"!==typeof e&&"object"!==typeof e){if(n._owner){if(n=n._owner){if(1!==n.tag)throw Error(l(309));var r=n.stateNode}if(!r)throw Error(l(147,e));var a=r,i=""+e;return null!==t&&null!==t.ref&&"function"===typeof t.ref&&t.ref._stringRef===i?t.ref:(t=function(e){var t=a.refs;null===e?delete t[i]:t[i]=e},t._stringRef=i,t)}if("string"!==typeof e)throw Error(l(284));if(!n._owner)throw Error(l(290,e))}return e}function gl(e,t){throw e=Object.prototype.toString.call(t),Error(l(31,"[object Object]"===e?"object with keys {"+Object.keys(t).join(", ")+"}":e))}function yl(e){return(0,e._init)(e._payload)}function bl(e){function t(t,n){if(e){var r=t.deletions;null===r?(t.deletions=[n],t.flags|=16):r.push(n)}}function n(n,r){if(!e)return null;for(;null!==r;)t(n,r),r=r.sibling;return null}function r(e,t){for(e=new Map;null!==t;)null!==t.key?e.set(t.key,t):e.set(t.index,t),t=t.sibling;return e}function a(e,t){return(e=Rs(e,t)).index=0,e.sibling=null,e}function i(t,n,r){return t.index=r,e?null!==(r=t.alternate)?(r=r.index)<n?(t.flags|=2,n):r:(t.flags|=2,n):(t.flags|=1048576,n)}function o(t){return e&&null===t.alternate&&(t.flags|=2),t}function u(e,t,n,r){return null===t||6!==t.tag?((t=Is(n,e.mode,r)).return=e,t):((t=a(t,n)).return=e,t)}function s(e,t,n,r){var l=n.type;return l===S?f(e,t,n.props.children,r,n.key):null!==t&&(t.elementType===l||"object"===typeof l&&null!==l&&l.$$typeof===M&&yl(l)===t.type)?((r=a(t,n.props)).ref=vl(e,t,n),r.return=e,r):((r=Os(n.type,n.key,n.props,null,e.mode,r)).ref=vl(e,t,n),r.return=e,r)}function c(e,t,n,r){return null===t||4!==t.tag||t.stateNode.containerInfo!==n.containerInfo||t.stateNode.implementation!==n.implementation?((t=As(n,e.mode,r)).return=e,t):((t=a(t,n.children||[])).return=e,t)}function f(e,t,n,r,l){return null===t||7!==t.tag?((t=Fs(n,e.mode,r,l)).return=e,t):((t=a(t,n)).return=e,t)}function d(e,t,n){if("string"===typeof t&&""!==t||"number"===typeof t)return(t=Is(""+t,e.mode,n)).return=e,t;if("object"===typeof t&&null!==t){switch(t.$$typeof){case k:return(n=Os(t.type,t.key,t.props,null,e.mode,n)).ref=vl(e,null,t),n.return=e,n;case _:return(t=As(t,e.mode,n)).return=e,t;case M:return d(e,(0,t._init)(t._payload),n)}if(te(t)||F(t))return(t=Fs(t,e.mode,n,null)).return=e,t;gl(e,t)}return null}function p(e,t,n,r){var a=null!==t?t.key:null;if("string"===typeof n&&""!==n||"number"===typeof n)return null!==a?null:u(e,t,""+n,r);if("object"===typeof n&&null!==n){switch(n.$$typeof){case k:return n.key===a?s(e,t,n,r):null;case _:return n.key===a?c(e,t,n,r):null;case M:return p(e,t,(a=n._init)(n._payload),r)}if(te(n)||F(n))return null!==a?null:f(e,t,n,r,null);gl(e,n)}return null}function h(e,t,n,r,a){if("string"===typeof r&&""!==r||"number"===typeof r)return u(t,e=e.get(n)||null,""+r,a);if("object"===typeof r&&null!==r){switch(r.$$typeof){case k:return s(t,e=e.get(null===r.key?n:r.key)||null,r,a);case _:return c(t,e=e.get(null===r.key?n:r.key)||null,r,a);case M:return h(e,t,n,(0,r._init)(r._payload),a)}if(te(r)||F(r))return f(t,e=e.get(n)||null,r,a,null);gl(t,r)}return null}function m(a,l,o,u){for(var s=null,c=null,f=l,m=l=0,v=null;null!==f&&m<o.length;m++){f.index>m?(v=f,f=null):v=f.sibling;var g=p(a,f,o[m],u);if(null===g){null===f&&(f=v);break}e&&f&&null===g.alternate&&t(a,f),l=i(g,l,m),null===c?s=g:c.sibling=g,c=g,f=v}if(m===o.length)return n(a,f),al&&Za(a,m),s;if(null===f){for(;m<o.length;m++)null!==(f=d(a,o[m],u))&&(l=i(f,l,m),null===c?s=f:c.sibling=f,c=f);return al&&Za(a,m),s}for(f=r(a,f);m<o.length;m++)null!==(v=h(f,a,m,o[m],u))&&(e&&null!==v.alternate&&f.delete(null===v.key?m:v.key),l=i(v,l,m),null===c?s=v:c.sibling=v,c=v);return e&&f.forEach((function(e){return t(a,e)})),al&&Za(a,m),s}function v(a,o,u,s){var c=F(u);if("function"!==typeof c)throw Error(l(150));if(null==(u=c.call(u)))throw Error(l(151));for(var f=c=null,m=o,v=o=0,g=null,y=u.next();null!==m&&!y.done;v++,y=u.next()){m.index>v?(g=m,m=null):g=m.sibling;var b=p(a,m,y.value,s);if(null===b){null===m&&(m=g);break}e&&m&&null===b.alternate&&t(a,m),o=i(b,o,v),null===f?c=b:f.sibling=b,f=b,m=g}if(y.done)return n(a,m),al&&Za(a,v),c;if(null===m){for(;!y.done;v++,y=u.next())null!==(y=d(a,y.value,s))&&(o=i(y,o,v),null===f?c=y:f.sibling=y,f=y);return al&&Za(a,v),c}for(m=r(a,m);!y.done;v++,y=u.next())null!==(y=h(m,a,v,y.value,s))&&(e&&null!==y.alternate&&m.delete(null===y.key?v:y.key),o=i(y,o,v),null===f?c=y:f.sibling=y,f=y);return e&&m.forEach((function(e){return t(a,e)})),al&&Za(a,v),c}return function e(r,l,i,u){if("object"===typeof i&&null!==i&&i.type===S&&null===i.key&&(i=i.props.children),"object"===typeof i&&null!==i){switch(i.$$typeof){case k:e:{for(var s=i.key,c=l;null!==c;){if(c.key===s){if((s=i.type)===S){if(7===c.tag){n(r,c.sibling),(l=a(c,i.props.children)).return=r,r=l;break e}}else if(c.elementType===s||"object"===typeof s&&null!==s&&s.$$typeof===M&&yl(s)===c.type){n(r,c.sibling),(l=a(c,i.props)).ref=vl(r,c,i),l.return=r,r=l;break e}n(r,c);break}t(r,c),c=c.sibling}i.type===S?((l=Fs(i.props.children,r.mode,u,i.key)).return=r,r=l):((u=Os(i.type,i.key,i.props,null,r.mode,u)).ref=vl(r,l,i),u.return=r,r=u)}return o(r);case _:e:{for(c=i.key;null!==l;){if(l.key===c){if(4===l.tag&&l.stateNode.containerInfo===i.containerInfo&&l.stateNode.implementation===i.implementation){n(r,l.sibling),(l=a(l,i.children||[])).return=r,r=l;break e}n(r,l);break}t(r,l),l=l.sibling}(l=As(i,r.mode,u)).return=r,r=l}return o(r);case M:return e(r,l,(c=i._init)(i._payload),u)}if(te(i))return m(r,l,i,u);if(F(i))return v(r,l,i,u);gl(r,i)}return"string"===typeof i&&""!==i||"number"===typeof i?(i=""+i,null!==l&&6===l.tag?(n(r,l.sibling),(l=a(l,i)).return=r,r=l):(n(r,l),(l=Is(i,r.mode,u)).return=r,r=l),o(r)):n(r,l)}}var wl=bl(!0),kl=bl(!1),_l=xa(null),Sl=null,xl=null,El=null;function Cl(){El=xl=Sl=null}function Nl(e){var t=_l.current;Ea(_l),e._currentValue=t}function Pl(e,t,n){for(;null!==e;){var r=e.alternate;if((e.childLanes&t)!==t?(e.childLanes|=t,null!==r&&(r.childLanes|=t)):null!==r&&(r.childLanes&t)!==t&&(r.childLanes|=t),e===n)break;e=e.return}}function zl(e,t){Sl=e,El=xl=null,null!==(e=e.dependencies)&&null!==e.firstContext&&(0!==(e.lanes&t)&&(bo=!0),e.firstContext=null)}function Tl(e){var t=e._currentValue;if(El!==e)if(e={context:e,memoizedValue:t,next:null},null===xl){if(null===Sl)throw Error(l(308));xl=e,Sl.dependencies={lanes:0,firstContext:e}}else xl=xl.next=e;return t}var Ll=null;function Ml(e){null===Ll?Ll=[e]:Ll.push(e)}function Rl(e,t,n,r){var a=t.interleaved;return null===a?(n.next=n,Ml(t)):(n.next=a.next,a.next=n),t.interleaved=n,Ol(e,r)}function Ol(e,t){e.lanes|=t;var n=e.alternate;for(null!==n&&(n.lanes|=t),n=e,e=e.return;null!==e;)e.childLanes|=t,null!==(n=e.alternate)&&(n.childLanes|=t),n=e,e=e.return;return 3===n.tag?n.stateNode:null}var Fl=!1;function Dl(e){e.updateQueue={baseState:e.memoizedState,firstBaseUpdate:null,lastBaseUpdate:null,shared:{pending:null,interleaved:null,lanes:0},effects:null}}function Il(e,t){e=e.updateQueue,t.updateQueue===e&&(t.updateQueue={baseState:e.baseState,firstBaseUpdate:e.firstBaseUpdate,lastBaseUpdate:e.lastBaseUpdate,shared:e.shared,effects:e.effects})}function Al(e,t){return{eventTime:e,lane:t,tag:0,payload:null,callback:null,next:null}}function $l(e,t,n){var r=e.updateQueue;if(null===r)return null;if(r=r.shared,0!==(2&Pu)){var a=r.pending;return null===a?t.next=t:(t.next=a.next,a.next=t),r.pending=t,Ol(e,n)}return null===(a=r.interleaved)?(t.next=t,Ml(r)):(t.next=a.next,a.next=t),r.interleaved=t,Ol(e,n)}function Ul(e,t,n){if(null!==(t=t.updateQueue)&&(t=t.shared,0!==(4194240&n))){var r=t.lanes;n|=r&=e.pendingLanes,t.lanes=n,yt(e,n)}}function jl(e,t){var n=e.updateQueue,r=e.alternate;if(null!==r&&n===(r=r.updateQueue)){var a=null,l=null;if(null!==(n=n.firstBaseUpdate)){do{var i={eventTime:n.eventTime,lane:n.lane,tag:n.tag,payload:n.payload,callback:n.callback,next:null};null===l?a=l=i:l=l.next=i,n=n.next}while(null!==n);null===l?a=l=t:l=l.next=t}else a=l=t;return n={baseState:r.baseState,firstBaseUpdate:a,lastBaseUpdate:l,shared:r.shared,effects:r.effects},void(e.updateQueue=n)}null===(e=n.lastBaseUpdate)?n.firstBaseUpdate=t:e.next=t,n.lastBaseUpdate=t}function Vl(e,t,n,r){var a=e.updateQueue;Fl=!1;var l=a.firstBaseUpdate,i=a.lastBaseUpdate,o=a.shared.pending;if(null!==o){a.shared.pending=null;var u=o,s=u.next;u.next=null,null===i?l=s:i.next=s,i=u;var c=e.alternate;null!==c&&((o=(c=c.updateQueue).lastBaseUpdate)!==i&&(null===o?c.firstBaseUpdate=s:o.next=s,c.lastBaseUpdate=u))}if(null!==l){var f=a.baseState;for(i=0,c=s=u=null,o=l;;){var d=o.lane,p=o.eventTime;if((r&d)===d){null!==c&&(c=c.next={eventTime:p,lane:0,tag:o.tag,payload:o.payload,callback:o.callback,next:null});e:{var h=e,m=o;switch(d=t,p=n,m.tag){case 1:if("function"===typeof(h=m.payload)){f=h.call(p,f,d);break e}f=h;break e;case 3:h.flags=-65537&h.flags|128;case 0:if(null===(d="function"===typeof(h=m.payload)?h.call(p,f,d):h)||void 0===d)break e;f=I({},f,d);break e;case 2:Fl=!0}}null!==o.callback&&0!==o.lane&&(e.flags|=64,null===(d=a.effects)?a.effects=[o]:d.push(o))}else p={eventTime:p,lane:d,tag:o.tag,payload:o.payload,callback:o.callback,next:null},null===c?(s=c=p,u=f):c=c.next=p,i|=d;if(null===(o=o.next)){if(null===(o=a.shared.pending))break;o=(d=o).next,d.next=null,a.lastBaseUpdate=d,a.shared.pending=null}}if(null===c&&(u=f),a.baseState=u,a.firstBaseUpdate=s,a.lastBaseUpdate=c,null!==(t=a.shared.interleaved)){a=t;do{i|=a.lane,a=a.next}while(a!==t)}else null===l&&(a.shared.lanes=0);Du|=i,e.lanes=i,e.memoizedState=f}}function Bl(e,t,n){if(e=t.effects,t.effects=null,null!==e)for(t=0;t<e.length;t++){var r=e[t],a=r.callback;if(null!==a){if(r.callback=null,r=n,"function"!==typeof a)throw Error(l(191,a));a.call(r)}}}var Hl={},Wl=xa(Hl),Ql=xa(Hl),ql=xa(Hl);function Xl(e){if(e===Hl)throw Error(l(174));return e}function Yl(e,t){switch(Ca(ql,t),Ca(Ql,e),Ca(Wl,Hl),e=t.nodeType){case 9:case 11:t=(t=t.documentElement)?t.namespaceURI:ue(null,"");break;default:t=ue(t=(e=8===e?t.parentNode:t).namespaceURI||null,e=e.tagName)}Ea(Wl),Ca(Wl,t)}function Kl(){Ea(Wl),Ea(Ql),Ea(ql)}function Gl(e){Xl(ql.current);var t=Xl(Wl.current),n=ue(t,e.type);t!==n&&(Ca(Ql,e),Ca(Wl,n))}function Zl(e){Ql.current===e&&(Ea(Wl),Ea(Ql))}var Jl=xa(0);function ei(e){for(var t=e;null!==t;){if(13===t.tag){var n=t.memoizedState;if(null!==n&&(null===(n=n.dehydrated)||"$?"===n.data||"$!"===n.data))return t}else if(19===t.tag&&void 0!==t.memoizedProps.revealOrder){if(0!==(128&t.flags))return t}else if(null!==t.child){t.child.return=t,t=t.child;continue}if(t===e)break;for(;null===t.sibling;){if(null===t.return||t.return===e)return null;t=t.return}t.sibling.return=t.return,t=t.sibling}return null}var ti=[];function ni(){for(var e=0;e<ti.length;e++)ti[e]._workInProgressVersionPrimary=null;ti.length=0}var ri=w.ReactCurrentDispatcher,ai=w.ReactCurrentBatchConfig,li=0,ii=null,oi=null,ui=null,si=!1,ci=!1,fi=0,di=0;function pi(){throw Error(l(321))}function hi(e,t){if(null===t)return!1;for(var n=0;n<t.length&&n<e.length;n++)if(!or(e[n],t[n]))return!1;return!0}function mi(e,t,n,r,a,i){if(li=i,ii=t,t.memoizedState=null,t.updateQueue=null,t.lanes=0,ri.current=null===e||null===e.memoizedState?Zi:Ji,e=n(r,a),ci){i=0;do{if(ci=!1,fi=0,25<=i)throw Error(l(301));i+=1,ui=oi=null,t.updateQueue=null,ri.current=eo,e=n(r,a)}while(ci)}if(ri.current=Gi,t=null!==oi&&null!==oi.next,li=0,ui=oi=ii=null,si=!1,t)throw Error(l(300));return e}function vi(){var e=0!==fi;return fi=0,e}function gi(){var e={memoizedState:null,baseState:null,baseQueue:null,queue:null,next:null};return null===ui?ii.memoizedState=ui=e:ui=ui.next=e,ui}function yi(){if(null===oi){var e=ii.alternate;e=null!==e?e.memoizedState:null}else e=oi.next;var t=null===ui?ii.memoizedState:ui.next;if(null!==t)ui=t,oi=e;else{if(null===e)throw Error(l(310));e={memoizedState:(oi=e).memoizedState,baseState:oi.baseState,baseQueue:oi.baseQueue,queue:oi.queue,next:null},null===ui?ii.memoizedState=ui=e:ui=ui.next=e}return ui}function bi(e,t){return"function"===typeof t?t(e):t}function wi(e){var t=yi(),n=t.queue;if(null===n)throw Error(l(311));n.lastRenderedReducer=e;var r=oi,a=r.baseQueue,i=n.pending;if(null!==i){if(null!==a){var o=a.next;a.next=i.next,i.next=o}r.baseQueue=a=i,n.pending=null}if(null!==a){i=a.next,r=r.baseState;var u=o=null,s=null,c=i;do{var f=c.lane;if((li&f)===f)null!==s&&(s=s.next={lane:0,action:c.action,hasEagerState:c.hasEagerState,eagerState:c.eagerState,next:null}),r=c.hasEagerState?c.eagerState:e(r,c.action);else{var d={lane:f,action:c.action,hasEagerState:c.hasEagerState,eagerState:c.eagerState,next:null};null===s?(u=s=d,o=r):s=s.next=d,ii.lanes|=f,Du|=f}c=c.next}while(null!==c&&c!==i);null===s?o=r:s.next=u,or(r,t.memoizedState)||(bo=!0),t.memoizedState=r,t.baseState=o,t.baseQueue=s,n.lastRenderedState=r}if(null!==(e=n.interleaved)){a=e;do{i=a.lane,ii.lanes|=i,Du|=i,a=a.next}while(a!==e)}else null===a&&(n.lanes=0);return[t.memoizedState,n.dispatch]}function ki(e){var t=yi(),n=t.queue;if(null===n)throw Error(l(311));n.lastRenderedReducer=e;var r=n.dispatch,a=n.pending,i=t.memoizedState;if(null!==a){n.pending=null;var o=a=a.next;do{i=e(i,o.action),o=o.next}while(o!==a);or(i,t.memoizedState)||(bo=!0),t.memoizedState=i,null===t.baseQueue&&(t.baseState=i),n.lastRenderedState=i}return[i,r]}function _i(){}function Si(e,t){var n=ii,r=yi(),a=t(),i=!or(r.memoizedState,a);if(i&&(r.memoizedState=a,bo=!0),r=r.queue,Fi(Ci.bind(null,n,r,e),[e]),r.getSnapshot!==t||i||null!==ui&&1&ui.memoizedState.tag){if(n.flags|=2048,Ti(9,Ei.bind(null,n,r,a,t),void 0,null),null===zu)throw Error(l(349));0!==(30&li)||xi(n,t,a)}return a}function xi(e,t,n){e.flags|=16384,e={getSnapshot:t,value:n},null===(t=ii.updateQueue)?(t={lastEffect:null,stores:null},ii.updateQueue=t,t.stores=[e]):null===(n=t.stores)?t.stores=[e]:n.push(e)}function Ei(e,t,n,r){t.value=n,t.getSnapshot=r,Ni(t)&&Pi(e)}function Ci(e,t,n){return n((function(){Ni(t)&&Pi(e)}))}function Ni(e){var t=e.getSnapshot;e=e.value;try{var n=t();return!or(e,n)}catch(r){return!0}}function Pi(e){var t=Ol(e,1);null!==t&&ns(t,e,1,-1)}function zi(e){var t=gi();return"function"===typeof e&&(e=e()),t.memoizedState=t.baseState=e,e={pending:null,interleaved:null,lanes:0,dispatch:null,lastRenderedReducer:bi,lastRenderedState:e},t.queue=e,e=e.dispatch=qi.bind(null,ii,e),[t.memoizedState,e]}function Ti(e,t,n,r){return e={tag:e,create:t,destroy:n,deps:r,next:null},null===(t=ii.updateQueue)?(t={lastEffect:null,stores:null},ii.updateQueue=t,t.lastEffect=e.next=e):null===(n=t.lastEffect)?t.lastEffect=e.next=e:(r=n.next,n.next=e,e.next=r,t.lastEffect=e),e}function Li(){return yi().memoizedState}function Mi(e,t,n,r){var a=gi();ii.flags|=e,a.memoizedState=Ti(1|t,n,void 0,void 0===r?null:r)}function Ri(e,t,n,r){var a=yi();r=void 0===r?null:r;var l=void 0;if(null!==oi){var i=oi.memoizedState;if(l=i.destroy,null!==r&&hi(r,i.deps))return void(a.memoizedState=Ti(t,n,l,r))}ii.flags|=e,a.memoizedState=Ti(1|t,n,l,r)}function Oi(e,t){return Mi(8390656,8,e,t)}function Fi(e,t){return Ri(2048,8,e,t)}function Di(e,t){return Ri(4,2,e,t)}function Ii(e,t){return Ri(4,4,e,t)}function Ai(e,t){return"function"===typeof t?(e=e(),t(e),function(){t(null)}):null!==t&&void 0!==t?(e=e(),t.current=e,function(){t.current=null}):void 0}function $i(e,t,n){return n=null!==n&&void 0!==n?n.concat([e]):null,Ri(4,4,Ai.bind(null,t,e),n)}function Ui(){}function ji(e,t){var n=yi();t=void 0===t?null:t;var r=n.memoizedState;return null!==r&&null!==t&&hi(t,r[1])?r[0]:(n.memoizedState=[e,t],e)}function Vi(e,t){var n=yi();t=void 0===t?null:t;var r=n.memoizedState;return null!==r&&null!==t&&hi(t,r[1])?r[0]:(e=e(),n.memoizedState=[e,t],e)}function Bi(e,t,n){return 0===(21&li)?(e.baseState&&(e.baseState=!1,bo=!0),e.memoizedState=n):(or(n,t)||(n=mt(),ii.lanes|=n,Du|=n,e.baseState=!0),t)}function Hi(e,t){var n=bt;bt=0!==n&&4>n?n:4,e(!0);var r=ai.transition;ai.transition={};try{e(!1),t()}finally{bt=n,ai.transition=r}}function Wi(){return yi().memoizedState}function Qi(e,t,n){var r=ts(e);if(n={lane:r,action:n,hasEagerState:!1,eagerState:null,next:null},Xi(e))Yi(t,n);else if(null!==(n=Rl(e,t,n,r))){ns(n,e,r,es()),Ki(n,t,r)}}function qi(e,t,n){var r=ts(e),a={lane:r,action:n,hasEagerState:!1,eagerState:null,next:null};if(Xi(e))Yi(t,a);else{var l=e.alternate;if(0===e.lanes&&(null===l||0===l.lanes)&&null!==(l=t.lastRenderedReducer))try{var i=t.lastRenderedState,o=l(i,n);if(a.hasEagerState=!0,a.eagerState=o,or(o,i)){var u=t.interleaved;return null===u?(a.next=a,Ml(t)):(a.next=u.next,u.next=a),void(t.interleaved=a)}}catch(s){}null!==(n=Rl(e,t,a,r))&&(ns(n,e,r,a=es()),Ki(n,t,r))}}function Xi(e){var t=e.alternate;return e===ii||null!==t&&t===ii}function Yi(e,t){ci=si=!0;var n=e.pending;null===n?t.next=t:(t.next=n.next,n.next=t),e.pending=t}function Ki(e,t,n){if(0!==(4194240&n)){var r=t.lanes;n|=r&=e.pendingLanes,t.lanes=n,yt(e,n)}}var Gi={readContext:Tl,useCallback:pi,useContext:pi,useEffect:pi,useImperativeHandle:pi,useInsertionEffect:pi,useLayoutEffect:pi,useMemo:pi,useReducer:pi,useRef:pi,useState:pi,useDebugValue:pi,useDeferredValue:pi,useTransition:pi,useMutableSource:pi,useSyncExternalStore:pi,useId:pi,unstable_isNewReconciler:!1},Zi={readContext:Tl,useCallback:function(e,t){return gi().memoizedState=[e,void 0===t?null:t],e},useContext:Tl,useEffect:Oi,useImperativeHandle:function(e,t,n){return n=null!==n&&void 0!==n?n.concat([e]):null,Mi(4194308,4,Ai.bind(null,t,e),n)},useLayoutEffect:function(e,t){return Mi(4194308,4,e,t)},useInsertionEffect:function(e,t){return Mi(4,2,e,t)},useMemo:function(e,t){var n=gi();return t=void 0===t?null:t,e=e(),n.memoizedState=[e,t],e},useReducer:function(e,t,n){var r=gi();return t=void 0!==n?n(t):t,r.memoizedState=r.baseState=t,e={pending:null,interleaved:null,lanes:0,dispatch:null,lastRenderedReducer:e,lastRenderedState:t},r.queue=e,e=e.dispatch=Qi.bind(null,ii,e),[r.memoizedState,e]},useRef:function(e){return e={current:e},gi().memoizedState=e},useState:zi,useDebugValue:Ui,useDeferredValue:function(e){return gi().memoizedState=e},useTransition:function(){var e=zi(!1),t=e[0];return e=Hi.bind(null,e[1]),gi().memoizedState=e,[t,e]},useMutableSource:function(){},useSyncExternalStore:function(e,t,n){var r=ii,a=gi();if(al){if(void 0===n)throw Error(l(407));n=n()}else{if(n=t(),null===zu)throw Error(l(349));0!==(30&li)||xi(r,t,n)}a.memoizedState=n;var i={value:n,getSnapshot:t};return a.queue=i,Oi(Ci.bind(null,r,i,e),[e]),r.flags|=2048,Ti(9,Ei.bind(null,r,i,n,t),void 0,null),n},useId:function(){var e=gi(),t=zu.identifierPrefix;if(al){var n=Ga;t=":"+t+"R"+(n=(Ka&~(1<<32-it(Ka)-1)).toString(32)+n),0<(n=fi++)&&(t+="H"+n.toString(32)),t+=":"}else t=":"+t+"r"+(n=di++).toString(32)+":";return e.memoizedState=t},unstable_isNewReconciler:!1},Ji={readContext:Tl,useCallback:ji,useContext:Tl,useEffect:Fi,useImperativeHandle:$i,useInsertionEffect:Di,useLayoutEffect:Ii,useMemo:Vi,useReducer:wi,useRef:Li,useState:function(){return wi(bi)},useDebugValue:Ui,useDeferredValue:function(e){return Bi(yi(),oi.memoizedState,e)},useTransition:function(){return[wi(bi)[0],yi().memoizedState]},useMutableSource:_i,useSyncExternalStore:Si,useId:Wi,unstable_isNewReconciler:!1},eo={readContext:Tl,useCallback:ji,useContext:Tl,useEffect:Fi,useImperativeHandle:$i,useInsertionEffect:Di,useLayoutEffect:Ii,useMemo:Vi,useReducer:ki,useRef:Li,useState:function(){return ki(bi)},useDebugValue:Ui,useDeferredValue:function(e){var t=yi();return null===oi?t.memoizedState=e:Bi(t,oi.memoizedState,e)},useTransition:function(){return[ki(bi)[0],yi().memoizedState]},useMutableSource:_i,useSyncExternalStore:Si,useId:Wi,unstable_isNewReconciler:!1};function to(e,t){if(e&&e.defaultProps){for(var n in t=I({},t),e=e.defaultProps)void 0===t[n]&&(t[n]=e[n]);return t}return t}function no(e,t,n,r){n=null===(n=n(r,t=e.memoizedState))||void 0===n?t:I({},t,n),e.memoizedState=n,0===e.lanes&&(e.updateQueue.baseState=n)}var ro={isMounted:function(e){return!!(e=e._reactInternals)&&Ve(e)===e},enqueueSetState:function(e,t,n){e=e._reactInternals;var r=es(),a=ts(e),l=Al(r,a);l.payload=t,void 0!==n&&null!==n&&(l.callback=n),null!==(t=$l(e,l,a))&&(ns(t,e,a,r),Ul(t,e,a))},enqueueReplaceState:function(e,t,n){e=e._reactInternals;var r=es(),a=ts(e),l=Al(r,a);l.tag=1,l.payload=t,void 0!==n&&null!==n&&(l.callback=n),null!==(t=$l(e,l,a))&&(ns(t,e,a,r),Ul(t,e,a))},enqueueForceUpdate:function(e,t){e=e._reactInternals;var n=es(),r=ts(e),a=Al(n,r);a.tag=2,void 0!==t&&null!==t&&(a.callback=t),null!==(t=$l(e,a,r))&&(ns(t,e,r,n),Ul(t,e,r))}};function ao(e,t,n,r,a,l,i){return"function"===typeof(e=e.stateNode).shouldComponentUpdate?e.shouldComponentUpdate(r,l,i):!t.prototype||!t.prototype.isPureReactComponent||(!ur(n,r)||!ur(a,l))}function lo(e,t,n){var r=!1,a=Na,l=t.contextType;return"object"===typeof l&&null!==l?l=Tl(l):(a=Ma(t)?Ta:Pa.current,l=(r=null!==(r=t.contextTypes)&&void 0!==r)?La(e,a):Na),t=new t(n,l),e.memoizedState=null!==t.state&&void 0!==t.state?t.state:null,t.updater=ro,e.stateNode=t,t._reactInternals=e,r&&((e=e.stateNode).__reactInternalMemoizedUnmaskedChildContext=a,e.__reactInternalMemoizedMaskedChildContext=l),t}function io(e,t,n,r){e=t.state,"function"===typeof t.componentWillReceiveProps&&t.componentWillReceiveProps(n,r),"function"===typeof t.UNSAFE_componentWillReceiveProps&&t.UNSAFE_componentWillReceiveProps(n,r),t.state!==e&&ro.enqueueReplaceState(t,t.state,null)}function oo(e,t,n,r){var a=e.stateNode;a.props=n,a.state=e.memoizedState,a.refs={},Dl(e);var l=t.contextType;"object"===typeof l&&null!==l?a.context=Tl(l):(l=Ma(t)?Ta:Pa.current,a.context=La(e,l)),a.state=e.memoizedState,"function"===typeof(l=t.getDerivedStateFromProps)&&(no(e,t,l,n),a.state=e.memoizedState),"function"===typeof t.getDerivedStateFromProps||"function"===typeof a.getSnapshotBeforeUpdate||"function"!==typeof a.UNSAFE_componentWillMount&&"function"!==typeof a.componentWillMount||(t=a.state,"function"===typeof a.componentWillMount&&a.componentWillMount(),"function"===typeof a.UNSAFE_componentWillMount&&a.UNSAFE_componentWillMount(),t!==a.state&&ro.enqueueReplaceState(a,a.state,null),Vl(e,n,a,r),a.state=e.memoizedState),"function"===typeof a.componentDidMount&&(e.flags|=4194308)}function uo(e,t){try{var n="",r=t;do{n+=j(r),r=r.return}while(r);var a=n}catch(l){a="\nError generating stack: "+l.message+"\n"+l.stack}return{value:e,source:t,stack:a,digest:null}}function so(e,t,n){return{value:e,source:null,stack:null!=n?n:null,digest:null!=t?t:null}}function co(e,t){try{console.error(t.value)}catch(n){setTimeout((function(){throw n}))}}var fo="function"===typeof WeakMap?WeakMap:Map;function po(e,t,n){(n=Al(-1,n)).tag=3,n.payload={element:null};var r=t.value;return n.callback=function(){Hu||(Hu=!0,Wu=r),co(0,t)},n}function ho(e,t,n){(n=Al(-1,n)).tag=3;var r=e.type.getDerivedStateFromError;if("function"===typeof r){var a=t.value;n.payload=function(){return r(a)},n.callback=function(){co(0,t)}}var l=e.stateNode;return null!==l&&"function"===typeof l.componentDidCatch&&(n.callback=function(){co(0,t),"function"!==typeof r&&(null===Qu?Qu=new Set([this]):Qu.add(this));var e=t.stack;this.componentDidCatch(t.value,{componentStack:null!==e?e:""})}),n}function mo(e,t,n){var r=e.pingCache;if(null===r){r=e.pingCache=new fo;var a=new Set;r.set(t,a)}else void 0===(a=r.get(t))&&(a=new Set,r.set(t,a));a.has(n)||(a.add(n),e=Es.bind(null,e,t,n),t.then(e,e))}function vo(e){do{var t;if((t=13===e.tag)&&(t=null===(t=e.memoizedState)||null!==t.dehydrated),t)return e;e=e.return}while(null!==e);return null}function go(e,t,n,r,a){return 0===(1&e.mode)?(e===t?e.flags|=65536:(e.flags|=128,n.flags|=131072,n.flags&=-52805,1===n.tag&&(null===n.alternate?n.tag=17:((t=Al(-1,1)).tag=2,$l(n,t,1))),n.lanes|=1),e):(e.flags|=65536,e.lanes=a,e)}var yo=w.ReactCurrentOwner,bo=!1;function wo(e,t,n,r){t.child=null===e?kl(t,null,n,r):wl(t,e.child,n,r)}function ko(e,t,n,r,a){n=n.render;var l=t.ref;return zl(t,a),r=mi(e,t,n,r,l,a),n=vi(),null===e||bo?(al&&n&&el(t),t.flags|=1,wo(e,t,r,a),t.child):(t.updateQueue=e.updateQueue,t.flags&=-2053,e.lanes&=~a,Ho(e,t,a))}function _o(e,t,n,r,a){if(null===e){var l=n.type;return"function"!==typeof l||Ms(l)||void 0!==l.defaultProps||null!==n.compare||void 0!==n.defaultProps?((e=Os(n.type,null,r,t,t.mode,a)).ref=t.ref,e.return=t,t.child=e):(t.tag=15,t.type=l,So(e,t,l,r,a))}if(l=e.child,0===(e.lanes&a)){var i=l.memoizedProps;if((n=null!==(n=n.compare)?n:ur)(i,r)&&e.ref===t.ref)return Ho(e,t,a)}return t.flags|=1,(e=Rs(l,r)).ref=t.ref,e.return=t,t.child=e}function So(e,t,n,r,a){if(null!==e){var l=e.memoizedProps;if(ur(l,r)&&e.ref===t.ref){if(bo=!1,t.pendingProps=r=l,0===(e.lanes&a))return t.lanes=e.lanes,Ho(e,t,a);0!==(131072&e.flags)&&(bo=!0)}}return Co(e,t,n,r,a)}function xo(e,t,n){var r=t.pendingProps,a=r.children,l=null!==e?e.memoizedState:null;if("hidden"===r.mode)if(0===(1&t.mode))t.memoizedState={baseLanes:0,cachePool:null,transitions:null},Ca(Ru,Mu),Mu|=n;else{if(0===(1073741824&n))return e=null!==l?l.baseLanes|n:n,t.lanes=t.childLanes=1073741824,t.memoizedState={baseLanes:e,cachePool:null,transitions:null},t.updateQueue=null,Ca(Ru,Mu),Mu|=e,null;t.memoizedState={baseLanes:0,cachePool:null,transitions:null},r=null!==l?l.baseLanes:n,Ca(Ru,Mu),Mu|=r}else null!==l?(r=l.baseLanes|n,t.memoizedState=null):r=n,Ca(Ru,Mu),Mu|=r;return wo(e,t,a,n),t.child}function Eo(e,t){var n=t.ref;(null===e&&null!==n||null!==e&&e.ref!==n)&&(t.flags|=512,t.flags|=2097152)}function Co(e,t,n,r,a){var l=Ma(n)?Ta:Pa.current;return l=La(t,l),zl(t,a),n=mi(e,t,n,r,l,a),r=vi(),null===e||bo?(al&&r&&el(t),t.flags|=1,wo(e,t,n,a),t.child):(t.updateQueue=e.updateQueue,t.flags&=-2053,e.lanes&=~a,Ho(e,t,a))}function No(e,t,n,r,a){if(Ma(n)){var l=!0;Da(t)}else l=!1;if(zl(t,a),null===t.stateNode)Bo(e,t),lo(t,n,r),oo(t,n,r,a),r=!0;else if(null===e){var i=t.stateNode,o=t.memoizedProps;i.props=o;var u=i.context,s=n.contextType;"object"===typeof s&&null!==s?s=Tl(s):s=La(t,s=Ma(n)?Ta:Pa.current);var c=n.getDerivedStateFromProps,f="function"===typeof c||"function"===typeof i.getSnapshotBeforeUpdate;f||"function"!==typeof i.UNSAFE_componentWillReceiveProps&&"function"!==typeof i.componentWillReceiveProps||(o!==r||u!==s)&&io(t,i,r,s),Fl=!1;var d=t.memoizedState;i.state=d,Vl(t,r,i,a),u=t.memoizedState,o!==r||d!==u||za.current||Fl?("function"===typeof c&&(no(t,n,c,r),u=t.memoizedState),(o=Fl||ao(t,n,o,r,d,u,s))?(f||"function"!==typeof i.UNSAFE_componentWillMount&&"function"!==typeof i.componentWillMount||("function"===typeof i.componentWillMount&&i.componentWillMount(),"function"===typeof i.UNSAFE_componentWillMount&&i.UNSAFE_componentWillMount()),"function"===typeof i.componentDidMount&&(t.flags|=4194308)):("function"===typeof i.componentDidMount&&(t.flags|=4194308),t.memoizedProps=r,t.memoizedState=u),i.props=r,i.state=u,i.context=s,r=o):("function"===typeof i.componentDidMount&&(t.flags|=4194308),r=!1)}else{i=t.stateNode,Il(e,t),o=t.memoizedProps,s=t.type===t.elementType?o:to(t.type,o),i.props=s,f=t.pendingProps,d=i.context,"object"===typeof(u=n.contextType)&&null!==u?u=Tl(u):u=La(t,u=Ma(n)?Ta:Pa.current);var p=n.getDerivedStateFromProps;(c="function"===typeof p||"function"===typeof i.getSnapshotBeforeUpdate)||"function"!==typeof i.UNSAFE_componentWillReceiveProps&&"function"!==typeof i.componentWillReceiveProps||(o!==f||d!==u)&&io(t,i,r,u),Fl=!1,d=t.memoizedState,i.state=d,Vl(t,r,i,a);var h=t.memoizedState;o!==f||d!==h||za.current||Fl?("function"===typeof p&&(no(t,n,p,r),h=t.memoizedState),(s=Fl||ao(t,n,s,r,d,h,u)||!1)?(c||"function"!==typeof i.UNSAFE_componentWillUpdate&&"function"!==typeof i.componentWillUpdate||("function"===typeof i.componentWillUpdate&&i.componentWillUpdate(r,h,u),"function"===typeof i.UNSAFE_componentWillUpdate&&i.UNSAFE_componentWillUpdate(r,h,u)),"function"===typeof i.componentDidUpdate&&(t.flags|=4),"function"===typeof i.getSnapshotBeforeUpdate&&(t.flags|=1024)):("function"!==typeof i.componentDidUpdate||o===e.memoizedProps&&d===e.memoizedState||(t.flags|=4),"function"!==typeof i.getSnapshotBeforeUpdate||o===e.memoizedProps&&d===e.memoizedState||(t.flags|=1024),t.memoizedProps=r,t.memoizedState=h),i.props=r,i.state=h,i.context=u,r=s):("function"!==typeof i.componentDidUpdate||o===e.memoizedProps&&d===e.memoizedState||(t.flags|=4),"function"!==typeof i.getSnapshotBeforeUpdate||o===e.memoizedProps&&d===e.memoizedState||(t.flags|=1024),r=!1)}return Po(e,t,n,r,l,a)}function Po(e,t,n,r,a,l){Eo(e,t);var i=0!==(128&t.flags);if(!r&&!i)return a&&Ia(t,n,!1),Ho(e,t,l);r=t.stateNode,yo.current=t;var o=i&&"function"!==typeof n.getDerivedStateFromError?null:r.render();return t.flags|=1,null!==e&&i?(t.child=wl(t,e.child,null,l),t.child=wl(t,null,o,l)):wo(e,t,o,l),t.memoizedState=r.state,a&&Ia(t,n,!0),t.child}function zo(e){var t=e.stateNode;t.pendingContext?Oa(0,t.pendingContext,t.pendingContext!==t.context):t.context&&Oa(0,t.context,!1),Yl(e,t.containerInfo)}function To(e,t,n,r,a){return pl(),hl(a),t.flags|=256,wo(e,t,n,r),t.child}var Lo,Mo,Ro,Oo,Fo={dehydrated:null,treeContext:null,retryLane:0};function Do(e){return{baseLanes:e,cachePool:null,transitions:null}}function Io(e,t,n){var r,a=t.pendingProps,i=Jl.current,o=!1,u=0!==(128&t.flags);if((r=u)||(r=(null===e||null!==e.memoizedState)&&0!==(2&i)),r?(o=!0,t.flags&=-129):null!==e&&null===e.memoizedState||(i|=1),Ca(Jl,1&i),null===e)return sl(t),null!==(e=t.memoizedState)&&null!==(e=e.dehydrated)?(0===(1&t.mode)?t.lanes=1:"$!"===e.data?t.lanes=8:t.lanes=1073741824,null):(u=a.children,e=a.fallback,o?(a=t.mode,o=t.child,u={mode:"hidden",children:u},0===(1&a)&&null!==o?(o.childLanes=0,o.pendingProps=u):o=Ds(u,a,0,null),e=Fs(e,a,n,null),o.return=t,e.return=t,o.sibling=e,t.child=o,t.child.memoizedState=Do(n),t.memoizedState=Fo,e):Ao(t,u));if(null!==(i=e.memoizedState)&&null!==(r=i.dehydrated))return function(e,t,n,r,a,i,o){if(n)return 256&t.flags?(t.flags&=-257,$o(e,t,o,r=so(Error(l(422))))):null!==t.memoizedState?(t.child=e.child,t.flags|=128,null):(i=r.fallback,a=t.mode,r=Ds({mode:"visible",children:r.children},a,0,null),(i=Fs(i,a,o,null)).flags|=2,r.return=t,i.return=t,r.sibling=i,t.child=r,0!==(1&t.mode)&&wl(t,e.child,null,o),t.child.memoizedState=Do(o),t.memoizedState=Fo,i);if(0===(1&t.mode))return $o(e,t,o,null);if("$!"===a.data){if(r=a.nextSibling&&a.nextSibling.dataset)var u=r.dgst;return r=u,$o(e,t,o,r=so(i=Error(l(419)),r,void 0))}if(u=0!==(o&e.childLanes),bo||u){if(null!==(r=zu)){switch(o&-o){case 4:a=2;break;case 16:a=8;break;case 64:case 128:case 256:case 512:case 1024:case 2048:case 4096:case 8192:case 16384:case 32768:case 65536:case 131072:case 262144:case 524288:case 1048576:case 2097152:case 4194304:case 8388608:case 16777216:case 33554432:case 67108864:a=32;break;case 536870912:a=268435456;break;default:a=0}0!==(a=0!==(a&(r.suspendedLanes|o))?0:a)&&a!==i.retryLane&&(i.retryLane=a,Ol(e,a),ns(r,e,a,-1))}return ms(),$o(e,t,o,r=so(Error(l(421))))}return"$?"===a.data?(t.flags|=128,t.child=e.child,t=Ns.bind(null,e),a._reactRetry=t,null):(e=i.treeContext,rl=sa(a.nextSibling),nl=t,al=!0,ll=null,null!==e&&(qa[Xa++]=Ka,qa[Xa++]=Ga,qa[Xa++]=Ya,Ka=e.id,Ga=e.overflow,Ya=t),t=Ao(t,r.children),t.flags|=4096,t)}(e,t,u,a,r,i,n);if(o){o=a.fallback,u=t.mode,r=(i=e.child).sibling;var s={mode:"hidden",children:a.children};return 0===(1&u)&&t.child!==i?((a=t.child).childLanes=0,a.pendingProps=s,t.deletions=null):(a=Rs(i,s)).subtreeFlags=14680064&i.subtreeFlags,null!==r?o=Rs(r,o):(o=Fs(o,u,n,null)).flags|=2,o.return=t,a.return=t,a.sibling=o,t.child=a,a=o,o=t.child,u=null===(u=e.child.memoizedState)?Do(n):{baseLanes:u.baseLanes|n,cachePool:null,transitions:u.transitions},o.memoizedState=u,o.childLanes=e.childLanes&~n,t.memoizedState=Fo,a}return e=(o=e.child).sibling,a=Rs(o,{mode:"visible",children:a.children}),0===(1&t.mode)&&(a.lanes=n),a.return=t,a.sibling=null,null!==e&&(null===(n=t.deletions)?(t.deletions=[e],t.flags|=16):n.push(e)),t.child=a,t.memoizedState=null,a}function Ao(e,t){return(t=Ds({mode:"visible",children:t},e.mode,0,null)).return=e,e.child=t}function $o(e,t,n,r){return null!==r&&hl(r),wl(t,e.child,null,n),(e=Ao(t,t.pendingProps.children)).flags|=2,t.memoizedState=null,e}function Uo(e,t,n){e.lanes|=t;var r=e.alternate;null!==r&&(r.lanes|=t),Pl(e.return,t,n)}function jo(e,t,n,r,a){var l=e.memoizedState;null===l?e.memoizedState={isBackwards:t,rendering:null,renderingStartTime:0,last:r,tail:n,tailMode:a}:(l.isBackwards=t,l.rendering=null,l.renderingStartTime=0,l.last=r,l.tail=n,l.tailMode=a)}function Vo(e,t,n){var r=t.pendingProps,a=r.revealOrder,l=r.tail;if(wo(e,t,r.children,n),0!==(2&(r=Jl.current)))r=1&r|2,t.flags|=128;else{if(null!==e&&0!==(128&e.flags))e:for(e=t.child;null!==e;){if(13===e.tag)null!==e.memoizedState&&Uo(e,n,t);else if(19===e.tag)Uo(e,n,t);else if(null!==e.child){e.child.return=e,e=e.child;continue}if(e===t)break e;for(;null===e.sibling;){if(null===e.return||e.return===t)break e;e=e.return}e.sibling.return=e.return,e=e.sibling}r&=1}if(Ca(Jl,r),0===(1&t.mode))t.memoizedState=null;else switch(a){case"forwards":for(n=t.child,a=null;null!==n;)null!==(e=n.alternate)&&null===ei(e)&&(a=n),n=n.sibling;null===(n=a)?(a=t.child,t.child=null):(a=n.sibling,n.sibling=null),jo(t,!1,a,n,l);break;case"backwards":for(n=null,a=t.child,t.child=null;null!==a;){if(null!==(e=a.alternate)&&null===ei(e)){t.child=a;break}e=a.sibling,a.sibling=n,n=a,a=e}jo(t,!0,n,null,l);break;case"together":jo(t,!1,null,null,void 0);break;default:t.memoizedState=null}return t.child}function Bo(e,t){0===(1&t.mode)&&null!==e&&(e.alternate=null,t.alternate=null,t.flags|=2)}function Ho(e,t,n){if(null!==e&&(t.dependencies=e.dependencies),Du|=t.lanes,0===(n&t.childLanes))return null;if(null!==e&&t.child!==e.child)throw Error(l(153));if(null!==t.child){for(n=Rs(e=t.child,e.pendingProps),t.child=n,n.return=t;null!==e.sibling;)e=e.sibling,(n=n.sibling=Rs(e,e.pendingProps)).return=t;n.sibling=null}return t.child}function Wo(e,t){if(!al)switch(e.tailMode){case"hidden":t=e.tail;for(var n=null;null!==t;)null!==t.alternate&&(n=t),t=t.sibling;null===n?e.tail=null:n.sibling=null;break;case"collapsed":n=e.tail;for(var r=null;null!==n;)null!==n.alternate&&(r=n),n=n.sibling;null===r?t||null===e.tail?e.tail=null:e.tail.sibling=null:r.sibling=null}}function Qo(e){var t=null!==e.alternate&&e.alternate.child===e.child,n=0,r=0;if(t)for(var a=e.child;null!==a;)n|=a.lanes|a.childLanes,r|=14680064&a.subtreeFlags,r|=14680064&a.flags,a.return=e,a=a.sibling;else for(a=e.child;null!==a;)n|=a.lanes|a.childLanes,r|=a.subtreeFlags,r|=a.flags,a.return=e,a=a.sibling;return e.subtreeFlags|=r,e.childLanes=n,t}function qo(e,t,n){var r=t.pendingProps;switch(tl(t),t.tag){case 2:case 16:case 15:case 0:case 11:case 7:case 8:case 12:case 9:case 14:return Qo(t),null;case 1:case 17:return Ma(t.type)&&Ra(),Qo(t),null;case 3:return r=t.stateNode,Kl(),Ea(za),Ea(Pa),ni(),r.pendingContext&&(r.context=r.pendingContext,r.pendingContext=null),null!==e&&null!==e.child||(fl(t)?t.flags|=4:null===e||e.memoizedState.isDehydrated&&0===(256&t.flags)||(t.flags|=1024,null!==ll&&(is(ll),ll=null))),Mo(e,t),Qo(t),null;case 5:Zl(t);var a=Xl(ql.current);if(n=t.type,null!==e&&null!=t.stateNode)Ro(e,t,n,r,a),e.ref!==t.ref&&(t.flags|=512,t.flags|=2097152);else{if(!r){if(null===t.stateNode)throw Error(l(166));return Qo(t),null}if(e=Xl(Wl.current),fl(t)){r=t.stateNode,n=t.type;var i=t.memoizedProps;switch(r[da]=t,r[pa]=i,e=0!==(1&t.mode),n){case"dialog":$r("cancel",r),$r("close",r);break;case"iframe":case"object":case"embed":$r("load",r);break;case"video":case"audio":for(a=0;a<Fr.length;a++)$r(Fr[a],r);break;case"source":$r("error",r);break;case"img":case"image":case"link":$r("error",r),$r("load",r);break;case"details":$r("toggle",r);break;case"input":K(r,i),$r("invalid",r);break;case"select":r._wrapperState={wasMultiple:!!i.multiple},$r("invalid",r);break;case"textarea":ae(r,i),$r("invalid",r)}for(var u in ye(n,i),a=null,i)if(i.hasOwnProperty(u)){var s=i[u];"children"===u?"string"===typeof s?r.textContent!==s&&(!0!==i.suppressHydrationWarning&&Zr(r.textContent,s,e),a=["children",s]):"number"===typeof s&&r.textContent!==""+s&&(!0!==i.suppressHydrationWarning&&Zr(r.textContent,s,e),a=["children",""+s]):o.hasOwnProperty(u)&&null!=s&&"onScroll"===u&&$r("scroll",r)}switch(n){case"input":Q(r),J(r,i,!0);break;case"textarea":Q(r),ie(r);break;case"select":case"option":break;default:"function"===typeof i.onClick&&(r.onclick=Jr)}r=a,t.updateQueue=r,null!==r&&(t.flags|=4)}else{u=9===a.nodeType?a:a.ownerDocument,"http://www.w3.org/1999/xhtml"===e&&(e=oe(n)),"http://www.w3.org/1999/xhtml"===e?"script"===n?((e=u.createElement("div")).innerHTML="<script><\/script>",e=e.removeChild(e.firstChild)):"string"===typeof r.is?e=u.createElement(n,{is:r.is}):(e=u.createElement(n),"select"===n&&(u=e,r.multiple?u.multiple=!0:r.size&&(u.size=r.size))):e=u.createElementNS(e,n),e[da]=t,e[pa]=r,Lo(e,t,!1,!1),t.stateNode=e;e:{switch(u=be(n,r),n){case"dialog":$r("cancel",e),$r("close",e),a=r;break;case"iframe":case"object":case"embed":$r("load",e),a=r;break;case"video":case"audio":for(a=0;a<Fr.length;a++)$r(Fr[a],e);a=r;break;case"source":$r("error",e),a=r;break;case"img":case"image":case"link":$r("error",e),$r("load",e),a=r;break;case"details":$r("toggle",e),a=r;break;case"input":K(e,r),a=Y(e,r),$r("invalid",e);break;case"option":default:a=r;break;case"select":e._wrapperState={wasMultiple:!!r.multiple},a=I({},r,{value:void 0}),$r("invalid",e);break;case"textarea":ae(e,r),a=re(e,r),$r("invalid",e)}for(i in ye(n,a),s=a)if(s.hasOwnProperty(i)){var c=s[i];"style"===i?ve(e,c):"dangerouslySetInnerHTML"===i?null!=(c=c?c.__html:void 0)&&fe(e,c):"children"===i?"string"===typeof c?("textarea"!==n||""!==c)&&de(e,c):"number"===typeof c&&de(e,""+c):"suppressContentEditableWarning"!==i&&"suppressHydrationWarning"!==i&&"autoFocus"!==i&&(o.hasOwnProperty(i)?null!=c&&"onScroll"===i&&$r("scroll",e):null!=c&&b(e,i,c,u))}switch(n){case"input":Q(e),J(e,r,!1);break;case"textarea":Q(e),ie(e);break;case"option":null!=r.value&&e.setAttribute("value",""+H(r.value));break;case"select":e.multiple=!!r.multiple,null!=(i=r.value)?ne(e,!!r.multiple,i,!1):null!=r.defaultValue&&ne(e,!!r.multiple,r.defaultValue,!0);break;default:"function"===typeof a.onClick&&(e.onclick=Jr)}switch(n){case"button":case"input":case"select":case"textarea":r=!!r.autoFocus;break e;case"img":r=!0;break e;default:r=!1}}r&&(t.flags|=4)}null!==t.ref&&(t.flags|=512,t.flags|=2097152)}return Qo(t),null;case 6:if(e&&null!=t.stateNode)Oo(e,t,e.memoizedProps,r);else{if("string"!==typeof r&&null===t.stateNode)throw Error(l(166));if(n=Xl(ql.current),Xl(Wl.current),fl(t)){if(r=t.stateNode,n=t.memoizedProps,r[da]=t,(i=r.nodeValue!==n)&&null!==(e=nl))switch(e.tag){case 3:Zr(r.nodeValue,n,0!==(1&e.mode));break;case 5:!0!==e.memoizedProps.suppressHydrationWarning&&Zr(r.nodeValue,n,0!==(1&e.mode))}i&&(t.flags|=4)}else(r=(9===n.nodeType?n:n.ownerDocument).createTextNode(r))[da]=t,t.stateNode=r}return Qo(t),null;case 13:if(Ea(Jl),r=t.memoizedState,null===e||null!==e.memoizedState&&null!==e.memoizedState.dehydrated){if(al&&null!==rl&&0!==(1&t.mode)&&0===(128&t.flags))dl(),pl(),t.flags|=98560,i=!1;else if(i=fl(t),null!==r&&null!==r.dehydrated){if(null===e){if(!i)throw Error(l(318));if(!(i=null!==(i=t.memoizedState)?i.dehydrated:null))throw Error(l(317));i[da]=t}else pl(),0===(128&t.flags)&&(t.memoizedState=null),t.flags|=4;Qo(t),i=!1}else null!==ll&&(is(ll),ll=null),i=!0;if(!i)return 65536&t.flags?t:null}return 0!==(128&t.flags)?(t.lanes=n,t):((r=null!==r)!==(null!==e&&null!==e.memoizedState)&&r&&(t.child.flags|=8192,0!==(1&t.mode)&&(null===e||0!==(1&Jl.current)?0===Ou&&(Ou=3):ms())),null!==t.updateQueue&&(t.flags|=4),Qo(t),null);case 4:return Kl(),Mo(e,t),null===e&&Vr(t.stateNode.containerInfo),Qo(t),null;case 10:return Nl(t.type._context),Qo(t),null;case 19:if(Ea(Jl),null===(i=t.memoizedState))return Qo(t),null;if(r=0!==(128&t.flags),null===(u=i.rendering))if(r)Wo(i,!1);else{if(0!==Ou||null!==e&&0!==(128&e.flags))for(e=t.child;null!==e;){if(null!==(u=ei(e))){for(t.flags|=128,Wo(i,!1),null!==(r=u.updateQueue)&&(t.updateQueue=r,t.flags|=4),t.subtreeFlags=0,r=n,n=t.child;null!==n;)e=r,(i=n).flags&=14680066,null===(u=i.alternate)?(i.childLanes=0,i.lanes=e,i.child=null,i.subtreeFlags=0,i.memoizedProps=null,i.memoizedState=null,i.updateQueue=null,i.dependencies=null,i.stateNode=null):(i.childLanes=u.childLanes,i.lanes=u.lanes,i.child=u.child,i.subtreeFlags=0,i.deletions=null,i.memoizedProps=u.memoizedProps,i.memoizedState=u.memoizedState,i.updateQueue=u.updateQueue,i.type=u.type,e=u.dependencies,i.dependencies=null===e?null:{lanes:e.lanes,firstContext:e.firstContext}),n=n.sibling;return Ca(Jl,1&Jl.current|2),t.child}e=e.sibling}null!==i.tail&&Ge()>Vu&&(t.flags|=128,r=!0,Wo(i,!1),t.lanes=4194304)}else{if(!r)if(null!==(e=ei(u))){if(t.flags|=128,r=!0,null!==(n=e.updateQueue)&&(t.updateQueue=n,t.flags|=4),Wo(i,!0),null===i.tail&&"hidden"===i.tailMode&&!u.alternate&&!al)return Qo(t),null}else 2*Ge()-i.renderingStartTime>Vu&&1073741824!==n&&(t.flags|=128,r=!0,Wo(i,!1),t.lanes=4194304);i.isBackwards?(u.sibling=t.child,t.child=u):(null!==(n=i.last)?n.sibling=u:t.child=u,i.last=u)}return null!==i.tail?(t=i.tail,i.rendering=t,i.tail=t.sibling,i.renderingStartTime=Ge(),t.sibling=null,n=Jl.current,Ca(Jl,r?1&n|2:1&n),t):(Qo(t),null);case 22:case 23:return fs(),r=null!==t.memoizedState,null!==e&&null!==e.memoizedState!==r&&(t.flags|=8192),r&&0!==(1&t.mode)?0!==(1073741824&Mu)&&(Qo(t),6&t.subtreeFlags&&(t.flags|=8192)):Qo(t),null;case 24:case 25:return null}throw Error(l(156,t.tag))}function Xo(e,t){switch(tl(t),t.tag){case 1:return Ma(t.type)&&Ra(),65536&(e=t.flags)?(t.flags=-65537&e|128,t):null;case 3:return Kl(),Ea(za),Ea(Pa),ni(),0!==(65536&(e=t.flags))&&0===(128&e)?(t.flags=-65537&e|128,t):null;case 5:return Zl(t),null;case 13:if(Ea(Jl),null!==(e=t.memoizedState)&&null!==e.dehydrated){if(null===t.alternate)throw Error(l(340));pl()}return 65536&(e=t.flags)?(t.flags=-65537&e|128,t):null;case 19:return Ea(Jl),null;case 4:return Kl(),null;case 10:return Nl(t.type._context),null;case 22:case 23:return fs(),null;default:return null}}Lo=function(e,t){for(var n=t.child;null!==n;){if(5===n.tag||6===n.tag)e.appendChild(n.stateNode);else if(4!==n.tag&&null!==n.child){n.child.return=n,n=n.child;continue}if(n===t)break;for(;null===n.sibling;){if(null===n.return||n.return===t)return;n=n.return}n.sibling.return=n.return,n=n.sibling}},Mo=function(){},Ro=function(e,t,n,r){var a=e.memoizedProps;if(a!==r){e=t.stateNode,Xl(Wl.current);var l,i=null;switch(n){case"input":a=Y(e,a),r=Y(e,r),i=[];break;case"select":a=I({},a,{value:void 0}),r=I({},r,{value:void 0}),i=[];break;case"textarea":a=re(e,a),r=re(e,r),i=[];break;default:"function"!==typeof a.onClick&&"function"===typeof r.onClick&&(e.onclick=Jr)}for(c in ye(n,r),n=null,a)if(!r.hasOwnProperty(c)&&a.hasOwnProperty(c)&&null!=a[c])if("style"===c){var u=a[c];for(l in u)u.hasOwnProperty(l)&&(n||(n={}),n[l]="")}else"dangerouslySetInnerHTML"!==c&&"children"!==c&&"suppressContentEditableWarning"!==c&&"suppressHydrationWarning"!==c&&"autoFocus"!==c&&(o.hasOwnProperty(c)?i||(i=[]):(i=i||[]).push(c,null));for(c in r){var s=r[c];if(u=null!=a?a[c]:void 0,r.hasOwnProperty(c)&&s!==u&&(null!=s||null!=u))if("style"===c)if(u){for(l in u)!u.hasOwnProperty(l)||s&&s.hasOwnProperty(l)||(n||(n={}),n[l]="");for(l in s)s.hasOwnProperty(l)&&u[l]!==s[l]&&(n||(n={}),n[l]=s[l])}else n||(i||(i=[]),i.push(c,n)),n=s;else"dangerouslySetInnerHTML"===c?(s=s?s.__html:void 0,u=u?u.__html:void 0,null!=s&&u!==s&&(i=i||[]).push(c,s)):"children"===c?"string"!==typeof s&&"number"!==typeof s||(i=i||[]).push(c,""+s):"suppressContentEditableWarning"!==c&&"suppressHydrationWarning"!==c&&(o.hasOwnProperty(c)?(null!=s&&"onScroll"===c&&$r("scroll",e),i||u===s||(i=[])):(i=i||[]).push(c,s))}n&&(i=i||[]).push("style",n);var c=i;(t.updateQueue=c)&&(t.flags|=4)}},Oo=function(e,t,n,r){n!==r&&(t.flags|=4)};var Yo=!1,Ko=!1,Go="function"===typeof WeakSet?WeakSet:Set,Zo=null;function Jo(e,t){var n=e.ref;if(null!==n)if("function"===typeof n)try{n(null)}catch(r){xs(e,t,r)}else n.current=null}function eu(e,t,n){try{n()}catch(r){xs(e,t,r)}}var tu=!1;function nu(e,t,n){var r=t.updateQueue;if(null!==(r=null!==r?r.lastEffect:null)){var a=r=r.next;do{if((a.tag&e)===e){var l=a.destroy;a.destroy=void 0,void 0!==l&&eu(t,n,l)}a=a.next}while(a!==r)}}function ru(e,t){if(null!==(t=null!==(t=t.updateQueue)?t.lastEffect:null)){var n=t=t.next;do{if((n.tag&e)===e){var r=n.create;n.destroy=r()}n=n.next}while(n!==t)}}function au(e){var t=e.ref;if(null!==t){var n=e.stateNode;e.tag,e=n,"function"===typeof t?t(e):t.current=e}}function lu(e){var t=e.alternate;null!==t&&(e.alternate=null,lu(t)),e.child=null,e.deletions=null,e.sibling=null,5===e.tag&&(null!==(t=e.stateNode)&&(delete t[da],delete t[pa],delete t[ma],delete t[va],delete t[ga])),e.stateNode=null,e.return=null,e.dependencies=null,e.memoizedProps=null,e.memoizedState=null,e.pendingProps=null,e.stateNode=null,e.updateQueue=null}function iu(e){return 5===e.tag||3===e.tag||4===e.tag}function ou(e){e:for(;;){for(;null===e.sibling;){if(null===e.return||iu(e.return))return null;e=e.return}for(e.sibling.return=e.return,e=e.sibling;5!==e.tag&&6!==e.tag&&18!==e.tag;){if(2&e.flags)continue e;if(null===e.child||4===e.tag)continue e;e.child.return=e,e=e.child}if(!(2&e.flags))return e.stateNode}}function uu(e,t,n){var r=e.tag;if(5===r||6===r)e=e.stateNode,t?8===n.nodeType?n.parentNode.insertBefore(e,t):n.insertBefore(e,t):(8===n.nodeType?(t=n.parentNode).insertBefore(e,n):(t=n).appendChild(e),null!==(n=n._reactRootContainer)&&void 0!==n||null!==t.onclick||(t.onclick=Jr));else if(4!==r&&null!==(e=e.child))for(uu(e,t,n),e=e.sibling;null!==e;)uu(e,t,n),e=e.sibling}function su(e,t,n){var r=e.tag;if(5===r||6===r)e=e.stateNode,t?n.insertBefore(e,t):n.appendChild(e);else if(4!==r&&null!==(e=e.child))for(su(e,t,n),e=e.sibling;null!==e;)su(e,t,n),e=e.sibling}var cu=null,fu=!1;function du(e,t,n){for(n=n.child;null!==n;)pu(e,t,n),n=n.sibling}function pu(e,t,n){if(lt&&"function"===typeof lt.onCommitFiberUnmount)try{lt.onCommitFiberUnmount(at,n)}catch(o){}switch(n.tag){case 5:Ko||Jo(n,t);case 6:var r=cu,a=fu;cu=null,du(e,t,n),fu=a,null!==(cu=r)&&(fu?(e=cu,n=n.stateNode,8===e.nodeType?e.parentNode.removeChild(n):e.removeChild(n)):cu.removeChild(n.stateNode));break;case 18:null!==cu&&(fu?(e=cu,n=n.stateNode,8===e.nodeType?ua(e.parentNode,n):1===e.nodeType&&ua(e,n),Vt(e)):ua(cu,n.stateNode));break;case 4:r=cu,a=fu,cu=n.stateNode.containerInfo,fu=!0,du(e,t,n),cu=r,fu=a;break;case 0:case 11:case 14:case 15:if(!Ko&&(null!==(r=n.updateQueue)&&null!==(r=r.lastEffect))){a=r=r.next;do{var l=a,i=l.destroy;l=l.tag,void 0!==i&&(0!==(2&l)||0!==(4&l))&&eu(n,t,i),a=a.next}while(a!==r)}du(e,t,n);break;case 1:if(!Ko&&(Jo(n,t),"function"===typeof(r=n.stateNode).componentWillUnmount))try{r.props=n.memoizedProps,r.state=n.memoizedState,r.componentWillUnmount()}catch(o){xs(n,t,o)}du(e,t,n);break;case 21:du(e,t,n);break;case 22:1&n.mode?(Ko=(r=Ko)||null!==n.memoizedState,du(e,t,n),Ko=r):du(e,t,n);break;default:du(e,t,n)}}function hu(e){var t=e.updateQueue;if(null!==t){e.updateQueue=null;var n=e.stateNode;null===n&&(n=e.stateNode=new Go),t.forEach((function(t){var r=Ps.bind(null,e,t);n.has(t)||(n.add(t),t.then(r,r))}))}}function mu(e,t){var n=t.deletions;if(null!==n)for(var r=0;r<n.length;r++){var a=n[r];try{var i=e,o=t,u=o;e:for(;null!==u;){switch(u.tag){case 5:cu=u.stateNode,fu=!1;break e;case 3:case 4:cu=u.stateNode.containerInfo,fu=!0;break e}u=u.return}if(null===cu)throw Error(l(160));pu(i,o,a),cu=null,fu=!1;var s=a.alternate;null!==s&&(s.return=null),a.return=null}catch(c){xs(a,t,c)}}if(12854&t.subtreeFlags)for(t=t.child;null!==t;)vu(t,e),t=t.sibling}function vu(e,t){var n=e.alternate,r=e.flags;switch(e.tag){case 0:case 11:case 14:case 15:if(mu(t,e),gu(e),4&r){try{nu(3,e,e.return),ru(3,e)}catch(v){xs(e,e.return,v)}try{nu(5,e,e.return)}catch(v){xs(e,e.return,v)}}break;case 1:mu(t,e),gu(e),512&r&&null!==n&&Jo(n,n.return);break;case 5:if(mu(t,e),gu(e),512&r&&null!==n&&Jo(n,n.return),32&e.flags){var a=e.stateNode;try{de(a,"")}catch(v){xs(e,e.return,v)}}if(4&r&&null!=(a=e.stateNode)){var i=e.memoizedProps,o=null!==n?n.memoizedProps:i,u=e.type,s=e.updateQueue;if(e.updateQueue=null,null!==s)try{"input"===u&&"radio"===i.type&&null!=i.name&&G(a,i),be(u,o);var c=be(u,i);for(o=0;o<s.length;o+=2){var f=s[o],d=s[o+1];"style"===f?ve(a,d):"dangerouslySetInnerHTML"===f?fe(a,d):"children"===f?de(a,d):b(a,f,d,c)}switch(u){case"input":Z(a,i);break;case"textarea":le(a,i);break;case"select":var p=a._wrapperState.wasMultiple;a._wrapperState.wasMultiple=!!i.multiple;var h=i.value;null!=h?ne(a,!!i.multiple,h,!1):p!==!!i.multiple&&(null!=i.defaultValue?ne(a,!!i.multiple,i.defaultValue,!0):ne(a,!!i.multiple,i.multiple?[]:"",!1))}a[pa]=i}catch(v){xs(e,e.return,v)}}break;case 6:if(mu(t,e),gu(e),4&r){if(null===e.stateNode)throw Error(l(162));a=e.stateNode,i=e.memoizedProps;try{a.nodeValue=i}catch(v){xs(e,e.return,v)}}break;case 3:if(mu(t,e),gu(e),4&r&&null!==n&&n.memoizedState.isDehydrated)try{Vt(t.containerInfo)}catch(v){xs(e,e.return,v)}break;case 4:default:mu(t,e),gu(e);break;case 13:mu(t,e),gu(e),8192&(a=e.child).flags&&(i=null!==a.memoizedState,a.stateNode.isHidden=i,!i||null!==a.alternate&&null!==a.alternate.memoizedState||(ju=Ge())),4&r&&hu(e);break;case 22:if(f=null!==n&&null!==n.memoizedState,1&e.mode?(Ko=(c=Ko)||f,mu(t,e),Ko=c):mu(t,e),gu(e),8192&r){if(c=null!==e.memoizedState,(e.stateNode.isHidden=c)&&!f&&0!==(1&e.mode))for(Zo=e,f=e.child;null!==f;){for(d=Zo=f;null!==Zo;){switch(h=(p=Zo).child,p.tag){case 0:case 11:case 14:case 15:nu(4,p,p.return);break;case 1:Jo(p,p.return);var m=p.stateNode;if("function"===typeof m.componentWillUnmount){r=p,n=p.return;try{t=r,m.props=t.memoizedProps,m.state=t.memoizedState,m.componentWillUnmount()}catch(v){xs(r,n,v)}}break;case 5:Jo(p,p.return);break;case 22:if(null!==p.memoizedState){ku(d);continue}}null!==h?(h.return=p,Zo=h):ku(d)}f=f.sibling}e:for(f=null,d=e;;){if(5===d.tag){if(null===f){f=d;try{a=d.stateNode,c?"function"===typeof(i=a.style).setProperty?i.setProperty("display","none","important"):i.display="none":(u=d.stateNode,o=void 0!==(s=d.memoizedProps.style)&&null!==s&&s.hasOwnProperty("display")?s.display:null,u.style.display=me("display",o))}catch(v){xs(e,e.return,v)}}}else if(6===d.tag){if(null===f)try{d.stateNode.nodeValue=c?"":d.memoizedProps}catch(v){xs(e,e.return,v)}}else if((22!==d.tag&&23!==d.tag||null===d.memoizedState||d===e)&&null!==d.child){d.child.return=d,d=d.child;continue}if(d===e)break e;for(;null===d.sibling;){if(null===d.return||d.return===e)break e;f===d&&(f=null),d=d.return}f===d&&(f=null),d.sibling.return=d.return,d=d.sibling}}break;case 19:mu(t,e),gu(e),4&r&&hu(e);case 21:}}function gu(e){var t=e.flags;if(2&t){try{e:{for(var n=e.return;null!==n;){if(iu(n)){var r=n;break e}n=n.return}throw Error(l(160))}switch(r.tag){case 5:var a=r.stateNode;32&r.flags&&(de(a,""),r.flags&=-33),su(e,ou(e),a);break;case 3:case 4:var i=r.stateNode.containerInfo;uu(e,ou(e),i);break;default:throw Error(l(161))}}catch(o){xs(e,e.return,o)}e.flags&=-3}4096&t&&(e.flags&=-4097)}function yu(e,t,n){Zo=e,bu(e,t,n)}function bu(e,t,n){for(var r=0!==(1&e.mode);null!==Zo;){var a=Zo,l=a.child;if(22===a.tag&&r){var i=null!==a.memoizedState||Yo;if(!i){var o=a.alternate,u=null!==o&&null!==o.memoizedState||Ko;o=Yo;var s=Ko;if(Yo=i,(Ko=u)&&!s)for(Zo=a;null!==Zo;)u=(i=Zo).child,22===i.tag&&null!==i.memoizedState?_u(a):null!==u?(u.return=i,Zo=u):_u(a);for(;null!==l;)Zo=l,bu(l,t,n),l=l.sibling;Zo=a,Yo=o,Ko=s}wu(e)}else 0!==(8772&a.subtreeFlags)&&null!==l?(l.return=a,Zo=l):wu(e)}}function wu(e){for(;null!==Zo;){var t=Zo;if(0!==(8772&t.flags)){var n=t.alternate;try{if(0!==(8772&t.flags))switch(t.tag){case 0:case 11:case 15:Ko||ru(5,t);break;case 1:var r=t.stateNode;if(4&t.flags&&!Ko)if(null===n)r.componentDidMount();else{var a=t.elementType===t.type?n.memoizedProps:to(t.type,n.memoizedProps);r.componentDidUpdate(a,n.memoizedState,r.__reactInternalSnapshotBeforeUpdate)}var i=t.updateQueue;null!==i&&Bl(t,i,r);break;case 3:var o=t.updateQueue;if(null!==o){if(n=null,null!==t.child)switch(t.child.tag){case 5:case 1:n=t.child.stateNode}Bl(t,o,n)}break;case 5:var u=t.stateNode;if(null===n&&4&t.flags){n=u;var s=t.memoizedProps;switch(t.type){case"button":case"input":case"select":case"textarea":s.autoFocus&&n.focus();break;case"img":s.src&&(n.src=s.src)}}break;case 6:case 4:case 12:case 19:case 17:case 21:case 22:case 23:case 25:break;case 13:if(null===t.memoizedState){var c=t.alternate;if(null!==c){var f=c.memoizedState;if(null!==f){var d=f.dehydrated;null!==d&&Vt(d)}}}break;default:throw Error(l(163))}Ko||512&t.flags&&au(t)}catch(p){xs(t,t.return,p)}}if(t===e){Zo=null;break}if(null!==(n=t.sibling)){n.return=t.return,Zo=n;break}Zo=t.return}}function ku(e){for(;null!==Zo;){var t=Zo;if(t===e){Zo=null;break}var n=t.sibling;if(null!==n){n.return=t.return,Zo=n;break}Zo=t.return}}function _u(e){for(;null!==Zo;){var t=Zo;try{switch(t.tag){case 0:case 11:case 15:var n=t.return;try{ru(4,t)}catch(u){xs(t,n,u)}break;case 1:var r=t.stateNode;if("function"===typeof r.componentDidMount){var a=t.return;try{r.componentDidMount()}catch(u){xs(t,a,u)}}var l=t.return;try{au(t)}catch(u){xs(t,l,u)}break;case 5:var i=t.return;try{au(t)}catch(u){xs(t,i,u)}}}catch(u){xs(t,t.return,u)}if(t===e){Zo=null;break}var o=t.sibling;if(null!==o){o.return=t.return,Zo=o;break}Zo=t.return}}var Su,xu=Math.ceil,Eu=w.ReactCurrentDispatcher,Cu=w.ReactCurrentOwner,Nu=w.ReactCurrentBatchConfig,Pu=0,zu=null,Tu=null,Lu=0,Mu=0,Ru=xa(0),Ou=0,Fu=null,Du=0,Iu=0,Au=0,$u=null,Uu=null,ju=0,Vu=1/0,Bu=null,Hu=!1,Wu=null,Qu=null,qu=!1,Xu=null,Yu=0,Ku=0,Gu=null,Zu=-1,Ju=0;function es(){return 0!==(6&Pu)?Ge():-1!==Zu?Zu:Zu=Ge()}function ts(e){return 0===(1&e.mode)?1:0!==(2&Pu)&&0!==Lu?Lu&-Lu:null!==ml.transition?(0===Ju&&(Ju=mt()),Ju):0!==(e=bt)?e:e=void 0===(e=window.event)?16:Kt(e.type)}function ns(e,t,n,r){if(50<Ku)throw Ku=0,Gu=null,Error(l(185));gt(e,n,r),0!==(2&Pu)&&e===zu||(e===zu&&(0===(2&Pu)&&(Iu|=n),4===Ou&&os(e,Lu)),rs(e,r),1===n&&0===Pu&&0===(1&t.mode)&&(Vu=Ge()+500,$a&&Va()))}function rs(e,t){var n=e.callbackNode;!function(e,t){for(var n=e.suspendedLanes,r=e.pingedLanes,a=e.expirationTimes,l=e.pendingLanes;0<l;){var i=31-it(l),o=1<<i,u=a[i];-1===u?0!==(o&n)&&0===(o&r)||(a[i]=pt(o,t)):u<=t&&(e.expiredLanes|=o),l&=~o}}(e,t);var r=dt(e,e===zu?Lu:0);if(0===r)null!==n&&Xe(n),e.callbackNode=null,e.callbackPriority=0;else if(t=r&-r,e.callbackPriority!==t){if(null!=n&&Xe(n),1===t)0===e.tag?function(e){$a=!0,ja(e)}(us.bind(null,e)):ja(us.bind(null,e)),ia((function(){0===(6&Pu)&&Va()})),n=null;else{switch(wt(r)){case 1:n=Je;break;case 4:n=et;break;case 16:default:n=tt;break;case 536870912:n=rt}n=zs(n,as.bind(null,e))}e.callbackPriority=t,e.callbackNode=n}}function as(e,t){if(Zu=-1,Ju=0,0!==(6&Pu))throw Error(l(327));var n=e.callbackNode;if(_s()&&e.callbackNode!==n)return null;var r=dt(e,e===zu?Lu:0);if(0===r)return null;if(0!==(30&r)||0!==(r&e.expiredLanes)||t)t=vs(e,r);else{t=r;var a=Pu;Pu|=2;var i=hs();for(zu===e&&Lu===t||(Bu=null,Vu=Ge()+500,ds(e,t));;)try{ys();break}catch(u){ps(e,u)}Cl(),Eu.current=i,Pu=a,null!==Tu?t=0:(zu=null,Lu=0,t=Ou)}if(0!==t){if(2===t&&(0!==(a=ht(e))&&(r=a,t=ls(e,a))),1===t)throw n=Fu,ds(e,0),os(e,r),rs(e,Ge()),n;if(6===t)os(e,r);else{if(a=e.current.alternate,0===(30&r)&&!function(e){for(var t=e;;){if(16384&t.flags){var n=t.updateQueue;if(null!==n&&null!==(n=n.stores))for(var r=0;r<n.length;r++){var a=n[r],l=a.getSnapshot;a=a.value;try{if(!or(l(),a))return!1}catch(o){return!1}}}if(n=t.child,16384&t.subtreeFlags&&null!==n)n.return=t,t=n;else{if(t===e)break;for(;null===t.sibling;){if(null===t.return||t.return===e)return!0;t=t.return}t.sibling.return=t.return,t=t.sibling}}return!0}(a)&&(2===(t=vs(e,r))&&(0!==(i=ht(e))&&(r=i,t=ls(e,i))),1===t))throw n=Fu,ds(e,0),os(e,r),rs(e,Ge()),n;switch(e.finishedWork=a,e.finishedLanes=r,t){case 0:case 1:throw Error(l(345));case 2:case 5:ks(e,Uu,Bu);break;case 3:if(os(e,r),(130023424&r)===r&&10<(t=ju+500-Ge())){if(0!==dt(e,0))break;if(((a=e.suspendedLanes)&r)!==r){es(),e.pingedLanes|=e.suspendedLanes&a;break}e.timeoutHandle=ra(ks.bind(null,e,Uu,Bu),t);break}ks(e,Uu,Bu);break;case 4:if(os(e,r),(4194240&r)===r)break;for(t=e.eventTimes,a=-1;0<r;){var o=31-it(r);i=1<<o,(o=t[o])>a&&(a=o),r&=~i}if(r=a,10<(r=(120>(r=Ge()-r)?120:480>r?480:1080>r?1080:1920>r?1920:3e3>r?3e3:4320>r?4320:1960*xu(r/1960))-r)){e.timeoutHandle=ra(ks.bind(null,e,Uu,Bu),r);break}ks(e,Uu,Bu);break;default:throw Error(l(329))}}}return rs(e,Ge()),e.callbackNode===n?as.bind(null,e):null}function ls(e,t){var n=$u;return e.current.memoizedState.isDehydrated&&(ds(e,t).flags|=256),2!==(e=vs(e,t))&&(t=Uu,Uu=n,null!==t&&is(t)),e}function is(e){null===Uu?Uu=e:Uu.push.apply(Uu,e)}function os(e,t){for(t&=~Au,t&=~Iu,e.suspendedLanes|=t,e.pingedLanes&=~t,e=e.expirationTimes;0<t;){var n=31-it(t),r=1<<n;e[n]=-1,t&=~r}}function us(e){if(0!==(6&Pu))throw Error(l(327));_s();var t=dt(e,0);if(0===(1&t))return rs(e,Ge()),null;var n=vs(e,t);if(0!==e.tag&&2===n){var r=ht(e);0!==r&&(t=r,n=ls(e,r))}if(1===n)throw n=Fu,ds(e,0),os(e,t),rs(e,Ge()),n;if(6===n)throw Error(l(345));return e.finishedWork=e.current.alternate,e.finishedLanes=t,ks(e,Uu,Bu),rs(e,Ge()),null}function ss(e,t){var n=Pu;Pu|=1;try{return e(t)}finally{0===(Pu=n)&&(Vu=Ge()+500,$a&&Va())}}function cs(e){null!==Xu&&0===Xu.tag&&0===(6&Pu)&&_s();var t=Pu;Pu|=1;var n=Nu.transition,r=bt;try{if(Nu.transition=null,bt=1,e)return e()}finally{bt=r,Nu.transition=n,0===(6&(Pu=t))&&Va()}}function fs(){Mu=Ru.current,Ea(Ru)}function ds(e,t){e.finishedWork=null,e.finishedLanes=0;var n=e.timeoutHandle;if(-1!==n&&(e.timeoutHandle=-1,aa(n)),null!==Tu)for(n=Tu.return;null!==n;){var r=n;switch(tl(r),r.tag){case 1:null!==(r=r.type.childContextTypes)&&void 0!==r&&Ra();break;case 3:Kl(),Ea(za),Ea(Pa),ni();break;case 5:Zl(r);break;case 4:Kl();break;case 13:case 19:Ea(Jl);break;case 10:Nl(r.type._context);break;case 22:case 23:fs()}n=n.return}if(zu=e,Tu=e=Rs(e.current,null),Lu=Mu=t,Ou=0,Fu=null,Au=Iu=Du=0,Uu=$u=null,null!==Ll){for(t=0;t<Ll.length;t++)if(null!==(r=(n=Ll[t]).interleaved)){n.interleaved=null;var a=r.next,l=n.pending;if(null!==l){var i=l.next;l.next=a,r.next=i}n.pending=r}Ll=null}return e}function ps(e,t){for(;;){var n=Tu;try{if(Cl(),ri.current=Gi,si){for(var r=ii.memoizedState;null!==r;){var a=r.queue;null!==a&&(a.pending=null),r=r.next}si=!1}if(li=0,ui=oi=ii=null,ci=!1,fi=0,Cu.current=null,null===n||null===n.return){Ou=1,Fu=t,Tu=null;break}e:{var i=e,o=n.return,u=n,s=t;if(t=Lu,u.flags|=32768,null!==s&&"object"===typeof s&&"function"===typeof s.then){var c=s,f=u,d=f.tag;if(0===(1&f.mode)&&(0===d||11===d||15===d)){var p=f.alternate;p?(f.updateQueue=p.updateQueue,f.memoizedState=p.memoizedState,f.lanes=p.lanes):(f.updateQueue=null,f.memoizedState=null)}var h=vo(o);if(null!==h){h.flags&=-257,go(h,o,u,0,t),1&h.mode&&mo(i,c,t),s=c;var m=(t=h).updateQueue;if(null===m){var v=new Set;v.add(s),t.updateQueue=v}else m.add(s);break e}if(0===(1&t)){mo(i,c,t),ms();break e}s=Error(l(426))}else if(al&&1&u.mode){var g=vo(o);if(null!==g){0===(65536&g.flags)&&(g.flags|=256),go(g,o,u,0,t),hl(uo(s,u));break e}}i=s=uo(s,u),4!==Ou&&(Ou=2),null===$u?$u=[i]:$u.push(i),i=o;do{switch(i.tag){case 3:i.flags|=65536,t&=-t,i.lanes|=t,jl(i,po(0,s,t));break e;case 1:u=s;var y=i.type,b=i.stateNode;if(0===(128&i.flags)&&("function"===typeof y.getDerivedStateFromError||null!==b&&"function"===typeof b.componentDidCatch&&(null===Qu||!Qu.has(b)))){i.flags|=65536,t&=-t,i.lanes|=t,jl(i,ho(i,u,t));break e}}i=i.return}while(null!==i)}ws(n)}catch(w){t=w,Tu===n&&null!==n&&(Tu=n=n.return);continue}break}}function hs(){var e=Eu.current;return Eu.current=Gi,null===e?Gi:e}function ms(){0!==Ou&&3!==Ou&&2!==Ou||(Ou=4),null===zu||0===(268435455&Du)&&0===(268435455&Iu)||os(zu,Lu)}function vs(e,t){var n=Pu;Pu|=2;var r=hs();for(zu===e&&Lu===t||(Bu=null,ds(e,t));;)try{gs();break}catch(a){ps(e,a)}if(Cl(),Pu=n,Eu.current=r,null!==Tu)throw Error(l(261));return zu=null,Lu=0,Ou}function gs(){for(;null!==Tu;)bs(Tu)}function ys(){for(;null!==Tu&&!Ye();)bs(Tu)}function bs(e){var t=Su(e.alternate,e,Mu);e.memoizedProps=e.pendingProps,null===t?ws(e):Tu=t,Cu.current=null}function ws(e){var t=e;do{var n=t.alternate;if(e=t.return,0===(32768&t.flags)){if(null!==(n=qo(n,t,Mu)))return void(Tu=n)}else{if(null!==(n=Xo(n,t)))return n.flags&=32767,void(Tu=n);if(null===e)return Ou=6,void(Tu=null);e.flags|=32768,e.subtreeFlags=0,e.deletions=null}if(null!==(t=t.sibling))return void(Tu=t);Tu=t=e}while(null!==t);0===Ou&&(Ou=5)}function ks(e,t,n){var r=bt,a=Nu.transition;try{Nu.transition=null,bt=1,function(e,t,n,r){do{_s()}while(null!==Xu);if(0!==(6&Pu))throw Error(l(327));n=e.finishedWork;var a=e.finishedLanes;if(null===n)return null;if(e.finishedWork=null,e.finishedLanes=0,n===e.current)throw Error(l(177));e.callbackNode=null,e.callbackPriority=0;var i=n.lanes|n.childLanes;if(function(e,t){var n=e.pendingLanes&~t;e.pendingLanes=t,e.suspendedLanes=0,e.pingedLanes=0,e.expiredLanes&=t,e.mutableReadLanes&=t,e.entangledLanes&=t,t=e.entanglements;var r=e.eventTimes;for(e=e.expirationTimes;0<n;){var a=31-it(n),l=1<<a;t[a]=0,r[a]=-1,e[a]=-1,n&=~l}}(e,i),e===zu&&(Tu=zu=null,Lu=0),0===(2064&n.subtreeFlags)&&0===(2064&n.flags)||qu||(qu=!0,zs(tt,(function(){return _s(),null}))),i=0!==(15990&n.flags),0!==(15990&n.subtreeFlags)||i){i=Nu.transition,Nu.transition=null;var o=bt;bt=1;var u=Pu;Pu|=4,Cu.current=null,function(e,t){if(ea=Ht,pr(e=dr())){if("selectionStart"in e)var n={start:e.selectionStart,end:e.selectionEnd};else e:{var r=(n=(n=e.ownerDocument)&&n.defaultView||window).getSelection&&n.getSelection();if(r&&0!==r.rangeCount){n=r.anchorNode;var a=r.anchorOffset,i=r.focusNode;r=r.focusOffset;try{n.nodeType,i.nodeType}catch(k){n=null;break e}var o=0,u=-1,s=-1,c=0,f=0,d=e,p=null;t:for(;;){for(var h;d!==n||0!==a&&3!==d.nodeType||(u=o+a),d!==i||0!==r&&3!==d.nodeType||(s=o+r),3===d.nodeType&&(o+=d.nodeValue.length),null!==(h=d.firstChild);)p=d,d=h;for(;;){if(d===e)break t;if(p===n&&++c===a&&(u=o),p===i&&++f===r&&(s=o),null!==(h=d.nextSibling))break;p=(d=p).parentNode}d=h}n=-1===u||-1===s?null:{start:u,end:s}}else n=null}n=n||{start:0,end:0}}else n=null;for(ta={focusedElem:e,selectionRange:n},Ht=!1,Zo=t;null!==Zo;)if(e=(t=Zo).child,0!==(1028&t.subtreeFlags)&&null!==e)e.return=t,Zo=e;else for(;null!==Zo;){t=Zo;try{var m=t.alternate;if(0!==(1024&t.flags))switch(t.tag){case 0:case 11:case 15:case 5:case 6:case 4:case 17:break;case 1:if(null!==m){var v=m.memoizedProps,g=m.memoizedState,y=t.stateNode,b=y.getSnapshotBeforeUpdate(t.elementType===t.type?v:to(t.type,v),g);y.__reactInternalSnapshotBeforeUpdate=b}break;case 3:var w=t.stateNode.containerInfo;1===w.nodeType?w.textContent="":9===w.nodeType&&w.documentElement&&w.removeChild(w.documentElement);break;default:throw Error(l(163))}}catch(k){xs(t,t.return,k)}if(null!==(e=t.sibling)){e.return=t.return,Zo=e;break}Zo=t.return}m=tu,tu=!1}(e,n),vu(n,e),hr(ta),Ht=!!ea,ta=ea=null,e.current=n,yu(n,e,a),Ke(),Pu=u,bt=o,Nu.transition=i}else e.current=n;if(qu&&(qu=!1,Xu=e,Yu=a),i=e.pendingLanes,0===i&&(Qu=null),function(e){if(lt&&"function"===typeof lt.onCommitFiberRoot)try{lt.onCommitFiberRoot(at,e,void 0,128===(128&e.current.flags))}catch(t){}}(n.stateNode),rs(e,Ge()),null!==t)for(r=e.onRecoverableError,n=0;n<t.length;n++)a=t[n],r(a.value,{componentStack:a.stack,digest:a.digest});if(Hu)throw Hu=!1,e=Wu,Wu=null,e;0!==(1&Yu)&&0!==e.tag&&_s(),i=e.pendingLanes,0!==(1&i)?e===Gu?Ku++:(Ku=0,Gu=e):Ku=0,Va()}(e,t,n,r)}finally{Nu.transition=a,bt=r}return null}function _s(){if(null!==Xu){var e=wt(Yu),t=Nu.transition,n=bt;try{if(Nu.transition=null,bt=16>e?16:e,null===Xu)var r=!1;else{if(e=Xu,Xu=null,Yu=0,0!==(6&Pu))throw Error(l(331));var a=Pu;for(Pu|=4,Zo=e.current;null!==Zo;){var i=Zo,o=i.child;if(0!==(16&Zo.flags)){var u=i.deletions;if(null!==u){for(var s=0;s<u.length;s++){var c=u[s];for(Zo=c;null!==Zo;){var f=Zo;switch(f.tag){case 0:case 11:case 15:nu(8,f,i)}var d=f.child;if(null!==d)d.return=f,Zo=d;else for(;null!==Zo;){var p=(f=Zo).sibling,h=f.return;if(lu(f),f===c){Zo=null;break}if(null!==p){p.return=h,Zo=p;break}Zo=h}}}var m=i.alternate;if(null!==m){var v=m.child;if(null!==v){m.child=null;do{var g=v.sibling;v.sibling=null,v=g}while(null!==v)}}Zo=i}}if(0!==(2064&i.subtreeFlags)&&null!==o)o.return=i,Zo=o;else e:for(;null!==Zo;){if(0!==(2048&(i=Zo).flags))switch(i.tag){case 0:case 11:case 15:nu(9,i,i.return)}var y=i.sibling;if(null!==y){y.return=i.return,Zo=y;break e}Zo=i.return}}var b=e.current;for(Zo=b;null!==Zo;){var w=(o=Zo).child;if(0!==(2064&o.subtreeFlags)&&null!==w)w.return=o,Zo=w;else e:for(o=b;null!==Zo;){if(0!==(2048&(u=Zo).flags))try{switch(u.tag){case 0:case 11:case 15:ru(9,u)}}catch(_){xs(u,u.return,_)}if(u===o){Zo=null;break e}var k=u.sibling;if(null!==k){k.return=u.return,Zo=k;break e}Zo=u.return}}if(Pu=a,Va(),lt&&"function"===typeof lt.onPostCommitFiberRoot)try{lt.onPostCommitFiberRoot(at,e)}catch(_){}r=!0}return r}finally{bt=n,Nu.transition=t}}return!1}function Ss(e,t,n){e=$l(e,t=po(0,t=uo(n,t),1),1),t=es(),null!==e&&(gt(e,1,t),rs(e,t))}function xs(e,t,n){if(3===e.tag)Ss(e,e,n);else for(;null!==t;){if(3===t.tag){Ss(t,e,n);break}if(1===t.tag){var r=t.stateNode;if("function"===typeof t.type.getDerivedStateFromError||"function"===typeof r.componentDidCatch&&(null===Qu||!Qu.has(r))){t=$l(t,e=ho(t,e=uo(n,e),1),1),e=es(),null!==t&&(gt(t,1,e),rs(t,e));break}}t=t.return}}function Es(e,t,n){var r=e.pingCache;null!==r&&r.delete(t),t=es(),e.pingedLanes|=e.suspendedLanes&n,zu===e&&(Lu&n)===n&&(4===Ou||3===Ou&&(130023424&Lu)===Lu&&500>Ge()-ju?ds(e,0):Au|=n),rs(e,t)}function Cs(e,t){0===t&&(0===(1&e.mode)?t=1:(t=ct,0===(130023424&(ct<<=1))&&(ct=4194304)));var n=es();null!==(e=Ol(e,t))&&(gt(e,t,n),rs(e,n))}function Ns(e){var t=e.memoizedState,n=0;null!==t&&(n=t.retryLane),Cs(e,n)}function Ps(e,t){var n=0;switch(e.tag){case 13:var r=e.stateNode,a=e.memoizedState;null!==a&&(n=a.retryLane);break;case 19:r=e.stateNode;break;default:throw Error(l(314))}null!==r&&r.delete(t),Cs(e,n)}function zs(e,t){return qe(e,t)}function Ts(e,t,n,r){this.tag=e,this.key=n,this.sibling=this.child=this.return=this.stateNode=this.type=this.elementType=null,this.index=0,this.ref=null,this.pendingProps=t,this.dependencies=this.memoizedState=this.updateQueue=this.memoizedProps=null,this.mode=r,this.subtreeFlags=this.flags=0,this.deletions=null,this.childLanes=this.lanes=0,this.alternate=null}function Ls(e,t,n,r){return new Ts(e,t,n,r)}function Ms(e){return!(!(e=e.prototype)||!e.isReactComponent)}function Rs(e,t){var n=e.alternate;return null===n?((n=Ls(e.tag,t,e.key,e.mode)).elementType=e.elementType,n.type=e.type,n.stateNode=e.stateNode,n.alternate=e,e.alternate=n):(n.pendingProps=t,n.type=e.type,n.flags=0,n.subtreeFlags=0,n.deletions=null),n.flags=14680064&e.flags,n.childLanes=e.childLanes,n.lanes=e.lanes,n.child=e.child,n.memoizedProps=e.memoizedProps,n.memoizedState=e.memoizedState,n.updateQueue=e.updateQueue,t=e.dependencies,n.dependencies=null===t?null:{lanes:t.lanes,firstContext:t.firstContext},n.sibling=e.sibling,n.index=e.index,n.ref=e.ref,n}function Os(e,t,n,r,a,i){var o=2;if(r=e,"function"===typeof e)Ms(e)&&(o=1);else if("string"===typeof e)o=5;else e:switch(e){case S:return Fs(n.children,a,i,t);case x:o=8,a|=8;break;case E:return(e=Ls(12,n,t,2|a)).elementType=E,e.lanes=i,e;case z:return(e=Ls(13,n,t,a)).elementType=z,e.lanes=i,e;case T:return(e=Ls(19,n,t,a)).elementType=T,e.lanes=i,e;case R:return Ds(n,a,i,t);default:if("object"===typeof e&&null!==e)switch(e.$$typeof){case C:o=10;break e;case N:o=9;break e;case P:o=11;break e;case L:o=14;break e;case M:o=16,r=null;break e}throw Error(l(130,null==e?e:typeof e,""))}return(t=Ls(o,n,t,a)).elementType=e,t.type=r,t.lanes=i,t}function Fs(e,t,n,r){return(e=Ls(7,e,r,t)).lanes=n,e}function Ds(e,t,n,r){return(e=Ls(22,e,r,t)).elementType=R,e.lanes=n,e.stateNode={isHidden:!1},e}function Is(e,t,n){return(e=Ls(6,e,null,t)).lanes=n,e}function As(e,t,n){return(t=Ls(4,null!==e.children?e.children:[],e.key,t)).lanes=n,t.stateNode={containerInfo:e.containerInfo,pendingChildren:null,implementation:e.implementation},t}function $s(e,t,n,r,a){this.tag=t,this.containerInfo=e,this.finishedWork=this.pingCache=this.current=this.pendingChildren=null,this.timeoutHandle=-1,this.callbackNode=this.pendingContext=this.context=null,this.callbackPriority=0,this.eventTimes=vt(0),this.expirationTimes=vt(-1),this.entangledLanes=this.finishedLanes=this.mutableReadLanes=this.expiredLanes=this.pingedLanes=this.suspendedLanes=this.pendingLanes=0,this.entanglements=vt(0),this.identifierPrefix=r,this.onRecoverableError=a,this.mutableSourceEagerHydrationData=null}function Us(e,t,n,r,a,l,i,o,u){return e=new $s(e,t,n,o,u),1===t?(t=1,!0===l&&(t|=8)):t=0,l=Ls(3,null,null,t),e.current=l,l.stateNode=e,l.memoizedState={element:r,isDehydrated:n,cache:null,transitions:null,pendingSuspenseBoundaries:null},Dl(l),e}function js(e){if(!e)return Na;e:{if(Ve(e=e._reactInternals)!==e||1!==e.tag)throw Error(l(170));var t=e;do{switch(t.tag){case 3:t=t.stateNode.context;break e;case 1:if(Ma(t.type)){t=t.stateNode.__reactInternalMemoizedMergedChildContext;break e}}t=t.return}while(null!==t);throw Error(l(171))}if(1===e.tag){var n=e.type;if(Ma(n))return Fa(e,n,t)}return t}function Vs(e,t,n,r,a,l,i,o,u){return(e=Us(n,r,!0,e,0,l,0,o,u)).context=js(null),n=e.current,(l=Al(r=es(),a=ts(n))).callback=void 0!==t&&null!==t?t:null,$l(n,l,a),e.current.lanes=a,gt(e,a,r),rs(e,r),e}function Bs(e,t,n,r){var a=t.current,l=es(),i=ts(a);return n=js(n),null===t.context?t.context=n:t.pendingContext=n,(t=Al(l,i)).payload={element:e},null!==(r=void 0===r?null:r)&&(t.callback=r),null!==(e=$l(a,t,i))&&(ns(e,a,i,l),Ul(e,a,i)),i}function Hs(e){return(e=e.current).child?(e.child.tag,e.child.stateNode):null}function Ws(e,t){if(null!==(e=e.memoizedState)&&null!==e.dehydrated){var n=e.retryLane;e.retryLane=0!==n&&n<t?n:t}}function Qs(e,t){Ws(e,t),(e=e.alternate)&&Ws(e,t)}Su=function(e,t,n){if(null!==e)if(e.memoizedProps!==t.pendingProps||za.current)bo=!0;else{if(0===(e.lanes&n)&&0===(128&t.flags))return bo=!1,function(e,t,n){switch(t.tag){case 3:zo(t),pl();break;case 5:Gl(t);break;case 1:Ma(t.type)&&Da(t);break;case 4:Yl(t,t.stateNode.containerInfo);break;case 10:var r=t.type._context,a=t.memoizedProps.value;Ca(_l,r._currentValue),r._currentValue=a;break;case 13:if(null!==(r=t.memoizedState))return null!==r.dehydrated?(Ca(Jl,1&Jl.current),t.flags|=128,null):0!==(n&t.child.childLanes)?Io(e,t,n):(Ca(Jl,1&Jl.current),null!==(e=Ho(e,t,n))?e.sibling:null);Ca(Jl,1&Jl.current);break;case 19:if(r=0!==(n&t.childLanes),0!==(128&e.flags)){if(r)return Vo(e,t,n);t.flags|=128}if(null!==(a=t.memoizedState)&&(a.rendering=null,a.tail=null,a.lastEffect=null),Ca(Jl,Jl.current),r)break;return null;case 22:case 23:return t.lanes=0,xo(e,t,n)}return Ho(e,t,n)}(e,t,n);bo=0!==(131072&e.flags)}else bo=!1,al&&0!==(1048576&t.flags)&&Ja(t,Qa,t.index);switch(t.lanes=0,t.tag){case 2:var r=t.type;Bo(e,t),e=t.pendingProps;var a=La(t,Pa.current);zl(t,n),a=mi(null,t,r,e,a,n);var i=vi();return t.flags|=1,"object"===typeof a&&null!==a&&"function"===typeof a.render&&void 0===a.$$typeof?(t.tag=1,t.memoizedState=null,t.updateQueue=null,Ma(r)?(i=!0,Da(t)):i=!1,t.memoizedState=null!==a.state&&void 0!==a.state?a.state:null,Dl(t),a.updater=ro,t.stateNode=a,a._reactInternals=t,oo(t,r,e,n),t=Po(null,t,r,!0,i,n)):(t.tag=0,al&&i&&el(t),wo(null,t,a,n),t=t.child),t;case 16:r=t.elementType;e:{switch(Bo(e,t),e=t.pendingProps,r=(a=r._init)(r._payload),t.type=r,a=t.tag=function(e){if("function"===typeof e)return Ms(e)?1:0;if(void 0!==e&&null!==e){if((e=e.$$typeof)===P)return 11;if(e===L)return 14}return 2}(r),e=to(r,e),a){case 0:t=Co(null,t,r,e,n);break e;case 1:t=No(null,t,r,e,n);break e;case 11:t=ko(null,t,r,e,n);break e;case 14:t=_o(null,t,r,to(r.type,e),n);break e}throw Error(l(306,r,""))}return t;case 0:return r=t.type,a=t.pendingProps,Co(e,t,r,a=t.elementType===r?a:to(r,a),n);case 1:return r=t.type,a=t.pendingProps,No(e,t,r,a=t.elementType===r?a:to(r,a),n);case 3:e:{if(zo(t),null===e)throw Error(l(387));r=t.pendingProps,a=(i=t.memoizedState).element,Il(e,t),Vl(t,r,null,n);var o=t.memoizedState;if(r=o.element,i.isDehydrated){if(i={element:r,isDehydrated:!1,cache:o.cache,pendingSuspenseBoundaries:o.pendingSuspenseBoundaries,transitions:o.transitions},t.updateQueue.baseState=i,t.memoizedState=i,256&t.flags){t=To(e,t,r,n,a=uo(Error(l(423)),t));break e}if(r!==a){t=To(e,t,r,n,a=uo(Error(l(424)),t));break e}for(rl=sa(t.stateNode.containerInfo.firstChild),nl=t,al=!0,ll=null,n=kl(t,null,r,n),t.child=n;n;)n.flags=-3&n.flags|4096,n=n.sibling}else{if(pl(),r===a){t=Ho(e,t,n);break e}wo(e,t,r,n)}t=t.child}return t;case 5:return Gl(t),null===e&&sl(t),r=t.type,a=t.pendingProps,i=null!==e?e.memoizedProps:null,o=a.children,na(r,a)?o=null:null!==i&&na(r,i)&&(t.flags|=32),Eo(e,t),wo(e,t,o,n),t.child;case 6:return null===e&&sl(t),null;case 13:return Io(e,t,n);case 4:return Yl(t,t.stateNode.containerInfo),r=t.pendingProps,null===e?t.child=wl(t,null,r,n):wo(e,t,r,n),t.child;case 11:return r=t.type,a=t.pendingProps,ko(e,t,r,a=t.elementType===r?a:to(r,a),n);case 7:return wo(e,t,t.pendingProps,n),t.child;case 8:case 12:return wo(e,t,t.pendingProps.children,n),t.child;case 10:e:{if(r=t.type._context,a=t.pendingProps,i=t.memoizedProps,o=a.value,Ca(_l,r._currentValue),r._currentValue=o,null!==i)if(or(i.value,o)){if(i.children===a.children&&!za.current){t=Ho(e,t,n);break e}}else for(null!==(i=t.child)&&(i.return=t);null!==i;){var u=i.dependencies;if(null!==u){o=i.child;for(var s=u.firstContext;null!==s;){if(s.context===r){if(1===i.tag){(s=Al(-1,n&-n)).tag=2;var c=i.updateQueue;if(null!==c){var f=(c=c.shared).pending;null===f?s.next=s:(s.next=f.next,f.next=s),c.pending=s}}i.lanes|=n,null!==(s=i.alternate)&&(s.lanes|=n),Pl(i.return,n,t),u.lanes|=n;break}s=s.next}}else if(10===i.tag)o=i.type===t.type?null:i.child;else if(18===i.tag){if(null===(o=i.return))throw Error(l(341));o.lanes|=n,null!==(u=o.alternate)&&(u.lanes|=n),Pl(o,n,t),o=i.sibling}else o=i.child;if(null!==o)o.return=i;else for(o=i;null!==o;){if(o===t){o=null;break}if(null!==(i=o.sibling)){i.return=o.return,o=i;break}o=o.return}i=o}wo(e,t,a.children,n),t=t.child}return t;case 9:return a=t.type,r=t.pendingProps.children,zl(t,n),r=r(a=Tl(a)),t.flags|=1,wo(e,t,r,n),t.child;case 14:return a=to(r=t.type,t.pendingProps),_o(e,t,r,a=to(r.type,a),n);case 15:return So(e,t,t.type,t.pendingProps,n);case 17:return r=t.type,a=t.pendingProps,a=t.elementType===r?a:to(r,a),Bo(e,t),t.tag=1,Ma(r)?(e=!0,Da(t)):e=!1,zl(t,n),lo(t,r,a),oo(t,r,a,n),Po(null,t,r,!0,e,n);case 19:return Vo(e,t,n);case 22:return xo(e,t,n)}throw Error(l(156,t.tag))};var qs="function"===typeof reportError?reportError:function(e){console.error(e)};function Xs(e){this._internalRoot=e}function Ys(e){this._internalRoot=e}function Ks(e){return!(!e||1!==e.nodeType&&9!==e.nodeType&&11!==e.nodeType)}function Gs(e){return!(!e||1!==e.nodeType&&9!==e.nodeType&&11!==e.nodeType&&(8!==e.nodeType||" react-mount-point-unstable "!==e.nodeValue))}function Zs(){}function Js(e,t,n,r,a){var l=n._reactRootContainer;if(l){var i=l;if("function"===typeof a){var o=a;a=function(){var e=Hs(i);o.call(e)}}Bs(t,i,e,a)}else i=function(e,t,n,r,a){if(a){if("function"===typeof r){var l=r;r=function(){var e=Hs(i);l.call(e)}}var i=Vs(t,r,e,0,null,!1,0,"",Zs);return e._reactRootContainer=i,e[ha]=i.current,Vr(8===e.nodeType?e.parentNode:e),cs(),i}for(;a=e.lastChild;)e.removeChild(a);if("function"===typeof r){var o=r;r=function(){var e=Hs(u);o.call(e)}}var u=Us(e,0,!1,null,0,!1,0,"",Zs);return e._reactRootContainer=u,e[ha]=u.current,Vr(8===e.nodeType?e.parentNode:e),cs((function(){Bs(t,u,n,r)})),u}(n,t,e,a,r);return Hs(i)}Ys.prototype.render=Xs.prototype.render=function(e){var t=this._internalRoot;if(null===t)throw Error(l(409));Bs(e,t,null,null)},Ys.prototype.unmount=Xs.prototype.unmount=function(){var e=this._internalRoot;if(null!==e){this._internalRoot=null;var t=e.containerInfo;cs((function(){Bs(null,e,null,null)})),t[ha]=null}},Ys.prototype.unstable_scheduleHydration=function(e){if(e){var t=xt();e={blockedOn:null,target:e,priority:t};for(var n=0;n<Rt.length&&0!==t&&t<Rt[n].priority;n++);Rt.splice(n,0,e),0===n&&It(e)}},kt=function(e){switch(e.tag){case 3:var t=e.stateNode;if(t.current.memoizedState.isDehydrated){var n=ft(t.pendingLanes);0!==n&&(yt(t,1|n),rs(t,Ge()),0===(6&Pu)&&(Vu=Ge()+500,Va()))}break;case 13:cs((function(){var t=Ol(e,1);if(null!==t){var n=es();ns(t,e,1,n)}})),Qs(e,1)}},_t=function(e){if(13===e.tag){var t=Ol(e,134217728);if(null!==t)ns(t,e,134217728,es());Qs(e,134217728)}},St=function(e){if(13===e.tag){var t=ts(e),n=Ol(e,t);if(null!==n)ns(n,e,t,es());Qs(e,t)}},xt=function(){return bt},Et=function(e,t){var n=bt;try{return bt=e,t()}finally{bt=n}},_e=function(e,t,n){switch(t){case"input":if(Z(e,n),t=n.name,"radio"===n.type&&null!=t){for(n=e;n.parentNode;)n=n.parentNode;for(n=n.querySelectorAll("input[name="+JSON.stringify(""+t)+'][type="radio"]'),t=0;t<n.length;t++){var r=n[t];if(r!==e&&r.form===e.form){var a=ka(r);if(!a)throw Error(l(90));q(r),Z(r,a)}}}break;case"textarea":le(e,n);break;case"select":null!=(t=n.value)&&ne(e,!!n.multiple,t,!1)}},Pe=ss,ze=cs;var ec={usingClientEntryPoint:!1,Events:[ba,wa,ka,Ce,Ne,ss]},tc={findFiberByHostInstance:ya,bundleType:0,version:"18.3.1",rendererPackageName:"react-dom"},nc={bundleType:tc.bundleType,version:tc.version,rendererPackageName:tc.rendererPackageName,rendererConfig:tc.rendererConfig,overrideHookState:null,overrideHookStateDeletePath:null,overrideHookStateRenamePath:null,overrideProps:null,overridePropsDeletePath:null,overridePropsRenamePath:null,setErrorHandler:null,setSuspenseHandler:null,scheduleUpdate:null,currentDispatcherRef:w.ReactCurrentDispatcher,findHostInstanceByFiber:function(e){return null===(e=We(e))?null:e.stateNode},findFiberByHostInstance:tc.findFiberByHostInstance||function(){return null},findHostInstancesForRefresh:null,scheduleRefresh:null,scheduleRoot:null,setRefreshHandler:null,getCurrentFiber:null,reconcilerVersion:"18.3.1-next-f1338f8080-20240426"};if("undefined"!==typeof __REACT_DEVTOOLS_GLOBAL_HOOK__){var rc=__REACT_DEVTOOLS_GLOBAL_HOOK__;if(!rc.isDisabled&&rc.supportsFiber)try{at=rc.inject(nc),lt=rc}catch(ce){}}t.__SECRET_INTERNALS_DO_NOT_USE_OR_YOU_WILL_BE_FIRED=ec,t.createPortal=function(e,t){var n=2<arguments.length&&void 0!==arguments[2]?arguments[2]:null;if(!Ks(t))throw Error(l(200));return function(e,t,n){var r=3<arguments.length&&void 0!==arguments[3]?arguments[3]:null;return{$$typeof:_,key:null==r?null:""+r,children:e,containerInfo:t,implementation:n}}(e,t,null,n)},t.createRoot=function(e,t){if(!Ks(e))throw Error(l(299));var n=!1,r="",a=qs;return null!==t&&void 0!==t&&(!0===t.unstable_strictMode&&(n=!0),void 0!==t.identifierPrefix&&(r=t.identifierPrefix),void 0!==t.onRecoverableError&&(a=t.onRecoverableError)),t=Us(e,1,!1,null,0,n,0,r,a),e[ha]=t.current,Vr(8===e.nodeType?e.parentNode:e),new Xs(t)},t.findDOMNode=function(e){if(null==e)return null;if(1===e.nodeType)return e;var t=e._reactInternals;if(void 0===t){if("function"===typeof e.render)throw Error(l(188));throw e=Object.keys(e).join(","),Error(l(268,e))}return e=null===(e=We(t))?null:e.stateNode},t.flushSync=function(e){return cs(e)},t.hydrate=function(e,t,n){if(!Gs(t))throw Error(l(200));return Js(null,e,t,!0,n)},t.hydrateRoot=function(e,t,n){if(!Ks(e))throw Error(l(405));var r=null!=n&&n.hydratedSources||null,a=!1,i="",o=qs;if(null!==n&&void 0!==n&&(!0===n.unstable_strictMode&&(a=!0),void 0!==n.identifierPrefix&&(i=n.identifierPrefix),void 0!==n.onRecoverableError&&(o=n.onRecoverableError)),t=Vs(t,null,e,1,null!=n?n:null,a,0,i,o),e[ha]=t.current,Vr(e),r)for(e=0;e<r.length;e++)a=(a=(n=r[e])._getVersion)(n._source),null==t.mutableSourceEagerHydrationData?t.mutableSourceEagerHydrationData=[n,a]:t.mutableSourceEagerHydrationData.push(n,a);return new Ys(t)},t.render=function(e,t,n){if(!Gs(t))throw Error(l(200));return Js(null,e,t,!1,n)},t.unmountComponentAtNode=function(e){if(!Gs(e))throw Error(l(40));return!!e._reactRootContainer&&(cs((function(){Js(null,null,e,!1,(function(){e._reactRootContainer=null,e[ha]=null}))})),!0)},t.unstable_batchedUpdates=ss,t.unstable_renderSubtreeIntoContainer=function(e,t,n,r){if(!Gs(n))throw Error(l(200));if(null==e||void 0===e._reactInternals)throw Error(l(38));return Js(e,t,n,!1,r)},t.version="18.3.1-next-f1338f8080-20240426"},950:(e,t,n)=>{!function e(){if("undefined"!==typeof __REACT_DEVTOOLS_GLOBAL_HOOK__&&"function"===typeof __REACT_DEVTOOLS_GLOBAL_HOOK__.checkDCE)try{__REACT_DEVTOOLS_GLOBAL_HOOK__.checkDCE(e)}catch(t){console.error(t)}}(),e.exports=n(730)},153:(e,t,n)=>{var r=n(43),a=Symbol.for("react.element"),l=Symbol.for("react.fragment"),i=Object.prototype.hasOwnProperty,o=r.__SECRET_INTERNALS_DO_NOT_USE_OR_YOU_WILL_BE_FIRED.ReactCurrentOwner,u={key:!0,ref:!0,__self:!0,__source:!0};function s(e,t,n){var r,l={},s=null,c=null;for(r in void 0!==n&&(s=""+n),void 0!==t.key&&(s=""+t.key),void 0!==t.ref&&(c=t.ref),t)i.call(t,r)&&!u.hasOwnProperty(r)&&(l[r]=t[r]);if(e&&e.defaultProps)for(r in t=e.defaultProps)void 0===l[r]&&(l[r]=t[r]);return{$$typeof:a,type:e,key:s,ref:c,props:l,_owner:o.current}}t.jsx=s},202:(e,t)=>{var n=Symbol.for("react.element"),r=Symbol.for("react.portal"),a=Symbol.for("react.fragment"),l=Symbol.for("react.strict_mode"),i=Symbol.for("react.profiler"),o=Symbol.for("react.provider"),u=Symbol.for("react.context"),s=Symbol.for("react.forward_ref"),c=Symbol.for("react.suspense"),f=Symbol.for("react.memo"),d=Symbol.for("react.lazy"),p=Symbol.iterator;var h={isMounted:function(){return!1},enqueueForceUpdate:function(){},enqueueReplaceState:function(){},enqueueSetState:function(){}},m=Object.assign,v={};function g(e,t,n){this.props=e,this.context=t,this.refs=v,this.updater=n||h}function y(){}function b(e,t,n){this.props=e,this.context=t,this.refs=v,this.updater=n||h}g.prototype.isReactComponent={},g.prototype.setState=function(e,t){if("object"!==typeof e&&"function"!==typeof e&&null!=e)throw Error("setState(...): takes an object of state variables to update or a function which returns an object of state variables.");this.updater.enqueueSetState(this,e,t,"setState")},g.prototype.forceUpdate=function(e){this.updater.enqueueForceUpdate(this,e,"forceUpdate")},y.prototype=g.prototype;var w=b.prototype=new y;w.constructor=b,m(w,g.prototype),w.isPureReactComponent=!0;var k=Array.isArray,_=Object.prototype.hasOwnProperty,S={current:null},x={key:!0,ref:!0,__self:!0,__source:!0};function E(e,t,r){var a,l={},i=null,o=null;if(null!=t)for(a in void 0!==t.ref&&(o=t.ref),void 0!==t.key&&(i=""+t.key),t)_.call(t,a)&&!x.hasOwnProperty(a)&&(l[a]=t[a]);var u=arguments.length-2;if(1===u)l.children=r;else if(1<u){for(var s=Array(u),c=0;c<u;c++)s[c]=arguments[c+2];l.children=s}if(e&&e.defaultProps)for(a in u=e.defaultProps)void 0===l[a]&&(l[a]=u[a]);return{$$typeof:n,type:e,key:i,ref:o,props:l,_owner:S.current}}function C(e){return"object"===typeof e&&null!==e&&e.$$typeof===n}var N=/\/+/g;function P(e,t){return"object"===typeof e&&null!==e&&null!=e.key?function(e){var t={"=":"=0",":":"=2"};return"$"+e.replace(/[=:]/g,(function(e){return t[e]}))}(""+e.key):t.toString(36)}function z(e,t,a,l,i){var o=typeof e;"undefined"!==o&&"boolean"!==o||(e=null);var u=!1;if(null===e)u=!0;else switch(o){case"string":case"number":u=!0;break;case"object":switch(e.$$typeof){case n:case r:u=!0}}if(u)return i=i(u=e),e=""===l?"."+P(u,0):l,k(i)?(a="",null!=e&&(a=e.replace(N,"$&/")+"/"),z(i,t,a,"",(function(e){return e}))):null!=i&&(C(i)&&(i=function(e,t){return{$$typeof:n,type:e.type,key:t,ref:e.ref,props:e.props,_owner:e._owner}}(i,a+(!i.key||u&&u.key===i.key?"":(""+i.key).replace(N,"$&/")+"/")+e)),t.push(i)),1;if(u=0,l=""===l?".":l+":",k(e))for(var s=0;s<e.length;s++){var c=l+P(o=e[s],s);u+=z(o,t,a,c,i)}else if(c=function(e){return null===e||"object"!==typeof e?null:"function"===typeof(e=p&&e[p]||e["@@iterator"])?e:null}(e),"function"===typeof c)for(e=c.call(e),s=0;!(o=e.next()).done;)u+=z(o=o.value,t,a,c=l+P(o,s++),i);else if("object"===o)throw t=String(e),Error("Objects are not valid as a React child (found: "+("[object Object]"===t?"object with keys {"+Object.keys(e).join(", ")+"}":t)+"). If you meant to render a collection of children, use an array instead.");return u}function T(e,t,n){if(null==e)return e;var r=[],a=0;return z(e,r,"","",(function(e){return t.call(n,e,a++)})),r}function L(e){if(-1===e._status){var t=e._result;(t=t()).then((function(t){0!==e._status&&-1!==e._status||(e._status=1,e._result=t)}),(function(t){0!==e._status&&-1!==e._status||(e._status=2,e._result=t)})),-1===e._status&&(e._status=0,e._result=t)}if(1===e._status)return e._result.default;throw e._result}var M={current:null},R={transition:null},O={ReactCurrentDispatcher:M,ReactCurrentBatchConfig:R,ReactCurrentOwner:S};function F(){throw Error("act(...) is not supported in production builds of React.")}t.Children={map:T,forEach:function(e,t,n){T(e,(function(){t.apply(this,arguments)}),n)},count:function(e){var t=0;return T(e,(function(){t++})),t},toArray:function(e){return T(e,(function(e){return e}))||[]},only:function(e){if(!C(e))throw Error("React.Children.only expected to receive a single React element child.");return e}},t.Component=g,t.Fragment=a,t.Profiler=i,t.PureComponent=b,t.StrictMode=l,t.Suspense=c,t.__SECRET_INTERNALS_DO_NOT_USE_OR_YOU_WILL_BE_FIRED=O,t.act=F,t.cloneElement=function(e,t,r){if(null===e||void 0===e)throw Error("React.cloneElement(...): The argument must be a React element, but you passed "+e+".");var a=m({},e.props),l=e.key,i=e.ref,o=e._owner;if(null!=t){if(void 0!==t.ref&&(i=t.ref,o=S.current),void 0!==t.key&&(l=""+t.key),e.type&&e.type.defaultProps)var u=e.type.defaultProps;for(s in t)_.call(t,s)&&!x.hasOwnProperty(s)&&(a[s]=void 0===t[s]&&void 0!==u?u[s]:t[s])}var s=arguments.length-2;if(1===s)a.children=r;else if(1<s){u=Array(s);for(var c=0;c<s;c++)u[c]=arguments[c+2];a.children=u}return{$$typeof:n,type:e.type,key:l,ref:i,props:a,_owner:o}},t.createContext=function(e){return(e={$$typeof:u,_currentValue:e,_currentValue2:e,_threadCount:0,Provider:null,Consumer:null,_defaultValue:null,_globalName:null}).Provider={$$typeof:o,_context:e},e.Consumer=e},t.createElement=E,t.createFactory=function(e){var t=E.bind(null,e);return t.type=e,t},t.createRef=function(){return{current:null}},t.forwardRef=function(e){return{$$typeof:s,render:e}},t.isValidElement=C,t.lazy=function(e){return{$$typeof:d,_payload:{_status:-1,_result:e},_init:L}},t.memo=function(e,t){return{$$typeof:f,type:e,compare:void 0===t?null:t}},t.startTransition=function(e){var t=R.transition;R.transition={};try{e()}finally{R.transition=t}},t.unstable_act=F,t.useCallback=function(e,t){return M.current.useCallback(e,t)},t.useContext=function(e){return M.current.useContext(e)},t.useDebugValue=function(){},t.useDeferredValue=function(e){return M.current.useDeferredValue(e)},t.useEffect=function(e,t){return M.current.useEffect(e,t)},t.useId=function(){return M.current.useId()},t.useImperativeHandle=function(e,t,n){return M.current.useImperativeHandle(e,t,n)},t.useInsertionEffect=function(e,t){return M.current.useInsertionEffect(e,t)},t.useLayoutEffect=function(e,t){return M.current.useLayoutEffect(e,t)},t.useMemo=function(e,t){return M.current.useMemo(e,t)},t.useReducer=function(e,t,n){return M.current.useReducer(e,t,n)},t.useRef=function(e){return M.current.useRef(e)},t.useState=function(e){return M.current.useState(e)},t.useSyncExternalStore=function(e,t,n){return M.current.useSyncExternalStore(e,t,n)},t.useTransition=function(){return M.current.useTransition()},t.version="18.3.1"},43:(e,t,n)=>{e.exports=n(202)},579:(e,t,n)=>{e.exports=n(153)},234:(e,t)=>{function n(e,t){var n=e.length;e.push(t);e:for(;0<n;){var r=n-1>>>1,a=e[r];if(!(0<l(a,t)))break e;e[r]=t,e[n]=a,n=r}}function r(e){return 0===e.length?null:e[0]}function a(e){if(0===e.length)return null;var t=e[0],n=e.pop();if(n!==t){e[0]=n;e:for(var r=0,a=e.length,i=a>>>1;r<i;){var o=2*(r+1)-1,u=e[o],s=o+1,c=e[s];if(0>l(u,n))s<a&&0>l(c,u)?(e[r]=c,e[s]=n,r=s):(e[r]=u,e[o]=n,r=o);else{if(!(s<a&&0>l(c,n)))break e;e[r]=c,e[s]=n,r=s}}}return t}function l(e,t){var n=e.sortIndex-t.sortIndex;return 0!==n?n:e.id-t.id}if("object"===typeof performance&&"function"===typeof performance.now){var i=performance;t.unstable_now=function(){return i.now()}}else{var o=Date,u=o.now();t.unstable_now=function(){return o.now()-u}}var s=[],c=[],f=1,d=null,p=3,h=!1,m=!1,v=!1,g="function"===typeof setTimeout?setTimeout:null,y="function"===typeof clearTimeout?clearTimeout:null,b="undefined"!==typeof setImmediate?setImmediate:null;function w(e){for(var t=r(c);null!==t;){if(null===t.callback)a(c);else{if(!(t.startTime<=e))break;a(c),t.sortIndex=t.expirationTime,n(s,t)}t=r(c)}}function k(e){if(v=!1,w(e),!m)if(null!==r(s))m=!0,R(_);else{var t=r(c);null!==t&&O(k,t.startTime-e)}}function _(e,n){m=!1,v&&(v=!1,y(C),C=-1),h=!0;var l=p;try{for(w(n),d=r(s);null!==d&&(!(d.expirationTime>n)||e&&!z());){var i=d.callback;if("function"===typeof i){d.callback=null,p=d.priorityLevel;var o=i(d.expirationTime<=n);n=t.unstable_now(),"function"===typeof o?d.callback=o:d===r(s)&&a(s),w(n)}else a(s);d=r(s)}if(null!==d)var u=!0;else{var f=r(c);null!==f&&O(k,f.startTime-n),u=!1}return u}finally{d=null,p=l,h=!1}}"undefined"!==typeof navigator&&void 0!==navigator.scheduling&&void 0!==navigator.scheduling.isInputPending&&navigator.scheduling.isInputPending.bind(navigator.scheduling);var S,x=!1,E=null,C=-1,N=5,P=-1;function z(){return!(t.unstable_now()-P<N)}function T(){if(null!==E){var e=t.unstable_now();P=e;var n=!0;try{n=E(!0,e)}finally{n?S():(x=!1,E=null)}}else x=!1}if("function"===typeof b)S=function(){b(T)};else if("undefined"!==typeof MessageChannel){var L=new MessageChannel,M=L.port2;L.port1.onmessage=T,S=function(){M.postMessage(null)}}else S=function(){g(T,0)};function R(e){E=e,x||(x=!0,S())}function O(e,n){C=g((function(){e(t.unstable_now())}),n)}t.unstable_IdlePriority=5,t.unstable_ImmediatePriority=1,t.unstable_LowPriority=4,t.unstable_NormalPriority=3,t.unstable_Profiling=null,t.unstable_UserBlockingPriority=2,t.unstable_cancelCallback=function(e){e.callback=null},t.unstable_continueExecution=function(){m||h||(m=!0,R(_))},t.unstable_forceFrameRate=function(e){0>e||125<e?console.error("forceFrameRate takes a positive int between 0 and 125, forcing frame rates higher than 125 fps is not supported"):N=0<e?Math.floor(1e3/e):5},t.unstable_getCurrentPriorityLevel=function(){return p},t.unstable_getFirstCallbackNode=function(){return r(s)},t.unstable_next=function(e){switch(p){case 1:case 2:case 3:var t=3;break;default:t=p}var n=p;p=t;try{return e()}finally{p=n}},t.unstable_pauseExecution=function(){},t.unstable_requestPaint=function(){},t.unstable_runWithPriority=function(e,t){switch(e){case 1:case 2:case 3:case 4:case 5:break;default:e=3}var n=p;p=e;try{return t()}finally{p=n}},t.unstable_scheduleCallback=function(e,a,l){var i=t.unstable_now();switch("object"===typeof l&&null!==l?l="number"===typeof(l=l.delay)&&0<l?i+l:i:l=i,e){case 1:var o=-1;break;case 2:o=250;break;case 5:o=1073741823;break;case 4:o=1e4;break;default:o=5e3}return e={id:f++,callback:a,priorityLevel:e,startTime:l,expirationTime:o=l+o,sortIndex:-1},l>i?(e.sortIndex=l,n(c,e),null===r(s)&&e===r(c)&&(v?(y(C),C=-1):v=!0,O(k,l-i))):(e.sortIndex=o,n(s,e),m||h||(m=!0,R(_))),e},t.unstable_shouldYield=z,t.unstable_wrapCallback=function(e){var t=p;return function(){var n=p;p=t;try{return e.apply(this,arguments)}finally{p=n}}}},853:(e,t,n)=>{e.exports=n(234)}},t={};function n(r){var a=t[r];if(void 0!==a)return a.exports;var l=t[r]={exports:{}};return e[r](l,l.exports,n),l.exports}n.m=e,n.d=(e,t)=>{for(var r in t)n.o(t,r)&&!n.o(e,r)&&Object.defineProperty(e,r,{enumerable:!0,get:t[r]})},n.f={},n.e=e=>Promise.all(Object.keys(n.f).reduce(((t,r)=>(n.f[r](e,t),t)),[])),n.u=e=>"static/js/"+e+".8ab44547.chunk.js",n.miniCssF=e=>{},n.o=(e,t)=>Object.prototype.hasOwnProperty.call(e,t),(()=>{var e={},t="frontend:";n.l=(r,a,l,i)=>{if(e[r])e[r].push(a);else{var o,u;if(void 0!==l)for(var s=document.getElementsByTagName("script"),c=0;c<s.length;c++){var f=s[c];if(f.getAttribute("src")==r||f.getAttribute("data-webpack")==t+l){o=f;break}}o||(u=!0,(o=document.createElement("script")).charset="utf-8",o.timeout=120,n.nc&&o.setAttribute("nonce",n.nc),o.setAttribute("data-webpack",t+l),o.src=r),e[r]=[a];var d=(t,n)=>{o.onerror=o.onload=null,clearTimeout(p);var a=e[r];if(delete e[r],o.parentNode&&o.parentNode.removeChild(o),a&&a.forEach((e=>e(n))),t)return t(n)},p=setTimeout(d.bind(null,void 0,{type:"timeout",target:o}),12e4);o.onerror=d.bind(null,o.onerror),o.onload=d.bind(null,o.onload),u&&document.head.appendChild(o)}}})(),n.r=e=>{"undefined"!==typeof Symbol&&Symbol.toStringTag&&Object.defineProperty(e,Symbol.toStringTag,{value:"Module"}),Object.defineProperty(e,"__esModule",{value:!0})},n.p="/",(()=>{var e={792:0};n.f.j=(t,r)=>{var a=n.o(e,t)?e[t]:void 0;if(0!==a)if(a)r.push(a[2]);else{var l=new Promise(((n,r)=>a=e[t]=[n,r]));r.push(a[2]=l);var i=n.p+n.u(t),o=new Error;n.l(i,(r=>{if(n.o(e,t)&&(0!==(a=e[t])&&(e[t]=void 0),a)){var l=r&&("load"===r.type?"missing":r.type),i=r&&r.target&&r.target.src;o.message="Loading chunk "+t+" failed.\n("+l+": "+i+")",o.name="ChunkLoadError",o.type=l,o.request=i,a[1](o)}}),"chunk-"+t,t)}};var t=(t,r)=>{var a,l,i=r[0],o=r[1],u=r[2],s=0;if(i.some((t=>0!==e[t]))){for(a in o)n.o(o,a)&&(n.m[a]=o[a]);if(u)u(n)}for(t&&t(r);s<i.length;s++)l=i[s],n.o(e,l)&&e[l]&&e[l][0](),e[l]=0},r=self.webpackChunkfrontend=self.webpackChunkfrontend||[];r.forEach(t.bind(null,0)),r.push=t.bind(null,r.push.bind(r))})();var r=n(43);function a(){}function l(e){return null==e?a:function(){return this.querySelector(e)}}function i(){return[]}function o(e){return null==e?i:function(){return this.querySelectorAll(e)}}function u(e){return function(){return null==(t=e.apply(this,arguments))?[]:Array.isArray(t)?t:Array.from(t);var t}}function s(e){return function(){return this.matches(e)}}function c(e){return function(t){return t.matches(e)}}var f=Array.prototype.find;function d(){return this.firstElementChild}var p=Array.prototype.filter;function h(){return Array.from(this.children)}function m(e){return new Array(e.length)}function v(e,t){this.ownerDocument=e.ownerDocument,this.namespaceURI=e.namespaceURI,this._next=null,this._parent=e,this.__data__=t}function g(e,t,n,r,a,l){for(var i,o=0,u=t.length,s=l.length;o<s;++o)(i=t[o])?(i.__data__=l[o],r[o]=i):n[o]=new v(e,l[o]);for(;o<u;++o)(i=t[o])&&(a[o]=i)}function y(e,t,n,r,a,l,i){var o,u,s,c=new Map,f=t.length,d=l.length,p=new Array(f);for(o=0;o<f;++o)(u=t[o])&&(p[o]=s=i.call(u,u.__data__,o,t)+"",c.has(s)?a[o]=u:c.set(s,u));for(o=0;o<d;++o)s=i.call(e,l[o],o,l)+"",(u=c.get(s))?(r[o]=u,u.__data__=l[o],c.delete(s)):n[o]=new v(e,l[o]);for(o=0;o<f;++o)(u=t[o])&&c.get(p[o])===u&&(a[o]=u)}function b(e){return e.__data__}function w(e){return"object"===typeof e&&"length"in e?e:Array.from(e)}function k(e,t){return e<t?-1:e>t?1:e>=t?0:NaN}v.prototype={constructor:v,appendChild:function(e){return this._parent.insertBefore(e,this._next)},insertBefore:function(e,t){return this._parent.insertBefore(e,t)},querySelector:function(e){return this._parent.querySelector(e)},querySelectorAll:function(e){return this._parent.querySelectorAll(e)}};var _="http://www.w3.org/1999/xhtml";const S={svg:"http://www.w3.org/2000/svg",xhtml:_,xlink:"http://www.w3.org/1999/xlink",xml:"http://www.w3.org/XML/1998/namespace",xmlns:"http://www.w3.org/2000/xmlns/"};function x(e){var t=e+="",n=t.indexOf(":");return n>=0&&"xmlns"!==(t=e.slice(0,n))&&(e=e.slice(n+1)),S.hasOwnProperty(t)?{space:S[t],local:e}:e}function E(e){return function(){this.removeAttribute(e)}}function C(e){return function(){this.removeAttributeNS(e.space,e.local)}}function N(e,t){return function(){this.setAttribute(e,t)}}function P(e,t){return function(){this.setAttributeNS(e.space,e.local,t)}}function z(e,t){return function(){var n=t.apply(this,arguments);null==n?this.removeAttribute(e):this.setAttribute(e,n)}}function T(e,t){return function(){var n=t.apply(this,arguments);null==n?this.removeAttributeNS(e.space,e.local):this.setAttributeNS(e.space,e.local,n)}}function L(e){return e.ownerDocument&&e.ownerDocument.defaultView||e.document&&e||e.defaultView}function M(e){return function(){this.style.removeProperty(e)}}function R(e,t,n){return function(){this.style.setProperty(e,t,n)}}function O(e,t,n){return function(){var r=t.apply(this,arguments);null==r?this.style.removeProperty(e):this.style.setProperty(e,r,n)}}function F(e,t){return e.style.getPropertyValue(t)||L(e).getComputedStyle(e,null).getPropertyValue(t)}function D(e){return function(){delete this[e]}}function I(e,t){return function(){this[e]=t}}function A(e,t){return function(){var n=t.apply(this,arguments);null==n?delete this[e]:this[e]=n}}function $(e){return e.trim().split(/^|\s+/)}function U(e){return e.classList||new j(e)}function j(e){this._node=e,this._names=$(e.getAttribute("class")||"")}function V(e,t){for(var n=U(e),r=-1,a=t.length;++r<a;)n.add(t[r])}function B(e,t){for(var n=U(e),r=-1,a=t.length;++r<a;)n.remove(t[r])}function H(e){return function(){V(this,e)}}function W(e){return function(){B(this,e)}}function Q(e,t){return function(){(t.apply(this,arguments)?V:B)(this,e)}}function q(){this.textContent=""}function X(e){return function(){this.textContent=e}}function Y(e){return function(){var t=e.apply(this,arguments);this.textContent=null==t?"":t}}function K(){this.innerHTML=""}function G(e){return function(){this.innerHTML=e}}function Z(e){return function(){var t=e.apply(this,arguments);this.innerHTML=null==t?"":t}}function J(){this.nextSibling&&this.parentNode.appendChild(this)}function ee(){this.previousSibling&&this.parentNode.insertBefore(this,this.parentNode.firstChild)}function te(e){return function(){var t=this.ownerDocument,n=this.namespaceURI;return n===_&&t.documentElement.namespaceURI===_?t.createElement(e):t.createElementNS(n,e)}}function ne(e){return function(){return this.ownerDocument.createElementNS(e.space,e.local)}}function re(e){var t=x(e);return(t.local?ne:te)(t)}function ae(){return null}function le(){var e=this.parentNode;e&&e.removeChild(this)}function ie(){var e=this.cloneNode(!1),t=this.parentNode;return t?t.insertBefore(e,this.nextSibling):e}function oe(){var e=this.cloneNode(!0),t=this.parentNode;return t?t.insertBefore(e,this.nextSibling):e}function ue(e){return function(){var t=this.__on;if(t){for(var n,r=0,a=-1,l=t.length;r<l;++r)n=t[r],e.type&&n.type!==e.type||n.name!==e.name?t[++a]=n:this.removeEventListener(n.type,n.listener,n.options);++a?t.length=a:delete this.__on}}}function se(e,t,n){return function(){var r,a=this.__on,l=function(e){return function(t){e.call(this,t,this.__data__)}}(t);if(a)for(var i=0,o=a.length;i<o;++i)if((r=a[i]).type===e.type&&r.name===e.name)return this.removeEventListener(r.type,r.listener,r.options),this.addEventListener(r.type,r.listener=l,r.options=n),void(r.value=t);this.addEventListener(e.type,l,n),r={type:e.type,name:e.name,value:t,listener:l,options:n},a?a.push(r):this.__on=[r]}}function ce(e,t,n){var r=L(e),a=r.CustomEvent;"function"===typeof a?a=new a(t,n):(a=r.document.createEvent("Event"),n?(a.initEvent(t,n.bubbles,n.cancelable),a.detail=n.detail):a.initEvent(t,!1,!1)),e.dispatchEvent(a)}function fe(e,t){return function(){return ce(this,e,t)}}function de(e,t){return function(){return ce(this,e,t.apply(this,arguments))}}j.prototype={add:function(e){this._names.indexOf(e)<0&&(this._names.push(e),this._node.setAttribute("class",this._names.join(" ")))},remove:function(e){var t=this._names.indexOf(e);t>=0&&(this._names.splice(t,1),this._node.setAttribute("class",this._names.join(" ")))},contains:function(e){return this._names.indexOf(e)>=0}};var pe=[null];function he(e,t){this._groups=e,this._parents=t}function me(){return new he([[document.documentElement]],pe)}he.prototype=me.prototype={constructor:he,select:function(e){"function"!==typeof e&&(e=l(e));for(var t=this._groups,n=t.length,r=new Array(n),a=0;a<n;++a)for(var i,o,u=t[a],s=u.length,c=r[a]=new Array(s),f=0;f<s;++f)(i=u[f])&&(o=e.call(i,i.__data__,f,u))&&("__data__"in i&&(o.__data__=i.__data__),c[f]=o);return new he(r,this._parents)},selectAll:function(e){e="function"===typeof e?u(e):o(e);for(var t=this._groups,n=t.length,r=[],a=[],l=0;l<n;++l)for(var i,s=t[l],c=s.length,f=0;f<c;++f)(i=s[f])&&(r.push(e.call(i,i.__data__,f,s)),a.push(i));return new he(r,a)},selectChild:function(e){return this.select(null==e?d:function(e){return function(){return f.call(this.children,e)}}("function"===typeof e?e:c(e)))},selectChildren:function(e){return this.selectAll(null==e?h:function(e){return function(){return p.call(this.children,e)}}("function"===typeof e?e:c(e)))},filter:function(e){"function"!==typeof e&&(e=s(e));for(var t=this._groups,n=t.length,r=new Array(n),a=0;a<n;++a)for(var l,i=t[a],o=i.length,u=r[a]=[],c=0;c<o;++c)(l=i[c])&&e.call(l,l.__data__,c,i)&&u.push(l);return new he(r,this._parents)},data:function(e,t){if(!arguments.length)return Array.from(this,b);var n,r=t?y:g,a=this._parents,l=this._groups;"function"!==typeof e&&(n=e,e=function(){return n});for(var i=l.length,o=new Array(i),u=new Array(i),s=new Array(i),c=0;c<i;++c){var f=a[c],d=l[c],p=d.length,h=w(e.call(f,f&&f.__data__,c,a)),m=h.length,v=u[c]=new Array(m),k=o[c]=new Array(m);r(f,d,v,k,s[c]=new Array(p),h,t);for(var _,S,x=0,E=0;x<m;++x)if(_=v[x]){for(x>=E&&(E=x+1);!(S=k[E])&&++E<m;);_._next=S||null}}return(o=new he(o,a))._enter=u,o._exit=s,o},enter:function(){return new he(this._enter||this._groups.map(m),this._parents)},exit:function(){return new he(this._exit||this._groups.map(m),this._parents)},join:function(e,t,n){var r=this.enter(),a=this,l=this.exit();return"function"===typeof e?(r=e(r))&&(r=r.selection()):r=r.append(e+""),null!=t&&(a=t(a))&&(a=a.selection()),null==n?l.remove():n(l),r&&a?r.merge(a).order():a},merge:function(e){for(var t=e.selection?e.selection():e,n=this._groups,r=t._groups,a=n.length,l=r.length,i=Math.min(a,l),o=new Array(a),u=0;u<i;++u)for(var s,c=n[u],f=r[u],d=c.length,p=o[u]=new Array(d),h=0;h<d;++h)(s=c[h]||f[h])&&(p[h]=s);for(;u<a;++u)o[u]=n[u];return new he(o,this._parents)},selection:function(){return this},order:function(){for(var e=this._groups,t=-1,n=e.length;++t<n;)for(var r,a=e[t],l=a.length-1,i=a[l];--l>=0;)(r=a[l])&&(i&&4^r.compareDocumentPosition(i)&&i.parentNode.insertBefore(r,i),i=r);return this},sort:function(e){function t(t,n){return t&&n?e(t.__data__,n.__data__):!t-!n}e||(e=k);for(var n=this._groups,r=n.length,a=new Array(r),l=0;l<r;++l){for(var i,o=n[l],u=o.length,s=a[l]=new Array(u),c=0;c<u;++c)(i=o[c])&&(s[c]=i);s.sort(t)}return new he(a,this._parents).order()},call:function(){var e=arguments[0];return arguments[0]=this,e.apply(null,arguments),this},nodes:function(){return Array.from(this)},node:function(){for(var e=this._groups,t=0,n=e.length;t<n;++t)for(var r=e[t],a=0,l=r.length;a<l;++a){var i=r[a];if(i)return i}return null},size:function(){let e=0;for(const t of this)++e;return e},empty:function(){return!this.node()},each:function(e){for(var t=this._groups,n=0,r=t.length;n<r;++n)for(var a,l=t[n],i=0,o=l.length;i<o;++i)(a=l[i])&&e.call(a,a.__data__,i,l);return this},attr:function(e,t){var n=x(e);if(arguments.length<2){var r=this.node();return n.local?r.getAttributeNS(n.space,n.local):r.getAttribute(n)}return this.each((null==t?n.local?C:E:"function"===typeof t?n.local?T:z:n.local?P:N)(n,t))},style:function(e,t,n){return arguments.length>1?this.each((null==t?M:"function"===typeof t?O:R)(e,t,null==n?"":n)):F(this.node(),e)},property:function(e,t){return arguments.length>1?this.each((null==t?D:"function"===typeof t?A:I)(e,t)):this.node()[e]},classed:function(e,t){var n=$(e+"");if(arguments.length<2){for(var r=U(this.node()),a=-1,l=n.length;++a<l;)if(!r.contains(n[a]))return!1;return!0}return this.each(("function"===typeof t?Q:t?H:W)(n,t))},text:function(e){return arguments.length?this.each(null==e?q:("function"===typeof e?Y:X)(e)):this.node().textContent},html:function(e){return arguments.length?this.each(null==e?K:("function"===typeof e?Z:G)(e)):this.node().innerHTML},raise:function(){return this.each(J)},lower:function(){return this.each(ee)},append:function(e){var t="function"===typeof e?e:re(e);return this.select((function(){return this.appendChild(t.apply(this,arguments))}))},insert:function(e,t){var n="function"===typeof e?e:re(e),r=null==t?ae:"function"===typeof t?t:l(t);return this.select((function(){return this.insertBefore(n.apply(this,arguments),r.apply(this,arguments)||null)}))},remove:function(){return this.each(le)},clone:function(e){return this.select(e?oe:ie)},datum:function(e){return arguments.length?this.property("__data__",e):this.node().__data__},on:function(e,t,n){var r,a,l=function(e){return e.trim().split(/^|\s+/).map((function(e){var t="",n=e.indexOf(".");return n>=0&&(t=e.slice(n+1),e=e.slice(0,n)),{type:e,name:t}}))}(e+""),i=l.length;if(!(arguments.length<2)){for(o=t?se:ue,r=0;r<i;++r)this.each(o(l[r],t,n));return this}var o=this.node().__on;if(o)for(var u,s=0,c=o.length;s<c;++s)for(r=0,u=o[s];r<i;++r)if((a=l[r]).type===u.type&&a.name===u.name)return u.value},dispatch:function(e,t){return this.each(("function"===typeof t?de:fe)(e,t))},[Symbol.iterator]:function*(){for(var e=this._groups,t=0,n=e.length;t<n;++t)for(var r,a=e[t],l=0,i=a.length;l<i;++l)(r=a[l])&&(yield r)}};const ve=me;var ge={value:()=>{}};function ye(){for(var e,t=0,n=arguments.length,r={};t<n;++t){if(!(e=arguments[t]+"")||e in r||/[\s.]/.test(e))throw new Error("illegal type: "+e);r[e]=[]}return new be(r)}function be(e){this._=e}function we(e,t){for(var n,r=0,a=e.length;r<a;++r)if((n=e[r]).name===t)return n.value}function ke(e,t,n){for(var r=0,a=e.length;r<a;++r)if(e[r].name===t){e[r]=ge,e=e.slice(0,r).concat(e.slice(r+1));break}return null!=n&&e.push({name:t,value:n}),e}be.prototype=ye.prototype={constructor:be,on:function(e,t){var n,r,a=this._,l=(r=a,(e+"").trim().split(/^|\s+/).map((function(e){var t="",n=e.indexOf(".");if(n>=0&&(t=e.slice(n+1),e=e.slice(0,n)),e&&!r.hasOwnProperty(e))throw new Error("unknown type: "+e);return{type:e,name:t}}))),i=-1,o=l.length;if(!(arguments.length<2)){if(null!=t&&"function"!==typeof t)throw new Error("invalid callback: "+t);for(;++i<o;)if(n=(e=l[i]).type)a[n]=ke(a[n],e.name,t);else if(null==t)for(n in a)a[n]=ke(a[n],e.name,null);return this}for(;++i<o;)if((n=(e=l[i]).type)&&(n=we(a[n],e.name)))return n},copy:function(){var e={},t=this._;for(var n in t)e[n]=t[n].slice();return new be(e)},call:function(e,t){if((n=arguments.length-2)>0)for(var n,r,a=new Array(n),l=0;l<n;++l)a[l]=arguments[l+2];if(!this._.hasOwnProperty(e))throw new Error("unknown type: "+e);for(l=0,n=(r=this._[e]).length;l<n;++l)r[l].value.apply(t,a)},apply:function(e,t,n){if(!this._.hasOwnProperty(e))throw new Error("unknown type: "+e);for(var r=this._[e],a=0,l=r.length;a<l;++a)r[a].value.apply(t,n)}};const _e=ye;var Se,xe,Ee=0,Ce=0,Ne=0,Pe=0,ze=0,Te=0,Le="object"===typeof performance&&performance.now?performance:Date,Me="object"===typeof window&&window.requestAnimationFrame?window.requestAnimationFrame.bind(window):function(e){setTimeout(e,17)};function Re(){return ze||(Me(Oe),ze=Le.now()+Te)}function Oe(){ze=0}function Fe(){this._call=this._time=this._next=null}function De(e,t,n){var r=new Fe;return r.restart(e,t,n),r}function Ie(){ze=(Pe=Le.now())+Te,Ee=Ce=0;try{!function(){Re(),++Ee;for(var e,t=Se;t;)(e=ze-t._time)>=0&&t._call.call(void 0,e),t=t._next;--Ee}()}finally{Ee=0,function(){var e,t,n=Se,r=1/0;for(;n;)n._call?(r>n._time&&(r=n._time),e=n,n=n._next):(t=n._next,n._next=null,n=e?e._next=t:Se=t);xe=e,$e(r)}(),ze=0}}function Ae(){var e=Le.now(),t=e-Pe;t>1e3&&(Te-=t,Pe=e)}function $e(e){Ee||(Ce&&(Ce=clearTimeout(Ce)),e-ze>24?(e<1/0&&(Ce=setTimeout(Ie,e-Le.now()-Te)),Ne&&(Ne=clearInterval(Ne))):(Ne||(Pe=Le.now(),Ne=setInterval(Ae,1e3)),Ee=1,Me(Ie)))}function Ue(e,t,n){var r=new Fe;return t=null==t?0:+t,r.restart((n=>{r.stop(),e(n+t)}),t,n),r}Fe.prototype=De.prototype={constructor:Fe,restart:function(e,t,n){if("function"!==typeof e)throw new TypeError("callback is not a function");n=(null==n?Re():+n)+(null==t?0:+t),this._next||xe===this||(xe?xe._next=this:Se=this,xe=this),this._call=e,this._time=n,$e()},stop:function(){this._call&&(this._call=null,this._time=1/0,$e())}};var je=_e("start","end","cancel","interrupt"),Ve=[];function Be(e,t,n,r,a,l){var i=e.__transition;if(i){if(n in i)return}else e.__transition={};!function(e,t,n){var r,a=e.__transition;function l(e){n.state=1,n.timer.restart(i,n.delay,n.time),n.delay<=e&&i(e-n.delay)}function i(l){var s,c,f,d;if(1!==n.state)return u();for(s in a)if((d=a[s]).name===n.name){if(3===d.state)return Ue(i);4===d.state?(d.state=6,d.timer.stop(),d.on.call("interrupt",e,e.__data__,d.index,d.group),delete a[s]):+s<t&&(d.state=6,d.timer.stop(),d.on.call("cancel",e,e.__data__,d.index,d.group),delete a[s])}if(Ue((function(){3===n.state&&(n.state=4,n.timer.restart(o,n.delay,n.time),o(l))})),n.state=2,n.on.call("start",e,e.__data__,n.index,n.group),2===n.state){for(n.state=3,r=new Array(f=n.tween.length),s=0,c=-1;s<f;++s)(d=n.tween[s].value.call(e,e.__data__,n.index,n.group))&&(r[++c]=d);r.length=c+1}}function o(t){for(var a=t<n.duration?n.ease.call(null,t/n.duration):(n.timer.restart(u),n.state=5,1),l=-1,i=r.length;++l<i;)r[l].call(e,a);5===n.state&&(n.on.call("end",e,e.__data__,n.index,n.group),u())}function u(){for(var r in n.state=6,n.timer.stop(),delete a[t],a)return;delete e.__transition}a[t]=n,n.timer=De(l,0,n.time)}(e,n,{name:t,index:r,group:a,on:je,tween:Ve,time:l.time,delay:l.delay,duration:l.duration,ease:l.ease,timer:null,state:0})}function He(e,t){var n=Qe(e,t);if(n.state>0)throw new Error("too late; already scheduled");return n}function We(e,t){var n=Qe(e,t);if(n.state>3)throw new Error("too late; already running");return n}function Qe(e,t){var n=e.__transition;if(!n||!(n=n[t]))throw new Error("transition not found");return n}function qe(e,t){return e=+e,t=+t,function(n){return e*(1-n)+t*n}}var Xe,Ye=180/Math.PI,Ke={translateX:0,translateY:0,rotate:0,skewX:0,scaleX:1,scaleY:1};function Ge(e,t,n,r,a,l){var i,o,u;return(i=Math.sqrt(e*e+t*t))&&(e/=i,t/=i),(u=e*n+t*r)&&(n-=e*u,r-=t*u),(o=Math.sqrt(n*n+r*r))&&(n/=o,r/=o,u/=o),e*r<t*n&&(e=-e,t=-t,u=-u,i=-i),{translateX:a,translateY:l,rotate:Math.atan2(t,e)*Ye,skewX:Math.atan(u)*Ye,scaleX:i,scaleY:o}}function Ze(e,t,n,r){function a(e){return e.length?e.pop()+" ":""}return function(l,i){var o=[],u=[];return l=e(l),i=e(i),function(e,r,a,l,i,o){if(e!==a||r!==l){var u=i.push("translate(",null,t,null,n);o.push({i:u-4,x:qe(e,a)},{i:u-2,x:qe(r,l)})}else(a||l)&&i.push("translate("+a+t+l+n)}(l.translateX,l.translateY,i.translateX,i.translateY,o,u),function(e,t,n,l){e!==t?(e-t>180?t+=360:t-e>180&&(e+=360),l.push({i:n.push(a(n)+"rotate(",null,r)-2,x:qe(e,t)})):t&&n.push(a(n)+"rotate("+t+r)}(l.rotate,i.rotate,o,u),function(e,t,n,l){e!==t?l.push({i:n.push(a(n)+"skewX(",null,r)-2,x:qe(e,t)}):t&&n.push(a(n)+"skewX("+t+r)}(l.skewX,i.skewX,o,u),function(e,t,n,r,l,i){if(e!==n||t!==r){var o=l.push(a(l)+"scale(",null,",",null,")");i.push({i:o-4,x:qe(e,n)},{i:o-2,x:qe(t,r)})}else 1===n&&1===r||l.push(a(l)+"scale("+n+","+r+")")}(l.scaleX,l.scaleY,i.scaleX,i.scaleY,o,u),l=i=null,function(e){for(var t,n=-1,r=u.length;++n<r;)o[(t=u[n]).i]=t.x(e);return o.join("")}}}var Je=Ze((function(e){const t=new("function"===typeof DOMMatrix?DOMMatrix:WebKitCSSMatrix)(e+"");return t.isIdentity?Ke:Ge(t.a,t.b,t.c,t.d,t.e,t.f)}),"px, ","px)","deg)"),et=Ze((function(e){return null==e?Ke:(Xe||(Xe=document.createElementNS("http://www.w3.org/2000/svg","g")),Xe.setAttribute("transform",e),(e=Xe.transform.baseVal.consolidate())?Ge((e=e.matrix).a,e.b,e.c,e.d,e.e,e.f):Ke)}),", ",")",")");function tt(e,t){var n,r;return function(){var a=We(this,e),l=a.tween;if(l!==n)for(var i=0,o=(r=n=l).length;i<o;++i)if(r[i].name===t){(r=r.slice()).splice(i,1);break}a.tween=r}}function nt(e,t,n){var r,a;if("function"!==typeof n)throw new Error;return function(){var l=We(this,e),i=l.tween;if(i!==r){a=(r=i).slice();for(var o={name:t,value:n},u=0,s=a.length;u<s;++u)if(a[u].name===t){a[u]=o;break}u===s&&a.push(o)}l.tween=a}}function rt(e,t,n){var r=e._id;return e.each((function(){var e=We(this,r);(e.value||(e.value={}))[t]=n.apply(this,arguments)})),function(e){return Qe(e,r).value[t]}}function at(e,t,n){e.prototype=t.prototype=n,n.constructor=e}function lt(e,t){var n=Object.create(e.prototype);for(var r in t)n[r]=t[r];return n}function it(){}var ot=.7,ut=1/ot,st="\\s*([+-]?\\d+)\\s*",ct="\\s*([+-]?(?:\\d*\\.)?\\d+(?:[eE][+-]?\\d+)?)\\s*",ft="\\s*([+-]?(?:\\d*\\.)?\\d+(?:[eE][+-]?\\d+)?)%\\s*",dt=/^#([0-9a-f]{3,8})$/,pt=new RegExp(`^rgb\\(${st},${st},${st}\\)$`),ht=new RegExp(`^rgb\\(${ft},${ft},${ft}\\)$`),mt=new RegExp(`^rgba\\(${st},${st},${st},${ct}\\)$`),vt=new RegExp(`^rgba\\(${ft},${ft},${ft},${ct}\\)$`),gt=new RegExp(`^hsl\\(${ct},${ft},${ft}\\)$`),yt=new RegExp(`^hsla\\(${ct},${ft},${ft},${ct}\\)$`),bt={aliceblue:15792383,antiquewhite:16444375,aqua:65535,aquamarine:8388564,azure:15794175,beige:16119260,bisque:16770244,black:0,blanchedalmond:16772045,blue:255,blueviolet:9055202,brown:10824234,burlywood:14596231,cadetblue:6266528,chartreuse:8388352,chocolate:13789470,coral:16744272,cornflowerblue:6591981,cornsilk:16775388,crimson:14423100,cyan:65535,darkblue:139,darkcyan:35723,darkgoldenrod:12092939,darkgray:11119017,darkgreen:25600,darkgrey:11119017,darkkhaki:12433259,darkmagenta:9109643,darkolivegreen:5597999,darkorange:16747520,darkorchid:10040012,darkred:9109504,darksalmon:15308410,darkseagreen:9419919,darkslateblue:4734347,darkslategray:3100495,darkslategrey:3100495,darkturquoise:52945,darkviolet:9699539,deeppink:16716947,deepskyblue:49151,dimgray:6908265,dimgrey:6908265,dodgerblue:2003199,firebrick:11674146,floralwhite:16775920,forestgreen:2263842,fuchsia:16711935,gainsboro:14474460,ghostwhite:16316671,gold:16766720,goldenrod:14329120,gray:8421504,green:32768,greenyellow:11403055,grey:8421504,honeydew:15794160,hotpink:16738740,indianred:13458524,indigo:4915330,ivory:16777200,khaki:15787660,lavender:15132410,lavenderblush:16773365,lawngreen:8190976,lemonchiffon:16775885,lightblue:11393254,lightcoral:15761536,lightcyan:14745599,lightgoldenrodyellow:16448210,lightgray:13882323,lightgreen:9498256,lightgrey:13882323,lightpink:16758465,lightsalmon:16752762,lightseagreen:2142890,lightskyblue:8900346,lightslategray:7833753,lightslategrey:7833753,lightsteelblue:11584734,lightyellow:16777184,lime:65280,limegreen:3329330,linen:16445670,magenta:16711935,maroon:8388608,mediumaquamarine:6737322,mediumblue:205,mediumorchid:12211667,mediumpurple:9662683,mediumseagreen:3978097,mediumslateblue:8087790,mediumspringgreen:64154,mediumturquoise:4772300,mediumvioletred:13047173,midnightblue:1644912,mintcream:16121850,mistyrose:16770273,moccasin:16770229,navajowhite:16768685,navy:128,oldlace:16643558,olive:8421376,olivedrab:7048739,orange:16753920,orangered:16729344,orchid:14315734,palegoldenrod:15657130,palegreen:10025880,paleturquoise:11529966,palevioletred:14381203,papayawhip:16773077,peachpuff:16767673,peru:13468991,pink:16761035,plum:14524637,powderblue:11591910,purple:8388736,rebeccapurple:6697881,red:16711680,rosybrown:12357519,royalblue:4286945,saddlebrown:9127187,salmon:16416882,sandybrown:16032864,seagreen:3050327,seashell:16774638,sienna:10506797,silver:12632256,skyblue:8900331,slateblue:6970061,slategray:7372944,slategrey:7372944,snow:16775930,springgreen:65407,steelblue:4620980,tan:13808780,teal:32896,thistle:14204888,tomato:16737095,turquoise:4251856,violet:15631086,wheat:16113331,white:16777215,whitesmoke:16119285,yellow:16776960,yellowgreen:10145074};function wt(){return this.rgb().formatHex()}function kt(){return this.rgb().formatRgb()}function _t(e){var t,n;return e=(e+"").trim().toLowerCase(),(t=dt.exec(e))?(n=t[1].length,t=parseInt(t[1],16),6===n?St(t):3===n?new Ct(t>>8&15|t>>4&240,t>>4&15|240&t,(15&t)<<4|15&t,1):8===n?xt(t>>24&255,t>>16&255,t>>8&255,(255&t)/255):4===n?xt(t>>12&15|t>>8&240,t>>8&15|t>>4&240,t>>4&15|240&t,((15&t)<<4|15&t)/255):null):(t=pt.exec(e))?new Ct(t[1],t[2],t[3],1):(t=ht.exec(e))?new Ct(255*t[1]/100,255*t[2]/100,255*t[3]/100,1):(t=mt.exec(e))?xt(t[1],t[2],t[3],t[4]):(t=vt.exec(e))?xt(255*t[1]/100,255*t[2]/100,255*t[3]/100,t[4]):(t=gt.exec(e))?Mt(t[1],t[2]/100,t[3]/100,1):(t=yt.exec(e))?Mt(t[1],t[2]/100,t[3]/100,t[4]):bt.hasOwnProperty(e)?St(bt[e]):"transparent"===e?new Ct(NaN,NaN,NaN,0):null}function St(e){return new Ct(e>>16&255,e>>8&255,255&e,1)}function xt(e,t,n,r){return r<=0&&(e=t=n=NaN),new Ct(e,t,n,r)}function Et(e,t,n,r){return 1===arguments.length?((a=e)instanceof it||(a=_t(a)),a?new Ct((a=a.rgb()).r,a.g,a.b,a.opacity):new Ct):new Ct(e,t,n,null==r?1:r);var a}function Ct(e,t,n,r){this.r=+e,this.g=+t,this.b=+n,this.opacity=+r}function Nt(){return`#${Lt(this.r)}${Lt(this.g)}${Lt(this.b)}`}function Pt(){const e=zt(this.opacity);return`${1===e?"rgb(":"rgba("}${Tt(this.r)}, ${Tt(this.g)}, ${Tt(this.b)}${1===e?")":`, ${e})`}`}function zt(e){return isNaN(e)?1:Math.max(0,Math.min(1,e))}function Tt(e){return Math.max(0,Math.min(255,Math.round(e)||0))}function Lt(e){return((e=Tt(e))<16?"0":"")+e.toString(16)}function Mt(e,t,n,r){return r<=0?e=t=n=NaN:n<=0||n>=1?e=t=NaN:t<=0&&(e=NaN),new Ot(e,t,n,r)}function Rt(e){if(e instanceof Ot)return new Ot(e.h,e.s,e.l,e.opacity);if(e instanceof it||(e=_t(e)),!e)return new Ot;if(e instanceof Ot)return e;var t=(e=e.rgb()).r/255,n=e.g/255,r=e.b/255,a=Math.min(t,n,r),l=Math.max(t,n,r),i=NaN,o=l-a,u=(l+a)/2;return o?(i=t===l?(n-r)/o+6*(n<r):n===l?(r-t)/o+2:(t-n)/o+4,o/=u<.5?l+a:2-l-a,i*=60):o=u>0&&u<1?0:i,new Ot(i,o,u,e.opacity)}function Ot(e,t,n,r){this.h=+e,this.s=+t,this.l=+n,this.opacity=+r}function Ft(e){return(e=(e||0)%360)<0?e+360:e}function Dt(e){return Math.max(0,Math.min(1,e||0))}function It(e,t,n){return 255*(e<60?t+(n-t)*e/60:e<180?n:e<240?t+(n-t)*(240-e)/60:t)}function At(e,t,n,r,a){var l=e*e,i=l*e;return((1-3*e+3*l-i)*t+(4-6*l+3*i)*n+(1+3*e+3*l-3*i)*r+i*a)/6}at(it,_t,{copy(e){return Object.assign(new this.constructor,this,e)},displayable(){return this.rgb().displayable()},hex:wt,formatHex:wt,formatHex8:function(){return this.rgb().formatHex8()},formatHsl:function(){return Rt(this).formatHsl()},formatRgb:kt,toString:kt}),at(Ct,Et,lt(it,{brighter(e){return e=null==e?ut:Math.pow(ut,e),new Ct(this.r*e,this.g*e,this.b*e,this.opacity)},darker(e){return e=null==e?ot:Math.pow(ot,e),new Ct(this.r*e,this.g*e,this.b*e,this.opacity)},rgb(){return this},clamp(){return new Ct(Tt(this.r),Tt(this.g),Tt(this.b),zt(this.opacity))},displayable(){return-.5<=this.r&&this.r<255.5&&-.5<=this.g&&this.g<255.5&&-.5<=this.b&&this.b<255.5&&0<=this.opacity&&this.opacity<=1},hex:Nt,formatHex:Nt,formatHex8:function(){return`#${Lt(this.r)}${Lt(this.g)}${Lt(this.b)}${Lt(255*(isNaN(this.opacity)?1:this.opacity))}`},formatRgb:Pt,toString:Pt})),at(Ot,(function(e,t,n,r){return 1===arguments.length?Rt(e):new Ot(e,t,n,null==r?1:r)}),lt(it,{brighter(e){return e=null==e?ut:Math.pow(ut,e),new Ot(this.h,this.s,this.l*e,this.opacity)},darker(e){return e=null==e?ot:Math.pow(ot,e),new Ot(this.h,this.s,this.l*e,this.opacity)},rgb(){var e=this.h%360+360*(this.h<0),t=isNaN(e)||isNaN(this.s)?0:this.s,n=this.l,r=n+(n<.5?n:1-n)*t,a=2*n-r;return new Ct(It(e>=240?e-240:e+120,a,r),It(e,a,r),It(e<120?e+240:e-120,a,r),this.opacity)},clamp(){return new Ot(Ft(this.h),Dt(this.s),Dt(this.l),zt(this.opacity))},displayable(){return(0<=this.s&&this.s<=1||isNaN(this.s))&&0<=this.l&&this.l<=1&&0<=this.opacity&&this.opacity<=1},formatHsl(){const e=zt(this.opacity);return`${1===e?"hsl(":"hsla("}${Ft(this.h)}, ${100*Dt(this.s)}%, ${100*Dt(this.l)}%${1===e?")":`, ${e})`}`}}));const $t=e=>()=>e;function Ut(e,t){return function(n){return e+n*t}}function jt(e){return 1===(e=+e)?Vt:function(t,n){return n-t?function(e,t,n){return e=Math.pow(e,n),t=Math.pow(t,n)-e,n=1/n,function(r){return Math.pow(e+r*t,n)}}(t,n,e):$t(isNaN(t)?n:t)}}function Vt(e,t){var n=t-e;return n?Ut(e,n):$t(isNaN(e)?t:e)}const Bt=function e(t){var n=jt(t);function r(e,t){var r=n((e=Et(e)).r,(t=Et(t)).r),a=n(e.g,t.g),l=n(e.b,t.b),i=Vt(e.opacity,t.opacity);return function(t){return e.r=r(t),e.g=a(t),e.b=l(t),e.opacity=i(t),e+""}}return r.gamma=e,r}(1);function Ht(e){return function(t){var n,r,a=t.length,l=new Array(a),i=new Array(a),o=new Array(a);for(n=0;n<a;++n)r=Et(t[n]),l[n]=r.r||0,i[n]=r.g||0,o[n]=r.b||0;return l=e(l),i=e(i),o=e(o),r.opacity=1,function(e){return r.r=l(e),r.g=i(e),r.b=o(e),r+""}}}Ht((function(e){var t=e.length-1;return function(n){var r=n<=0?n=0:n>=1?(n=1,t-1):Math.floor(n*t),a=e[r],l=e[r+1],i=r>0?e[r-1]:2*a-l,o=r<t-1?e[r+2]:2*l-a;return At((n-r/t)*t,i,a,l,o)}})),Ht((function(e){var t=e.length;return function(n){var r=Math.floor(((n%=1)<0?++n:n)*t),a=e[(r+t-1)%t],l=e[r%t],i=e[(r+1)%t],o=e[(r+2)%t];return At((n-r/t)*t,a,l,i,o)}}));var Wt=/[-+]?(?:\d+\.?\d*|\.?\d+)(?:[eE][-+]?\d+)?/g,Qt=new RegExp(Wt.source,"g");function qt(e,t){var n,r,a,l=Wt.lastIndex=Qt.lastIndex=0,i=-1,o=[],u=[];for(e+="",t+="";(n=Wt.exec(e))&&(r=Qt.exec(t));)(a=r.index)>l&&(a=t.slice(l,a),o[i]?o[i]+=a:o[++i]=a),(n=n[0])===(r=r[0])?o[i]?o[i]+=r:o[++i]=r:(o[++i]=null,u.push({i:i,x:qe(n,r)})),l=Qt.lastIndex;return l<t.length&&(a=t.slice(l),o[i]?o[i]+=a:o[++i]=a),o.length<2?u[0]?function(e){return function(t){return e(t)+""}}(u[0].x):function(e){return function(){return e}}(t):(t=u.length,function(e){for(var n,r=0;r<t;++r)o[(n=u[r]).i]=n.x(e);return o.join("")})}function Xt(e,t){var n;return("number"===typeof t?qe:t instanceof _t?Bt:(n=_t(t))?(t=n,Bt):qt)(e,t)}function Yt(e){return function(){this.removeAttribute(e)}}function Kt(e){return function(){this.removeAttributeNS(e.space,e.local)}}function Gt(e,t,n){var r,a,l=n+"";return function(){var i=this.getAttribute(e);return i===l?null:i===r?a:a=t(r=i,n)}}function Zt(e,t,n){var r,a,l=n+"";return function(){var i=this.getAttributeNS(e.space,e.local);return i===l?null:i===r?a:a=t(r=i,n)}}function Jt(e,t,n){var r,a,l;return function(){var i,o,u=n(this);if(null!=u)return(i=this.getAttribute(e))===(o=u+"")?null:i===r&&o===a?l:(a=o,l=t(r=i,u));this.removeAttribute(e)}}function en(e,t,n){var r,a,l;return function(){var i,o,u=n(this);if(null!=u)return(i=this.getAttributeNS(e.space,e.local))===(o=u+"")?null:i===r&&o===a?l:(a=o,l=t(r=i,u));this.removeAttributeNS(e.space,e.local)}}function tn(e,t){var n,r;function a(){var a=t.apply(this,arguments);return a!==r&&(n=(r=a)&&function(e,t){return function(n){this.setAttributeNS(e.space,e.local,t.call(this,n))}}(e,a)),n}return a._value=t,a}function nn(e,t){var n,r;function a(){var a=t.apply(this,arguments);return a!==r&&(n=(r=a)&&function(e,t){return function(n){this.setAttribute(e,t.call(this,n))}}(e,a)),n}return a._value=t,a}function rn(e,t){return function(){He(this,e).delay=+t.apply(this,arguments)}}function an(e,t){return t=+t,function(){He(this,e).delay=t}}function ln(e,t){return function(){We(this,e).duration=+t.apply(this,arguments)}}function on(e,t){return t=+t,function(){We(this,e).duration=t}}var un=ve.prototype.constructor;function sn(e){return function(){this.style.removeProperty(e)}}var cn=0;function fn(e,t,n,r){this._groups=e,this._parents=t,this._name=n,this._id=r}function dn(){return++cn}var pn=ve.prototype;fn.prototype=function(e){return ve().transition(e)}.prototype={constructor:fn,select:function(e){var t=this._name,n=this._id;"function"!==typeof e&&(e=l(e));for(var r=this._groups,a=r.length,i=new Array(a),o=0;o<a;++o)for(var u,s,c=r[o],f=c.length,d=i[o]=new Array(f),p=0;p<f;++p)(u=c[p])&&(s=e.call(u,u.__data__,p,c))&&("__data__"in u&&(s.__data__=u.__data__),d[p]=s,Be(d[p],t,n,p,d,Qe(u,n)));return new fn(i,this._parents,t,n)},selectAll:function(e){var t=this._name,n=this._id;"function"!==typeof e&&(e=o(e));for(var r=this._groups,a=r.length,l=[],i=[],u=0;u<a;++u)for(var s,c=r[u],f=c.length,d=0;d<f;++d)if(s=c[d]){for(var p,h=e.call(s,s.__data__,d,c),m=Qe(s,n),v=0,g=h.length;v<g;++v)(p=h[v])&&Be(p,t,n,v,h,m);l.push(h),i.push(s)}return new fn(l,i,t,n)},selectChild:pn.selectChild,selectChildren:pn.selectChildren,filter:function(e){"function"!==typeof e&&(e=s(e));for(var t=this._groups,n=t.length,r=new Array(n),a=0;a<n;++a)for(var l,i=t[a],o=i.length,u=r[a]=[],c=0;c<o;++c)(l=i[c])&&e.call(l,l.__data__,c,i)&&u.push(l);return new fn(r,this._parents,this._name,this._id)},merge:function(e){if(e._id!==this._id)throw new Error;for(var t=this._groups,n=e._groups,r=t.length,a=n.length,l=Math.min(r,a),i=new Array(r),o=0;o<l;++o)for(var u,s=t[o],c=n[o],f=s.length,d=i[o]=new Array(f),p=0;p<f;++p)(u=s[p]||c[p])&&(d[p]=u);for(;o<r;++o)i[o]=t[o];return new fn(i,this._parents,this._name,this._id)},selection:function(){return new un(this._groups,this._parents)},transition:function(){for(var e=this._name,t=this._id,n=dn(),r=this._groups,a=r.length,l=0;l<a;++l)for(var i,o=r[l],u=o.length,s=0;s<u;++s)if(i=o[s]){var c=Qe(i,t);Be(i,e,n,s,o,{time:c.time+c.delay+c.duration,delay:0,duration:c.duration,ease:c.ease})}return new fn(r,this._parents,e,n)},call:pn.call,nodes:pn.nodes,node:pn.node,size:pn.size,empty:pn.empty,each:pn.each,on:function(e,t){var n=this._id;return arguments.length<2?Qe(this.node(),n).on.on(e):this.each(function(e,t,n){var r,a,l=function(e){return(e+"").trim().split(/^|\s+/).every((function(e){var t=e.indexOf(".");return t>=0&&(e=e.slice(0,t)),!e||"start"===e}))}(t)?He:We;return function(){var i=l(this,e),o=i.on;o!==r&&(a=(r=o).copy()).on(t,n),i.on=a}}(n,e,t))},attr:function(e,t){var n=x(e),r="transform"===n?et:Xt;return this.attrTween(e,"function"===typeof t?(n.local?en:Jt)(n,r,rt(this,"attr."+e,t)):null==t?(n.local?Kt:Yt)(n):(n.local?Zt:Gt)(n,r,t))},attrTween:function(e,t){var n="attr."+e;if(arguments.length<2)return(n=this.tween(n))&&n._value;if(null==t)return this.tween(n,null);if("function"!==typeof t)throw new Error;var r=x(e);return this.tween(n,(r.local?tn:nn)(r,t))},style:function(e,t,n){var r="transform"===(e+="")?Je:Xt;return null==t?this.styleTween(e,function(e,t){var n,r,a;return function(){var l=F(this,e),i=(this.style.removeProperty(e),F(this,e));return l===i?null:l===n&&i===r?a:a=t(n=l,r=i)}}(e,r)).on("end.style."+e,sn(e)):"function"===typeof t?this.styleTween(e,function(e,t,n){var r,a,l;return function(){var i=F(this,e),o=n(this),u=o+"";return null==o&&(this.style.removeProperty(e),u=o=F(this,e)),i===u?null:i===r&&u===a?l:(a=u,l=t(r=i,o))}}(e,r,rt(this,"style."+e,t))).each(function(e,t){var n,r,a,l,i="style."+t,o="end."+i;return function(){var u=We(this,e),s=u.on,c=null==u.value[i]?l||(l=sn(t)):void 0;s===n&&a===c||(r=(n=s).copy()).on(o,a=c),u.on=r}}(this._id,e)):this.styleTween(e,function(e,t,n){var r,a,l=n+"";return function(){var i=F(this,e);return i===l?null:i===r?a:a=t(r=i,n)}}(e,r,t),n).on("end.style."+e,null)},styleTween:function(e,t,n){var r="style."+(e+="");if(arguments.length<2)return(r=this.tween(r))&&r._value;if(null==t)return this.tween(r,null);if("function"!==typeof t)throw new Error;return this.tween(r,function(e,t,n){var r,a;function l(){var l=t.apply(this,arguments);return l!==a&&(r=(a=l)&&function(e,t,n){return function(r){this.style.setProperty(e,t.call(this,r),n)}}(e,l,n)),r}return l._value=t,l}(e,t,null==n?"":n))},text:function(e){return this.tween("text","function"===typeof e?function(e){return function(){var t=e(this);this.textContent=null==t?"":t}}(rt(this,"text",e)):function(e){return function(){this.textContent=e}}(null==e?"":e+""))},textTween:function(e){var t="text";if(arguments.length<1)return(t=this.tween(t))&&t._value;if(null==e)return this.tween(t,null);if("function"!==typeof e)throw new Error;return this.tween(t,function(e){var t,n;function r(){var r=e.apply(this,arguments);return r!==n&&(t=(n=r)&&function(e){return function(t){this.textContent=e.call(this,t)}}(r)),t}return r._value=e,r}(e))},remove:function(){return this.on("end.remove",function(e){return function(){var t=this.parentNode;for(var n in this.__transition)if(+n!==e)return;t&&t.removeChild(this)}}(this._id))},tween:function(e,t){var n=this._id;if(e+="",arguments.length<2){for(var r,a=Qe(this.node(),n).tween,l=0,i=a.length;l<i;++l)if((r=a[l]).name===e)return r.value;return null}return this.each((null==t?tt:nt)(n,e,t))},delay:function(e){var t=this._id;return arguments.length?this.each(("function"===typeof e?rn:an)(t,e)):Qe(this.node(),t).delay},duration:function(e){var t=this._id;return arguments.length?this.each(("function"===typeof e?ln:on)(t,e)):Qe(this.node(),t).duration},ease:function(e){var t=this._id;return arguments.length?this.each(function(e,t){if("function"!==typeof t)throw new Error;return function(){We(this,e).ease=t}}(t,e)):Qe(this.node(),t).ease},easeVarying:function(e){if("function"!==typeof e)throw new Error;return this.each(function(e,t){return function(){var n=t.apply(this,arguments);if("function"!==typeof n)throw new Error;We(this,e).ease=n}}(this._id,e))},end:function(){var e,t,n=this,r=n._id,a=n.size();return new Promise((function(l,i){var o={value:i},u={value:function(){0===--a&&l()}};n.each((function(){var n=We(this,r),a=n.on;a!==e&&((t=(e=a).copy())._.cancel.push(o),t._.interrupt.push(o),t._.end.push(u)),n.on=t})),0===a&&l()}))},[Symbol.iterator]:pn[Symbol.iterator]};var hn={time:null,delay:0,duration:250,ease:function(e){return((e*=2)<=1?e*e*e:(e-=2)*e*e+2)/2}};function mn(e,t){for(var n;!(n=e.__transition)||!(n=n[t]);)if(!(e=e.parentNode))throw new Error(`transition ${t} not found`);return n}ve.prototype.interrupt=function(e){return this.each((function(){!function(e,t){var n,r,a,l=e.__transition,i=!0;if(l){for(a in t=null==t?null:t+"",l)(n=l[a]).name===t?(r=n.state>2&&n.state<5,n.state=6,n.timer.stop(),n.on.call(r?"interrupt":"cancel",e,e.__data__,n.index,n.group),delete l[a]):i=!1;i&&delete e.__transition}}(this,e)}))},ve.prototype.transition=function(e){var t,n;e instanceof fn?(t=e._id,e=e._name):(t=dn(),(n=hn).time=Re(),e=null==e?null:e+"");for(var r=this._groups,a=r.length,l=0;l<a;++l)for(var i,o=r[l],u=o.length,s=0;s<u;++s)(i=o[s])&&Be(i,e,t,s,o,n||mn(i,t));return new fn(r,this._parents,e,t)};const{abs:vn,max:gn,min:yn}=Math;function bn(e){return[+e[0],+e[1]]}function wn(e){return[bn(e[0]),bn(e[1])]}["w","e"].map(kn),["n","s"].map(kn),["n","w","e","s","nw","ne","sw","se"].map(kn);function kn(e){return{type:e}}function _n(e,t,n){this.k=e,this.x=t,this.y=n}_n.prototype={constructor:_n,scale:function(e){return 1===e?this:new _n(this.k*e,this.x,this.y)},translate:function(e,t){return 0===e&0===t?this:new _n(this.k,this.x+this.k*e,this.y+this.k*t)},apply:function(e){return[e[0]*this.k+this.x,e[1]*this.k+this.y]},applyX:function(e){return e*this.k+this.x},applyY:function(e){return e*this.k+this.y},invert:function(e){return[(e[0]-this.x)/this.k,(e[1]-this.y)/this.k]},invertX:function(e){return(e-this.x)/this.k},invertY:function(e){return(e-this.y)/this.k},rescaleX:function(e){return e.copy().domain(e.range().map(this.invertX,this).map(e.invert,e))},rescaleY:function(e){return e.copy().domain(e.range().map(this.invertY,this).map(e.invert,e))},toString:function(){return"translate("+this.x+","+this.y+") scale("+this.k+")"}};new _n(1,0,0);_n.prototype;var Sn=n(579);const xn=()=>((0,r.useEffect)((()=>{const e=function(e){return"string"===typeof e?new he([[document.querySelector(e)]],[document.documentElement]):new he([[e]],pe)}("svg");e.attr("width"),e.attr("height");async function t(e){const t=await fetch(e);return await t.json()}new EventSource("/api/visualizacion/stream").onmessage=function(t){!function(t){const n=400+100*t;e.append("circle").attr("cx",400).attr("cy",0).attr("r",5).style("fill","red").transition().duration(2e3).attr("cx",n).attr("cy",300)}(parseFloat(t.data))},async function(){const n=await t("/api/componentes/clavos");e.selectAll("circle.clavo").data(n).enter().append("circle").attr("class","clavo").attr("cx",((e,t)=>200+t%5*100)).attr("cy",((e,t)=>50+50*Math.floor(t/5))).attr("r",5).style("fill","black")}(),async function(){const n=await t("/api/componentes/contenedores");e.selectAll("rect.contenedor").data(n).enter().append("rect").attr("class","contenedor").attr("x",((e,t)=>200+100*t-25)).attr("y",300).attr("width",50).attr("height",200).style("fill","lightgrey")}()}),[]),(0,Sn.jsx)("svg",{width:"800",height:"600"})),En=e=>{e&&e instanceof Function&&n.e(453).then(n.bind(n,453)).then((t=>{let{getCLS:n,getFID:r,getFCP:a,getLCP:l,getTTFB:i}=t;n(e),r(e),a(e),l(e),i(e)}))};n(950).createRoot(document.getElementById("root")).render((0,Sn.jsx)(r.StrictMode,{children:(0,Sn.jsx)(xn,{})})),En()})();
+//# sourceMappingURL=main.fd488f23.js.map
