+++ src/main/resources/application.properties
@@ -1,3 +1,7 @@
 spring.application.name=Parcial-1Concurrente
 spring.docker.compose.lifecycle-management=start-only
 spring.profiles.active=mysql
+
+# Updated properties to serve static content
+spring.mvc.static-path-pattern=/**
+spring.web.resources.static-locations=classpath:/static/
