+++ src/main/resources/static/asset-manifest.json
@@ -1,0 +1,15 @@
+{
+  "files": {
+    "main.css": "/static/css/main.3282904e.css",
+    "main.js": "/static/js/main.fd488f23.js",
+    "static/js/453.8ab44547.chunk.js": "/static/js/453.8ab44547.chunk.js",
+    "index.html": "/index.html",
+    "main.3282904e.css.map": "/static/css/main.3282904e.css.map",
+    "main.fd488f23.js.map": "/static/js/main.fd488f23.js.map",
+    "453.8ab44547.chunk.js.map": "/static/js/453.8ab44547.chunk.js.map"
+  },
+  "entrypoints": [
+    "static/css/main.3282904e.css",
+    "static/js/main.fd488f23.js"
+  ]
+}
